Aggregation

Aggregation

Sampling

Weekdays

Temporal.mondays — Function.

mondays(x::TS)

Return a time series containing all observations occuring on Mondays of the given TS object.

source

Temporal.tuesdays — Function.

tuesdays(x::TS)

Return a time series containing all observations occuring on Tuesdays of the given TS object.

source

Temporal.wednesdays — Function.

wednesdays(x::TS)

Return a time series containing all observations occuring on Wednesdays of the given TS object.

source

Temporal.thursdays — Function.

thursdays(x::TS)

Return a time series containing all observations occuring on Thursdays of the given TS object.

source

Temporal.fridays — Function.

fridays(x::TS)

Return a time series containing all observations occuring on Fridays of the given TS object.

source

Temporal.saturdays — Function.

saturdays(x::TS)

Return a time series containing all observations occuring on Saturdays of the given TS object.

source

Temporal.sundays — Function.

sundays(x::TS)

Return a time series containing all observations occuring on Sundays of the given TS object.

source

Interval Boundaries

Weekly

Temporal.bow — Function.

bow(x::TS; cal::Bool=false)

Return a time series containing all observations occuring at the beginnings of the weeks of the input.

If cal is false, only observations occurring the last calendar day of the week are returned

If cal is true, all observation for which the previous index is a prior week are returned

source

Temporal.eow — Function.

eow(x::TS; cal::Bool=false)

Return a time series containing all observations occuring at the ends of the weeks of the input.

If cal is false, only observations occurring the last calendar day of the week are returned

If cal is true, all observation for which the next index is a new week are returned

source

Monthly

Temporal.bom — Function.

bom(x::TS; cal::Bool=false)

Return a time series containing all observations occuring at the beginnings of the months of the input.

If cal is false, only observations occurring the last calendar day of the month are returned

If cal is true, all observation for which the previous index is a prior month are returned

source

Temporal.eom — Function.

eom(x::TS; cal::Bool=false)

Return a time series containing all observations occuring at the ends of the months of the input.

If cal is false, only observations occurring the last calendar day of the month are returned

If cal is true, all observation for which the next index is a new month are returned

source

Quarterly

Temporal.boq — Function.

boq(x::TS; cal::Bool=false)

Return a time series containing all observations occuring at the beginnings of the quarters of the input.

If cal is false, only observations occurring the last calendar day of the quarter are returned

If cal is true, all observation for which the previous index is a prior quarter are returned

source

Temporal.eoq — Function.

eoq(x::TS; cal::Bool=false)

Return a time series containing all observations occuring at the ends of the quarters of the input.

If cal is false, only observations occurring the last calendar day of the quarter are returned

If cal is true, all observation for which the next index is a new quarter are returned

source

Yearly

Temporal.boy — Function.

boy(x::TS; cal::Bool=false)

Return a time series containing all observations occuring at the beginnings of the years of the input.

If cal is false, only observations occurring the last calendar day of the year are returned

If cal is true, all observation for which the previous index is a prior year are returned

source

Temporal.eoy — Function.

eoy(x::TS; cal::Bool=false)

Return a time series containing all observations occuring at the ends of the years of the input.

If cal is false, only observations occurring the last calendar day of the year are returned

If cal is true, all observation for which the next index is a new year are returned

source

Collapsing

Temporal.collapse — Function.

collapse(x::TS{V,T}, at::Function; fun::Function=last, args...)::TS{V,T} where {V,T}

Compute the function fun over period boundaries defined by at (e.g. eom, boq, etc.) and returng a time series of the results.

Keyword arguments (args...) are passed to the function fun.

source

julia> using Temporal, Statistics, Dates

julia> X = TS(randn(100, 4))
100x4 TS{Float64,Dates.Date}: 2018-12-09 to 2019-03-18

Index       A        B        C        D
2018-12-09  -0.916   0.7587   -1.3964  0.2799
2018-12-10  -0.9084  1.1872   0.8103   1.2296
2018-12-11  0.0885   -0.2209  1.3596   -0.8114
2018-12-12  -0.1008  -2.7561  -1.111   0.2708
2018-12-13  -0.1872  1.4361   -1.3869  -0.068
2018-12-14  1.6654   1.3934   0.4215   -1.0274
2018-12-15  -0.6294  1.2857   0.3907   -0.1297
⋮
2019-03-11  0.0348   -0.4745  -1.2021  0.5111
2019-03-12  -0.0296  0.7676   -0.9134  0.6325
2019-03-13  0.7523   -1.3058  -0.4922  -0.5569
2019-03-14  -0.489   0.2964   -0.9981  -1.8601
2019-03-15  -0.4425  2.4969   0.1003   0.4871
2019-03-16  0.5056   0.3984   -0.2271  0.7203
2019-03-17  -0.1927  -1.6219  0.0416   -0.6728
2019-03-18  1.0539   0.1177   0.7381   1.1237

julia> collapse(X, eom, fun=mean)
3x4 TS{Float64,Dates.Date}: 2018-12-31 to 2019-02-28

Index       A        B        C        D
2018-12-31  0.1497   0.2816   -0.2014  0.1677
2019-01-31  -0.0069  -0.2952  -0.1764  -0.0315
2019-02-28  0.0015   0.115    0.0593   -0.1231

julia> last_month = string(X.index[end])[1:7]
"2019-03"

julia> mean(X[last_month])
-0.007916687838570254

Summarizing

Temporal.apply — Function.

apply(x::TS{V}, dim::Int=1; fun=sum) where {V}

Apply function fun across dimension dim of a time series x.

If dim is 1, then apply fun to each row of x, returning a time series (TS) object of the results

If dim is 2, then apply fun to each column of x, returning an Array of the results

source

julia> using Temporal, Statistics

julia> X = TS(randn(100, 4))
100x4 TS{Float64,Dates.Date}: 2018-12-09 to 2019-03-18

Index       A        B        C        D
2018-12-09  -0.1279  0.9044   0.5279   -0.8471
2018-12-10  1.5493   -0.2482  -0.5599  1.1794
2018-12-11  -0.5904  -0.0858  0.0965   -0.0757
2018-12-12  -2.4017  -0.9037  -1.3738  1.5787
2018-12-13  0.8954   0.8432   -0.2799  0.0053
2018-12-14  -1.0709  0.4056   0.9473   0.3587
2018-12-15  -0.3267  -0.0721  -0.2727  -0.9444
⋮
2019-03-11  -0.2576  -0.6814  1.5383   1.4623
2019-03-12  -0.188   -0.8057  1.2761   1.1214
2019-03-13  -0.0166  3.0671   2.1328   -0.6898
2019-03-14  -0.5733  -0.1755  0.9702   -0.666
2019-03-15  0.7302   1.5593   -0.9611  -0.7578
2019-03-16  1.6863   1.473    -2.3154  -1.0308
2019-03-17  0.4832   0.4971   -0.7074  -1.5338
2019-03-18  0.7634   -1.1184  -0.9521  0.0391

julia> apply(X, 1, fun=sum)
100x1 TS{Float64,Dates.Date}: 2018-12-09 to 2019-03-18

Index       Sum
2018-12-09  0.4573
2018-12-10  1.9206
2018-12-11  -0.6555
2018-12-12  -3.1004
2018-12-13  1.464
2018-12-14  0.6406
2018-12-15  -1.6159
⋮
2019-03-11  2.0615
2019-03-12  1.4038
2019-03-13  4.4934
2019-03-14  -0.4446
2019-03-15  0.5707
2019-03-16  -0.1869
2019-03-17  -1.2609
2019-03-18  -1.268

julia> apply(X, 2, fun=sum)
1×4 LinearAlgebra.Adjoint{Float64,Array{Float64,1}}:
 10.4557  -0.312881  15.1183  10.3255