Combining

Joins

Outer Joins

One can perform a full outer join on the time indexes of two TS objects $x$ and $y$ in the following ways:

merge(x, y)

ojoin(x, y)

[x y]

hcat(x, y)

Where there are dates in the index of one that do not exist in the other, values will be filled with NaN objects. As the missing functionality matures in Julia's base syntax, it will eventually replace NaN in this context, since unfortunately NaN is only applicable for Float64 element types.

Base.merge — Function.

merge(x::TS,y::TS;join::Char='o')::TS

Merge two time series objects together by index with an optionally specified join type parameter.

...

Arguments

x::TS: Left side of the join.

y::TS: Right side of the join.

Optional args:

join::Char='o'::TS: Specifies the logic used to perform the merge, and may take on the values 'o' (outer join), 'i' (inner join), 'l' (left join), or 'r' (right join). Defaults to outer join, whose result is the same as hcat(x, y) or [x y].

...

source

Temporal.ojoin — Function.

ojoin(x::TS,y::TS)::TS

Outer join two TS objects by index.

Equivalent to x OUTER JOIN y ON x.index = y.index.

...

Arguments

x::TS: Left side of the join.

y::TS: Right side of the join.

...

source

Example

julia> x = TS(rand(252))
252x1 TS{Float64,Dates.Date}: 2019-02-22 to 2019-10-31

Index       A
2019-02-22  0.4709
2019-02-23  0.7173
2019-02-24  0.929
2019-02-25  0.2008
2019-02-26  0.698
2019-02-27  0.0843
2019-02-28  0.1383
⋮
2019-10-24  0.3034
2019-10-25  0.872
2019-10-26  0.8789
2019-10-27  0.8131
2019-10-28  0.1707
2019-10-29  0.0077
2019-10-30  0.7711
2019-10-31  0.4758

julia> y = TS(rand(252), x.index .- Month(6))
252x1 TS{Float64,Dates.Date}: 2018-08-22 to 2019-04-30

Index       A
2018-08-22  0.5
2018-08-23  0.9022
2018-08-24  0.9373
2018-08-25  0.4267
2018-08-26  0.6933
2018-08-27  0.4479
2018-08-28  0.8178
⋮
2019-04-24  0.434
2019-04-25  0.2474
2019-04-26  0.3573
2019-04-27  0.6212
2019-04-28  0.4275
2019-04-29  0.6416
2019-04-30  0.9795
2019-04-30  0.7951

julia> [x y]
431x2 TS{Float64,Dates.Date}: 2018-08-22 to 2019-10-31

Index       A       A
2018-08-22  NaN     0.5
2018-08-23  NaN     0.9022
2018-08-24  NaN     0.9373
2018-08-25  NaN     0.4267
2018-08-26  NaN     0.6933
2018-08-27  NaN     0.4479
2018-08-28  NaN     0.8178
⋮
2019-10-24  0.3034  NaN
2019-10-25  0.872   NaN
2019-10-26  0.8789  NaN
2019-10-27  0.8131  NaN
2019-10-28  0.1707  NaN
2019-10-29  0.0077  NaN
2019-10-30  0.7711  NaN
2019-10-31  0.4758  NaN

Inner Joins

You can do inner joins on TS objects using the ijoin function, which will remove any observations corresponding to time steps where at least one of the joined objects is missing a row. This will basically keep only the rows where the time index of the left side and the right side intersect.

Temporal.ijoin — Function.

ijoin(x::TS,y::TS)::TS

Inner join two TS objects by index.

Equivalent to x INNER JOIN y on x.index = y.index.

...

Arguments

x::TS: Left side of the join.

y::TS: Right side of the join.

...

source

Example

julia> x = TS(rand(252))
252x1 TS{Float64,Dates.Date}: 2019-02-22 to 2019-10-31

Index       A
2019-02-22  0.7325
2019-02-23  0.2349
2019-02-24  0.0834
2019-02-25  0.4714
2019-02-26  0.6905
2019-02-27  0.7022
2019-02-28  0.2997
⋮
2019-10-24  0.9365
2019-10-25  0.4683
2019-10-26  0.3854
2019-10-27  0.7364
2019-10-28  0.9663
2019-10-29  0.2793
2019-10-30  0.0586
2019-10-31  0.9931

julia> y = TS(rand(252), x.index .- Month(6))
252x1 TS{Float64,Dates.Date}: 2018-08-22 to 2019-04-30

Index       A
2018-08-22  0.9327
2018-08-23  0.3373
2018-08-24  0.7192
2018-08-25  0.7764
2018-08-26  0.8713
2018-08-27  0.2731
2018-08-28  0.1833
⋮
2019-04-24  0.6783
2019-04-25  0.1502
2019-04-26  0.1712
2019-04-27  0.0927
2019-04-28  0.0592
2019-04-29  0.2427
2019-04-30  0.1454
2019-04-30  0.7558

julia> ijoin(x, y)
ERROR: ArgumentError: number of rows of each array must match (got (67, 71))

Left/Right Joins

Left and right joins are performed similarly to inner joins and the typical SQL join queries using the index field each object as the joining key.

Left Join: keep all observations of the left side of the join, fill the right side with NaN's where missing the corresponding time index

Right Join: keep all observations of the right side of the join, fill the left side with NaN's where missing the corresponding time index

Temporal.ljoin — Function.

ljoin(x::TS, y::TS)::TS

Left join two TS objects by index.

Equivalent to x LEFT JOIN y ON x.index = y.index.

...

Arguments

x::TS: Left side of the join.

y::TS: Right side of the join.

...

source

Temporal.rjoin — Function.

rjoin(x::TS, y::TS)::TS

Right join two TS objects by index.

Equivalent to x RIGHT JOIN y ON x.index = y.index.

...

Arguments

x::TS: Left side of the join.

y::TS: Right side of the join.

...

source

Example

julia> x = TS(rand(252))
252x1 TS{Float64,Dates.Date}: 2019-02-22 to 2019-10-31

Index       A
2019-02-22  0.0599
2019-02-23  0.8793
2019-02-24  0.5537
2019-02-25  0.776
2019-02-26  0.2592
2019-02-27  0.1286
2019-02-28  0.9161
⋮
2019-10-24  0.8053
2019-10-25  0.4895
2019-10-26  0.7212
2019-10-27  0.6559
2019-10-28  0.6169
2019-10-29  0.3758
2019-10-30  0.0116
2019-10-31  0.6539

julia> y = TS(rand(252), x.index .- Month(6))
252x1 TS{Float64,Dates.Date}: 2018-08-22 to 2019-04-30

Index       A
2018-08-22  0.6687
2018-08-23  0.5618
2018-08-24  0.4914
2018-08-25  0.9989
2018-08-26  0.7093
2018-08-27  0.682
2018-08-28  0.0983
⋮
2019-04-24  0.2081
2019-04-25  0.4179
2019-04-26  0.7255
2019-04-27  0.2485
2019-04-28  0.1703
2019-04-29  0.4803
2019-04-30  0.441
2019-04-30  0.4178

julia> ljoin(x, y)
252x2 TS{Float64,Dates.Date}: 2019-02-22 to 2019-10-31

Index       A       A
2019-02-22  0.0599  0.9952
2019-02-23  0.8793  0.1679
2019-02-24  0.5537  0.121
2019-02-25  0.776   0.5999
2019-02-26  0.2592  0.2562
2019-02-27  0.1286  0.7764
2019-02-28  0.9161  0.8445
⋮
2019-10-24  0.8053  NaN
2019-10-25  0.4895  NaN
2019-10-26  0.7212  NaN
2019-10-27  0.6559  NaN
2019-10-28  0.6169  NaN
2019-10-29  0.3758  NaN
2019-10-30  0.0116  NaN
2019-10-31  0.6539  NaN

julia> rjoin(x, y)
246x2 TS{Float64,Dates.Date}: 2018-08-22 to 2019-04-30

Index       A       A
2018-08-22  NaN     0.6687
2018-08-23  NaN     0.5618
2018-08-24  NaN     0.4914
2018-08-25  NaN     0.9989
2018-08-26  NaN     0.7093
2018-08-27  NaN     0.682
2018-08-28  NaN     0.0983
⋮
2019-04-23  0.0148  0.4893
2019-04-24  0.5372  0.108
2019-04-25  0.3635  0.8487
2019-04-26  0.1542  0.5338
2019-04-27  0.9407  0.0312
2019-04-28  0.9111  0.7765
2019-04-29  0.4999  0.2081
2019-04-30  0.5692  0.4179