Aggregation

Aggregation

Sampling

Weekdays

Temporal.mondays — Function.

mondays(x::TS)

Return a time series containing all observations occuring on Mondays of the given TS object.

source

Temporal.tuesdays — Function.

tuesdays(x::TS)

Return a time series containing all observations occuring on Tuesdays of the given TS object.

source

Temporal.wednesdays — Function.

wednesdays(x::TS)

Return a time series containing all observations occuring on Wednesdays of the given TS object.

source

Temporal.thursdays — Function.

thursdays(x::TS)

Return a time series containing all observations occuring on Thursdays of the given TS object.

source

Temporal.fridays — Function.

fridays(x::TS)

Return a time series containing all observations occuring on Fridays of the given TS object.

source

Temporal.saturdays — Function.

saturdays(x::TS)

Return a time series containing all observations occuring on Saturdays of the given TS object.

source

Temporal.sundays — Function.

sundays(x::TS)

Return a time series containing all observations occuring on Sundays of the given TS object.

source

Interval Boundaries

Weekly

Temporal.bow — Function.

bow(x::TS; cal::Bool=false)

Return a time series containing all observations occuring at the beginnings of the weeks of the input.

If cal is false, only observations occurring the last calendar day of the week are returned

If cal is true, all observation for which the previous index is a prior week are returned

source

Temporal.eow — Function.

eow(x::TS; cal::Bool=false)

Return a time series containing all observations occuring at the ends of the weeks of the input.

If cal is false, only observations occurring the last calendar day of the week are returned

If cal is true, all observation for which the next index is a new week are returned

source

Monthly

Temporal.bom — Function.

bom(x::TS; cal::Bool=false)

Return a time series containing all observations occuring at the beginnings of the months of the input.

If cal is false, only observations occurring the last calendar day of the month are returned

If cal is true, all observation for which the previous index is a prior month are returned

source

Temporal.eom — Function.

eom(x::TS; cal::Bool=false)

Return a time series containing all observations occuring at the ends of the months of the input.

If cal is false, only observations occurring the last calendar day of the month are returned

If cal is true, all observation for which the next index is a new month are returned

source

Quarterly

Temporal.boq — Function.

boq(x::TS; cal::Bool=false)

Return a time series containing all observations occuring at the beginnings of the quarters of the input.

If cal is false, only observations occurring the last calendar day of the quarter are returned

If cal is true, all observation for which the previous index is a prior quarter are returned

source

Temporal.eoq — Function.

eoq(x::TS; cal::Bool=false)

Return a time series containing all observations occuring at the ends of the quarters of the input.

If cal is false, only observations occurring the last calendar day of the quarter are returned

If cal is true, all observation for which the next index is a new quarter are returned

source

Yearly

Temporal.boy — Function.

boy(x::TS; cal::Bool=false)

Return a time series containing all observations occuring at the beginnings of the years of the input.

If cal is false, only observations occurring the last calendar day of the year are returned

If cal is true, all observation for which the previous index is a prior year are returned

source

Temporal.eoy — Function.

eoy(x::TS; cal::Bool=false)

Return a time series containing all observations occuring at the ends of the years of the input.

If cal is false, only observations occurring the last calendar day of the year are returned

If cal is true, all observation for which the next index is a new year are returned

source

Collapsing

Temporal.collapse — Function.

collapse(x::TS{V,T}, at::Function; fun::Function=last, args...)::TS{V,T} where {V,T}

Compute the function fun over period boundaries defined by at (e.g. eom, boq, etc.) and returng a time series of the results.

Keyword arguments (args...) are passed to the function fun.

source

julia> using Temporal, Statistics, Dates

julia> X = TS(randn(100, 4))
100x4 TS{Float64,Dates.Date}: 2018-11-27 to 2019-03-06

Index       A        B        C        D
2018-11-27  -0.6731  0.1346   0.7062   1.4492
2018-11-28  0.5073   -1.7963  2.4987   0.5123
2018-11-29  1.5402   -0.6211  0.5895   -1.1703
2018-11-30  -0.4879  -1.1817  1.7008   0.1307
2018-12-01  0.0103   0.5472   -0.4373  -2.4079
2018-12-02  1.7969   0.9322   0.935    1.4514
2018-12-03  0.0204   -0.8405  -0.2689  -0.4435
⋮
2019-02-27  -0.026   0.7657   -2.0603  -0.7765
2019-02-28  -1.1533  0.4172   0.3868   0.6612
2019-03-01  -0.5636  -0.2072  0.4729   0.8722
2019-03-02  0.7491   -0.7273  1.5426   -0.3582
2019-03-03  -0.3273  -0.3109  0.1989   -2.6922
2019-03-04  0.1326   0.3203   0.6539   0.1478
2019-03-05  -0.208   -0.0648  0.2933   -1.3519
2019-03-06  0.3187   0.1668   -0.2458  -0.9865

julia> collapse(X, eom, fun=mean)
4x4 TS{Float64,Dates.Date}: 2018-11-30 to 2019-02-28

Index       A       B        C        D
2018-11-30  0.2216  -0.8661  1.3738   0.2305
2018-12-31  0.0258  0.196    -0.1157  0.0271
2019-01-31  0.1489  -0.0118  -0.4069  -0.3698
2019-02-28  0.0498  0.215    -0.1967  0.1226

julia> last_month = string(X.index[end])[1:7]
"2019-03"

julia> mean(X[last_month])
-0.0906034742617361

Summarizing

Temporal.apply — Function.

apply(x::TS{V}, dim::Int=1; fun=sum) where {V}

Apply function fun across dimension dim of a time series x.

If dim is 1, then apply fun to each row of x, returning a time series (TS) object of the results

If dim is 2, then apply fun to each column of x, returning an Array of the results

source

julia> using Temporal, Statistics

julia> X = TS(randn(100, 4))
100x4 TS{Float64,Dates.Date}: 2018-11-27 to 2019-03-06

Index       A        B        C        D
2018-11-27  -0.2595  0.1053   -0.8226  1.5342
2018-11-28  -0.361   0.4994   0.3487   -0.3253
2018-11-29  0.7563   -1.3519  -0.366   -1.1203
2018-11-30  0.2574   0.2553   1.4209   3.1548
2018-12-01  1.5216   0.3131   0.893    -1.391
2018-12-02  -0.5936  0.1432   2.2293   -0.4975
2018-12-03  0.769    -1.269   0.5974   1.4941
⋮
2019-02-27  -0.4137  -0.6026  -1.3186  -1.1843
2019-02-28  0.9433   -0.3733  1.2525   1.3248
2019-03-01  -0.4346  -1.0076  0.702    -0.9866
2019-03-02  1.7533   -1.6619  0.205    -0.0934
2019-03-03  0.5781   0.9346   -1.2177  1.1863
2019-03-04  -0.2898  0.8693   0.8615   -0.4625
2019-03-05  0.1942   0.1419   -0.3136  -0.0803
2019-03-06  -0.0304  1.0068   1.2888   1.5595

julia> apply(X, 1, fun=sum)
100x1 TS{Float64,Dates.Date}: 2018-11-27 to 2019-03-06

Index       Sum
2018-11-27  0.5574
2018-11-28  0.1618
2018-11-29  -2.0818
2018-11-30  5.0884
2018-12-01  1.3367
2018-12-02  1.2815
2018-12-03  1.5915
⋮
2019-02-27  -3.5193
2019-02-28  3.1472
2019-03-01  -1.7268
2019-03-02  0.203
2019-03-03  1.4813
2019-03-04  0.9785
2019-03-05  -0.0579
2019-03-06  3.8247

julia> apply(X, 2, fun=sum)
1×4 LinearAlgebra.Adjoint{Float64,Array{Float64,1}}:
 12.5648  -12.7102  17.0325  11.6794