Aggregation

Aggregation

Sampling

Weekdays

Temporal.mondays — Function.

mondays(x::TS)

Return a time series containing all observations occuring on Mondays of the given TS object.

source

Temporal.tuesdays — Function.

tuesdays(x::TS)

Return a time series containing all observations occuring on Tuesdays of the given TS object.

source

Temporal.wednesdays — Function.

wednesdays(x::TS)

Return a time series containing all observations occuring on Wednesdays of the given TS object.

source

Temporal.thursdays — Function.

thursdays(x::TS)

Return a time series containing all observations occuring on Thursdays of the given TS object.

source

Temporal.fridays — Function.

fridays(x::TS)

Return a time series containing all observations occuring on Fridays of the given TS object.

source

Temporal.saturdays — Function.

saturdays(x::TS)

Return a time series containing all observations occuring on Saturdays of the given TS object.

source

Temporal.sundays — Function.

sundays(x::TS)

Return a time series containing all observations occuring on Sundays of the given TS object.

source

Interval Boundaries

Weekly

Temporal.bow — Function.

bow(x::TS; cal::Bool=false)

Return a time series containing all observations occuring at the beginnings of the weeks of the input.

If cal is false, only observations occurring the last calendar day of the week are returned

If cal is true, all observation for which the previous index is a prior week are returned

source

Temporal.eow — Function.

eow(x::TS; cal::Bool=false)

Return a time series containing all observations occuring at the ends of the weeks of the input.

If cal is false, only observations occurring the last calendar day of the week are returned

If cal is true, all observation for which the next index is a new week are returned

source

Monthly

Temporal.bom — Function.

bom(x::TS; cal::Bool=false)

Return a time series containing all observations occuring at the beginnings of the months of the input.

If cal is false, only observations occurring the last calendar day of the month are returned

If cal is true, all observation for which the previous index is a prior month are returned

source

Temporal.eom — Function.

eom(x::TS; cal::Bool=false)

Return a time series containing all observations occuring at the ends of the months of the input.

If cal is false, only observations occurring the last calendar day of the month are returned

If cal is true, all observation for which the next index is a new month are returned

source

Quarterly

Temporal.boq — Function.

boq(x::TS; cal::Bool=false)

Return a time series containing all observations occuring at the beginnings of the quarters of the input.

If cal is false, only observations occurring the last calendar day of the quarter are returned

If cal is true, all observation for which the previous index is a prior quarter are returned

source

Temporal.eoq — Function.

eoq(x::TS; cal::Bool=false)

Return a time series containing all observations occuring at the ends of the quarters of the input.

If cal is false, only observations occurring the last calendar day of the quarter are returned

If cal is true, all observation for which the next index is a new quarter are returned

source

Yearly

Temporal.boy — Function.

boy(x::TS; cal::Bool=false)

Return a time series containing all observations occuring at the beginnings of the years of the input.

If cal is false, only observations occurring the last calendar day of the year are returned

If cal is true, all observation for which the previous index is a prior year are returned

source

Temporal.eoy — Function.

eoy(x::TS; cal::Bool=false)

Return a time series containing all observations occuring at the ends of the years of the input.

If cal is false, only observations occurring the last calendar day of the year are returned

If cal is true, all observation for which the next index is a new year are returned

source

Collapsing

Temporal.collapse — Function.

collapse(x::TS{V,T}, at::Function; fun::Function=last, args...)::TS{V,T} where {V,T}

Compute the function fun over period boundaries defined by at (e.g. eom, boq, etc.) and returng a time series of the results.

Keyword arguments (args...) are passed to the function fun.

source

julia> using Temporal, Statistics, Dates

julia> X = TS(randn(100, 4))
100x4 TS{Float64,Dates.Date}: 2018-11-06 to 2019-02-13

Index       A        B        C        D
2018-11-06  -2.6208  -0.0603  0.8877   -0.1903
2018-11-07  -0.5864  -1.1742  -0.3737  -0.424
2018-11-08  -0.0315  0.7758   0.0461   -0.7151
2018-11-09  0.9377   -0.8427  -0.8721  1.3059
2018-11-10  0.236    -0.4837  1.8588   -0.2221
2018-11-11  -0.3528  -0.1006  0.9167   -0.2889
2018-11-12  0.5351   0.004    0.1236   -2.0515
⋮
2019-02-06  0.6383   1.2161   0.2465   0.2234
2019-02-07  -2.1643  0.5972   -0.8305  -1.0838
2019-02-08  0.0464   1.2708   0.4423   -0.7181
2019-02-09  0.82     0.0411   -1.6112  -1.6904
2019-02-10  -1.7713  -0.2343  0.2268   1.1905
2019-02-11  1.0498   0.9486   -0.6322  -0.3334
2019-02-12  0.5614   0.8548   0.2865   -0.5864
2019-02-13  1.1065   1.1769   0.1779   0.8802

julia> collapse(X, eom, fun=mean)
3x4 TS{Float64,Dates.Date}: 2018-11-30 to 2019-01-31

Index       A        B        C        D
2018-11-30  0.0816   -0.0409  -0.2302  -0.2463
2018-12-31  0.1231   -0.0822  -0.0365  0.0054
2019-01-31  -0.1538  0.1271   0.0289   0.0484

julia> last_month = string(X.index[end])[1:7]
"2019-02"

julia> mean(X[last_month])
0.05907557763097778

Summarizing

Temporal.apply — Function.

apply(x::TS{V}, dim::Int=1; fun=sum) where {V}

Apply function fun across dimension dim of a time series x.

If dim is 1, then apply fun to each row of x, returning a time series (TS) object of the results

If dim is 2, then apply fun to each column of x, returning an Array of the results

source

julia> using Temporal, Statistics

julia> X = TS(randn(100, 4))
100x4 TS{Float64,Dates.Date}: 2018-11-06 to 2019-02-13

Index       A        B        C        D
2018-11-06  1.176    1.131    -1.8598  -0.3221
2018-11-07  -1.0676  1.2759   0.3486   1.9824
2018-11-08  0.748    -0.0823  0.7391   -0.0281
2018-11-09  0.6144   1.3055   -0.2876  1.3244
2018-11-10  0.0753   0.0525   -2.1742  2.5417
2018-11-11  -0.5322  2.4583   -1.0097  1.834
2018-11-12  -1.1824  -0.5438  -0.3359  -1.0095
⋮
2019-02-06  1.6089   -0.0236  0.1979   -2.0259
2019-02-07  0.4082   -1.1552  0.0236   -0.515
2019-02-08  -1.2663  -0.7979  -0.5236  -1.2526
2019-02-09  -0.0337  0.1609   0.7652   -1.4046
2019-02-10  -0.8257  1.662    -0.0847  0.2329
2019-02-11  -0.9888  -0.4772  0.3051   1.5381
2019-02-12  -1.8402  -0.1622  -0.2629  0.567
2019-02-13  -0.9044  -0.818   -0.1646  -0.0634

julia> apply(X, 1, fun=sum)
100x1 TS{Float64,Dates.Date}: 2018-11-06 to 2019-02-13

Index       Sum
2018-11-06  0.1252
2018-11-07  2.5393
2018-11-08  1.3766
2018-11-09  2.9567
2018-11-10  0.4953
2018-11-11  2.7504
2018-11-12  -3.0717
⋮
2019-02-06  -0.2426
2019-02-07  -1.2384
2019-02-08  -3.8405
2019-02-09  -0.5122
2019-02-10  0.9845
2019-02-11  0.3772
2019-02-12  -1.6983
2019-02-13  -1.9505

julia> apply(X, 2, fun=sum)
1×4 LinearAlgebra.Adjoint{Float64,Array{Float64,1}}:
 -2.46747  6.87184  -11.6787  0.745827