Aggregation

Aggregation

Sampling

Weekdays

Temporal.mondays — Function.

mondays(x::TS)

Return a time series containing all observations occuring on Mondays of the given TS object.

source

Temporal.tuesdays — Function.

tuesdays(x::TS)

Return a time series containing all observations occuring on Tuesdays of the given TS object.

source

Temporal.wednesdays — Function.

wednesdays(x::TS)

Return a time series containing all observations occuring on Wednesdays of the given TS object.

source

Temporal.thursdays — Function.

thursdays(x::TS)

Return a time series containing all observations occuring on Thursdays of the given TS object.

source

Temporal.fridays — Function.

fridays(x::TS)

Return a time series containing all observations occuring on Fridays of the given TS object.

source

Temporal.saturdays — Function.

saturdays(x::TS)

Return a time series containing all observations occuring on Saturdays of the given TS object.

source

Temporal.sundays — Function.

sundays(x::TS)

Return a time series containing all observations occuring on Sundays of the given TS object.

source

Interval Boundaries

Weekly

Temporal.bow — Function.

bow(x::TS; cal::Bool=false)

Return a time series containing all observations occuring at the beginnings of the weeks of the input.

If cal is false, only observations occurring the last calendar day of the week are returned

If cal is true, all observation for which the previous index is a prior week are returned

source

Temporal.eow — Function.

eow(x::TS; cal::Bool=false)

Return a time series containing all observations occuring at the ends of the weeks of the input.

If cal is false, only observations occurring the last calendar day of the week are returned

If cal is true, all observation for which the next index is a new week are returned

source

Monthly

Temporal.bom — Function.

bom(x::TS; cal::Bool=false)

Return a time series containing all observations occuring at the beginnings of the months of the input.

If cal is false, only observations occurring the last calendar day of the month are returned

If cal is true, all observation for which the previous index is a prior month are returned

source

Temporal.eom — Function.

eom(x::TS; cal::Bool=false)

Return a time series containing all observations occuring at the ends of the months of the input.

If cal is false, only observations occurring the last calendar day of the month are returned

If cal is true, all observation for which the next index is a new month are returned

source

Quarterly

Temporal.boq — Function.

boq(x::TS; cal::Bool=false)

Return a time series containing all observations occuring at the beginnings of the quarters of the input.

If cal is false, only observations occurring the last calendar day of the quarter are returned

If cal is true, all observation for which the previous index is a prior quarter are returned

source

Temporal.eoq — Function.

eoq(x::TS; cal::Bool=false)

Return a time series containing all observations occuring at the ends of the quarters of the input.

If cal is false, only observations occurring the last calendar day of the quarter are returned

If cal is true, all observation for which the next index is a new quarter are returned

source

Yearly

Temporal.boy — Function.

boy(x::TS; cal::Bool=false)

Return a time series containing all observations occuring at the beginnings of the years of the input.

If cal is false, only observations occurring the last calendar day of the year are returned

If cal is true, all observation for which the previous index is a prior year are returned

source

Temporal.eoy — Function.

eoy(x::TS; cal::Bool=false)

Return a time series containing all observations occuring at the ends of the years of the input.

If cal is false, only observations occurring the last calendar day of the year are returned

If cal is true, all observation for which the next index is a new year are returned

source

Collapsing

Temporal.collapse — Function.

collapse(x::TS{V,T}, at::Function; fun::Function=last, args...)::TS{V,T} where {V,T}

Compute the function fun over period boundaries defined by at (e.g. eom, boq, etc.) and returng a time series of the results.

Keyword arguments (args...) are passed to the function fun.

source

julia> using Temporal, Statistics, Dates

julia> X = TS(randn(100, 4))
100x4 TS{Float64,Dates.Date}: 2018-10-23 to 2019-01-30

Index       A        B        C        D
2018-10-23  1.8883   0.496    2.3985   -1.6618
2018-10-24  1.2996   -0.9074  -0.2786  -0.3347
2018-10-25  0.4561   0.4105   -0.2344  1.106
2018-10-26  -1.1397  -1.8783  -2.6248  -0.4542
2018-10-27  0.4185   -0.4355  -0.9554  0.4246
2018-10-28  -1.0863  -0.9992  0.9672   1.7586
2018-10-29  -0.9111  -0.0147  0.3486   0.0555
2018-10-30  -0.4394  0.4988   -0.5474  0.1292
⋮
2019-01-21  -0.1026  -0.217   1.4015   0.203    2019-01-22  -0.8605  -0.0648  -1.2638  0.2904   2019-01-23  -0.5835  0.5062   -0.5882  -0.306   2019-01-24  0.1913   -0.3041  0.3578   -0.2709  2019-01-25  -0.7285  0.0014   -0.4318  -1.5897  2019-01-26  -0.5003  -0.0161  0.2527   -0.0241  2019-01-27  -0.0721  0.0672   0.4283   1.1654   2019-01-28  -0.7243  -0.3988  0.8395   -0.0273  2019-01-29  -0.8485  -0.2328  0.0194   0.7768   2019-01-30  -0.6706  -1.2092  -1.3549  1.1918

julia> collapse(X, eom, fun=mean)
3x4 TS{Float64,Dates.Date}: 2018-10-31 to 2018-12-31

Index       A        B        C        D
2018-10-31  -0.1403  -0.4519  -0.0487  0.2887
2018-11-30  0.166    0.2189   -0.0766  -0.2112
2018-12-31  0.0636   0.2229   0.129    0.1283

julia> last_month = string(X.index[end])[1:7]
"2019-01"

julia> mean(X[last_month])
-0.05313333759147694

Summarizing

Temporal.apply — Function.

apply(x::TS{V}, dim::Int=1; fun=sum) where {V}

Apply function fun across dimension dim of a time series x.

If dim is 1, then apply fun to each row of x, returning a time series (TS) object of the results

If dim is 2, then apply fun to each column of x, returning an Array of the results

source

julia> using Temporal, Statistics

julia> X = TS(randn(100, 4))
100x4 TS{Float64,Dates.Date}: 2018-10-23 to 2019-01-30

Index       A        B        C        D
2018-10-23  -0.1712  -1.5096  -1.8568  -0.1379
2018-10-24  -0.2288  0.2315   -0.0761  -0.7843
2018-10-25  0.2609   -0.279   1.4975   -0.187
2018-10-26  -0.5216  -1.3988  0.9722   -0.7697
2018-10-27  0.2453   -0.3347  0.6025   -0.3401
2018-10-28  0.4018   0.8113   -1.0861  1.679
2018-10-29  0.4596   -0.3099  -0.9664  -1.0645
2018-10-30  0.2345   0.3815   0.5911   0.2381
⋮
2019-01-21  1.8639   -0.0588  -0.9395  -1.6198  2019-01-22  -0.0493  1.1845   0.8087   -0.2462  2019-01-23  -0.8448  0.5899   -0.4832  1.4424   2019-01-24  0.2904   0.4164   0.9846   -0.4985  2019-01-25  -1.84    -0.0586  -1.5735  0.6774   2019-01-26  -2.5544  -1.087   0.5415   1.4042   2019-01-27  -0.8004  -0.9345  0.3      -0.7965  2019-01-28  1.5857   -1.1992  -0.5144  -0.1173  2019-01-29  1.2906   -0.5472  0.878    1.3264   2019-01-30  -0.6025  1.1531   -0.108   -0.3425

julia> apply(X, 1, fun=sum)
100x1 TS{Float64,Dates.Date}: 2018-10-23 to 2019-01-30

Index       Sum
2018-10-23  -3.6755
2018-10-24  -0.8577
2018-10-25  1.2924
2018-10-26  -1.7179
2018-10-27  0.173
2018-10-28  1.806
2018-10-29  -1.8812
2018-10-30  1.4452
⋮
2019-01-21  -0.7543  2019-01-22  1.6977   2019-01-23  0.7043   2019-01-24  1.1928   2019-01-25  -2.7946  2019-01-26  -1.6957  2019-01-27  -2.2314  2019-01-28  -0.2453  2019-01-29  2.9479   2019-01-30  0.1002

julia> apply(X, 2, fun=sum)
1×4 LinearAlgebra.Adjoint{Float64,Array{Float64,1}}:
 9.83025  9.86977  -16.7448  -1.67304