Aggregation

Aggregation

Sampling

Weekdays

Temporal.mondays — Function.

mondays(x::TS)

Return a time series containing all observations occuring on Mondays of the given TS object.

source

Temporal.tuesdays — Function.

tuesdays(x::TS)

Return a time series containing all observations occuring on Tuesdays of the given TS object.

source

Temporal.wednesdays — Function.

wednesdays(x::TS)

Return a time series containing all observations occuring on Wednesdays of the given TS object.

source

Temporal.thursdays — Function.

thursdays(x::TS)

Return a time series containing all observations occuring on Thursdays of the given TS object.

source

Temporal.fridays — Function.

fridays(x::TS)

Return a time series containing all observations occuring on Fridays of the given TS object.

source

Temporal.saturdays — Function.

saturdays(x::TS)

Return a time series containing all observations occuring on Saturdays of the given TS object.

source

Temporal.sundays — Function.

sundays(x::TS)

Return a time series containing all observations occuring on Sundays of the given TS object.

source

Interval Boundaries

Weekly

Temporal.bow — Function.

bow(x::TS; cal::Bool=false)

Return a time series containing all observations occuring at the beginnings of the weeks of the input.

If cal is false, only observations occurring the last calendar day of the week are returned

If cal is true, all observation for which the previous index is a prior week are returned

source

Temporal.eow — Function.

eow(x::TS; cal::Bool=false)

Return a time series containing all observations occuring at the ends of the weeks of the input.

If cal is false, only observations occurring the last calendar day of the week are returned

If cal is true, all observation for which the next index is a new week are returned

source

Monthly

Temporal.bom — Function.

bom(x::TS; cal::Bool=false)

Return a time series containing all observations occuring at the beginnings of the months of the input.

If cal is false, only observations occurring the last calendar day of the month are returned

If cal is true, all observation for which the previous index is a prior month are returned

source

Temporal.eom — Function.

eom(x::TS; cal::Bool=false)

Return a time series containing all observations occuring at the ends of the months of the input.

If cal is false, only observations occurring the last calendar day of the month are returned

If cal is true, all observation for which the next index is a new month are returned

source

Quarterly

Temporal.boq — Function.

boq(x::TS; cal::Bool=false)

Return a time series containing all observations occuring at the beginnings of the quarters of the input.

If cal is false, only observations occurring the last calendar day of the quarter are returned

If cal is true, all observation for which the previous index is a prior quarter are returned

source

Temporal.eoq — Function.

eoq(x::TS; cal::Bool=false)

Return a time series containing all observations occuring at the ends of the quarters of the input.

If cal is false, only observations occurring the last calendar day of the quarter are returned

If cal is true, all observation for which the next index is a new quarter are returned

source

Yearly

Temporal.boy — Function.

boy(x::TS; cal::Bool=false)

Return a time series containing all observations occuring at the beginnings of the years of the input.

If cal is false, only observations occurring the last calendar day of the year are returned

If cal is true, all observation for which the previous index is a prior year are returned

source

Temporal.eoy — Function.

eoy(x::TS; cal::Bool=false)

Return a time series containing all observations occuring at the ends of the years of the input.

If cal is false, only observations occurring the last calendar day of the year are returned

If cal is true, all observation for which the next index is a new year are returned

source

Collapsing

Temporal.collapse — Function.

collapse(x::TS{V,T}, at::Function; fun::Function=last, args...)::TS{V,T} where {V,T}

Compute the function fun over period boundaries defined by at (e.g. eom, boq, etc.) and returng a time series of the results.

Keyword arguments (args...) are passed to the function fun.

source

julia> using Temporal, Statistics, Dates

julia> X = TS(randn(100, 4))
100x4 TS{Float64,Dates.Date}: 2018-10-23 to 2019-01-30

Index       A        B        C        D
2018-10-23  -0.965   1.3209   0.4198   0.0996
2018-10-24  -1.5534  1.1332   -0.7142  2.0123
2018-10-25  0.461    -0.6808  -1.0596  0.6537
2018-10-26  0.4762   -0.0714  -0.6212  -0.3001
2018-10-27  -1.2174  -0.217   0.4545   -1.8673
2018-10-28  1.0315   0.4587   0.8628   -0.7029
2018-10-29  -0.0595  -0.2417  -0.5597  -0.7231
2018-10-30  -1.5597  -1.3313  0.8635   -0.6808
⋮
2019-01-21  0.4187   -0.2005  -1.6102  -0.5862  2019-01-22  0.8004   -0.0887  -1.352   0.0981   2019-01-23  -0.1211  -1.0679  -0.1864  0.544    2019-01-24  1.406    0.7572   -1.5147  0.2481   2019-01-25  0.3575   1.7521   -0.782   -0.3215  2019-01-26  0.3428   0.2937   0.9249   -1.2412  2019-01-27  -1.3142  -0.4831  1.9267   0.6605   2019-01-28  0.3421   -0.3814  -0.1341  0.8516   2019-01-29  0.913    -0.0417  -0.3306  -0.4301  2019-01-30  -2.5396  -0.8181  1.4587   -0.6134

julia> collapse(X, eom, fun=mean)
3x4 TS{Float64,Dates.Date}: 2018-10-31 to 2018-12-31

Index       A        B        C        D
2018-10-31  -0.3207  0.1002   -0.1746  0.0011
2018-11-30  -0.1671  -0.1301  -0.0431  -0.1767
2018-12-31  0.3581   -0.0288  0.1475   0.0379

julia> last_month = string(X.index[end])[1:7]
"2019-01"

julia> mean(X[last_month])
-0.08632212695205653

Summarizing

Temporal.apply — Function.

apply(x::TS{V}, dim::Int=1; fun=sum) where {V}

Apply function fun across dimension dim of a time series x.

If dim is 1, then apply fun to each row of x, returning a time series (TS) object of the results

If dim is 2, then apply fun to each column of x, returning an Array of the results

source

julia> using Temporal, Statistics

julia> X = TS(randn(100, 4))
100x4 TS{Float64,Dates.Date}: 2018-10-23 to 2019-01-30

Index       A        B        C        D
2018-10-23  1.1215   -1.4953  -0.2583  -1.0897
2018-10-24  1.2125   0.138    0.5121   0.4036
2018-10-25  -0.755   -0.1681  -0.7977  -0.006
2018-10-26  0.6331   -0.3368  -0.3049  -0.0703
2018-10-27  0.4004   -0.7805  0.0489   -1.0422
2018-10-28  1.6576   1.8386   -0.1904  -0.2526
2018-10-29  0.6249   -1.3129  0.8941   0.6418
2018-10-30  0.3899   -1.6959  -1.0257  1.6126
⋮
2019-01-21  0.2536   0.0613   1.6327   0.2761   2019-01-22  0.0442   1.3242   -0.6617  -0.107   2019-01-23  0.8171   -0.0655  -0.87    -0.056   2019-01-24  1.9758   -1.742   -0.8139  0.9172   2019-01-25  0.489    1.0782   0.991    0.4549   2019-01-26  0.8768   0.0292   -0.0989  -0.3422  2019-01-27  -0.5651  -0.0893  0.7129   1.2386   2019-01-28  -0.0275  -1.0822  0.4905   0.4762   2019-01-29  0.4494   -1.7399  -0.6441  0.4038   2019-01-30  -0.8246  -0.7663  -0.5695  1.0121

julia> apply(X, 1, fun=sum)
100x1 TS{Float64,Dates.Date}: 2018-10-23 to 2019-01-30

Index       Sum
2018-10-23  -1.7218
2018-10-24  2.2662
2018-10-25  -1.7267
2018-10-26  -0.0788
2018-10-27  -1.3734
2018-10-28  3.0532
2018-10-29  0.8479
2018-10-30  -0.7191
⋮
2019-01-21  2.2237   2019-01-22  0.5998   2019-01-23  -0.1744  2019-01-24  0.3371   2019-01-25  3.0131   2019-01-26  0.4649   2019-01-27  1.2971   2019-01-28  -0.1429  2019-01-29  -1.5308  2019-01-30  -1.1483

julia> apply(X, 2, fun=sum)
1×4 LinearAlgebra.Adjoint{Float64,Array{Float64,1}}:
 15.1677  -14.7842  -5.52187  10.2131