Aggregation

Sampling

Weekdays

Temporal.mondays — Function

mondays(x::TS)

Return a time series containing all observations occuring on Mondays of the given TS object.

Temporal.tuesdays — Function

tuesdays(x::TS)

Return a time series containing all observations occuring on Tuesdays of the given TS object.

Temporal.wednesdays — Function

wednesdays(x::TS)

Return a time series containing all observations occuring on Wednesdays of the given TS object.

Temporal.thursdays — Function

thursdays(x::TS)

Return a time series containing all observations occuring on Thursdays of the given TS object.

Temporal.fridays — Function

fridays(x::TS)

Return a time series containing all observations occuring on Fridays of the given TS object.

Temporal.saturdays — Function

saturdays(x::TS)

Return a time series containing all observations occuring on Saturdays of the given TS object.

Temporal.sundays — Function

sundays(x::TS)

Return a time series containing all observations occuring on Sundays of the given TS object.

Interval Boundaries

Weekly

Temporal.bow — Function

bow(x::TS; cal::Bool=false)

Return a time series containing all observations occuring at the beginnings of the weeks of the input.

If cal is false, only observations occurring the last calendar day of the week are returned

If cal is true, all observation for which the previous index is a prior week are returned

Temporal.eow — Function

eow(x::TS; cal::Bool=false)

Return a time series containing all observations occuring at the ends of the weeks of the input.

If cal is false, only observations occurring the last calendar day of the week are returned

If cal is true, all observation for which the next index is a new week are returned

Monthly

Temporal.bom — Function

bom(x::TS; cal::Bool=false)

Return a time series containing all observations occuring at the beginnings of the months of the input.

If cal is false, only observations occurring the last calendar day of the month are returned

If cal is true, all observation for which the previous index is a prior month are returned

Temporal.eom — Function

eom(x::TS; cal::Bool=false)

Return a time series containing all observations occuring at the ends of the months of the input.

If cal is false, only observations occurring the last calendar day of the month are returned

If cal is true, all observation for which the next index is a new month are returned

Quarterly

Temporal.boq — Function

boq(x::TS; cal::Bool=false)

Return a time series containing all observations occuring at the beginnings of the quarters of the input.

If cal is false, only observations occurring the last calendar day of the quarter are returned

If cal is true, all observation for which the previous index is a prior quarter are returned

Temporal.eoq — Function

eoq(x::TS; cal::Bool=false)

Return a time series containing all observations occuring at the ends of the quarters of the input.

If cal is false, only observations occurring the last calendar day of the quarter are returned

If cal is true, all observation for which the next index is a new quarter are returned

Yearly

Temporal.boy — Function

boy(x::TS; cal::Bool=false)

Return a time series containing all observations occuring at the beginnings of the years of the input.

If cal is false, only observations occurring the last calendar day of the year are returned

If cal is true, all observation for which the previous index is a prior year are returned

Temporal.eoy — Function

eoy(x::TS; cal::Bool=false)

Return a time series containing all observations occuring at the ends of the years of the input.

If cal is false, only observations occurring the last calendar day of the year are returned

If cal is true, all observation for which the next index is a new year are returned

Collapsing

Temporal.collapse — Function

collapse(x::TS, at::Function; fun::Function=last, args...)::TS

Compute the function fun over period boundaries defined by at (e.g. eom, boq, etc.) and returng a time series of the results.

Keyword arguments (args...) are passed to the function fun.

julia> using Temporal, Statistics, Dates

julia> X = TS(randn(100, 4))
100x4 TS{Float64,Dates.Date}: 2019-09-10 to 2019-12-18

Index       A        B        C        D
2019-09-10  0.0303   -0.0485  -0.1408  1.1617
2019-09-11  -1.131   -0.8848  -0.3621  0.4302
2019-09-12  -1.6888  0.0301   0.3776   0.0375
2019-09-13  1.0147   0.4387   1.0982   -0.726
2019-09-14  -0.4134  0.0801   -0.1354  1.7397
2019-09-15  0.5132   -1.1566  -0.8389  0.9422
2019-09-16  0.2511   -2.6757  0.3663   -1.4698
⋮
2019-12-11  -0.9296  1.021    -1.7384  -1.082
2019-12-12  0.7794   -1.5124  2.2232   -0.0345
2019-12-13  -0.659   -1.325   -0.3999  2.0488
2019-12-14  -1.7524  1.826    -0.0756  0.0541
2019-12-15  1.6862   0.5086   -1.3191  0.6302
2019-12-16  0.1636   -0.98    -0.4034  -0.1747
2019-12-17  0.2114   0.0901   1.0595   -0.2814
2019-12-18  -0.0112  0.8897   -1.7225  -0.2439

julia> collapse(X, eom, fun=mean)
3x4 TS{Float64,Dates.Date}: 2019-09-30 to 2019-11-30

Index       A       B        C        D
2019-09-30  0.3824  -0.2589  0.1273   0.5438
2019-10-31  0.0294  -0.0308  0.3654   -0.2036
2019-11-30  0.0799  0.0901   -0.0294  -0.1332

julia> last_month = string(X.index[end])[1:7]
"2019-12"

julia> mean(X[last_month])
-0.08660703757238343

Summarizing

Temporal.apply — Function

apply(x::TS, dim::Int=1; fun::Function=sum)

Apply function fun across dimension dim of a time series x.

If dim is 1, then apply fun to each row of x, returning a time series (TS) object of the results

If dim is 2, then apply fun to each column of x, returning an Array of the results

julia> using Temporal, Statistics

julia> X = TS(randn(100, 4))
100x4 TS{Float64,Dates.Date}: 2019-09-10 to 2019-12-18

Index       A        B        C        D
2019-09-10  0.7248   -1.117   -0.9545  -1.5688
2019-09-11  0.8947   -0.1576  0.3626   0.7181
2019-09-12  1.0514   -0.78    -0.37    1.39
2019-09-13  -0.5118  -0.4639  -0.8325  -0.623
2019-09-14  1.2372   0.1562   0.0307   1.0169
2019-09-15  1.8271   -0.9851  0.0516   -0.2334
2019-09-16  1.1535   0.9995   1.6772   -0.0668
⋮
2019-12-11  0.4671   -0.6166  -1.0196  -0.0497
2019-12-12  1.2124   0.4668   0.1104   2.0338
2019-12-13  -0.9594  -0.7124  0.086    -2.068
2019-12-14  -0.2549  -0.6718  0.5209   -0.997
2019-12-15  -0.9002  -0.6137  -0.4606  -1.1939
2019-12-16  1.1445   1.1003   1.1806   -0.4621
2019-12-17  1.4796   -1.9777  1.8941   0.2564
2019-12-18  -1.4252  -0.1872  -2.1918  -0.8414

julia> apply(X, 1, fun=sum)
100x1 TS{Float64,Dates.Date}: 2019-09-10 to 2019-12-18

Index       Sum
2019-09-10  -2.9155
2019-09-11  1.8178
2019-09-12  1.2913
2019-09-13  -2.4313
2019-09-14  2.441
2019-09-15  0.6601
2019-09-16  3.7634
⋮
2019-12-11  -1.2188
2019-12-12  3.8234
2019-12-13  -3.6539
2019-12-14  -1.4028
2019-12-15  -3.1685
2019-12-16  2.9633
2019-12-17  1.6524
2019-12-18  -4.6456

julia> apply(X, 2, fun=sum)
1×4 LinearAlgebra.Adjoint{Float64,Array{Float64,1}}:
 -5.69472  2.80124  1.18759  -7.18014