Aggregation

Aggregation

Sampling

Weekdays

Temporal.mondays — Function.

mondays(x::TS)

Return a time series containing all observations occuring on Mondays of the given TS object.

source

Temporal.tuesdays — Function.

tuesdays(x::TS)

Return a time series containing all observations occuring on Tuesdays of the given TS object.

source

Temporal.wednesdays — Function.

wednesdays(x::TS)

Return a time series containing all observations occuring on Wednesdays of the given TS object.

source

Temporal.thursdays — Function.

thursdays(x::TS)

Return a time series containing all observations occuring on Thursdays of the given TS object.

source

Temporal.fridays — Function.

fridays(x::TS)

Return a time series containing all observations occuring on Fridays of the given TS object.

source

Temporal.saturdays — Function.

saturdays(x::TS)

Return a time series containing all observations occuring on Saturdays of the given TS object.

source

Temporal.sundays — Function.

sundays(x::TS)

Return a time series containing all observations occuring on Sundays of the given TS object.

source

Interval Boundaries

Weekly

Temporal.bow — Function.

bow(x::TS; cal::Bool=false)

Return a time series containing all observations occuring at the beginnings of the weeks of the input.

If cal is false, only observations occurring the last calendar day of the week are returned

If cal is true, all observation for which the previous index is a prior week are returned

source

Temporal.eow — Function.

eow(x::TS; cal::Bool=false)

Return a time series containing all observations occuring at the ends of the weeks of the input.

If cal is false, only observations occurring the last calendar day of the week are returned

If cal is true, all observation for which the next index is a new week are returned

source

Monthly

Temporal.bom — Function.

bom(x::TS; cal::Bool=false)

Return a time series containing all observations occuring at the beginnings of the months of the input.

If cal is false, only observations occurring the last calendar day of the month are returned

If cal is true, all observation for which the previous index is a prior month are returned

source

Temporal.eom — Function.

eom(x::TS; cal::Bool=false)

Return a time series containing all observations occuring at the ends of the months of the input.

If cal is false, only observations occurring the last calendar day of the month are returned

If cal is true, all observation for which the next index is a new month are returned

source

Quarterly

Temporal.boq — Function.

boq(x::TS; cal::Bool=false)

Return a time series containing all observations occuring at the beginnings of the quarters of the input.

If cal is false, only observations occurring the last calendar day of the quarter are returned

If cal is true, all observation for which the previous index is a prior quarter are returned

source

Temporal.eoq — Function.

eoq(x::TS; cal::Bool=false)

Return a time series containing all observations occuring at the ends of the quarters of the input.

If cal is false, only observations occurring the last calendar day of the quarter are returned

If cal is true, all observation for which the next index is a new quarter are returned

source

Yearly

Temporal.boy — Function.

boy(x::TS; cal::Bool=false)

Return a time series containing all observations occuring at the beginnings of the years of the input.

If cal is false, only observations occurring the last calendar day of the year are returned

If cal is true, all observation for which the previous index is a prior year are returned

source

Temporal.eoy — Function.

eoy(x::TS; cal::Bool=false)

Return a time series containing all observations occuring at the ends of the years of the input.

If cal is false, only observations occurring the last calendar day of the year are returned

If cal is true, all observation for which the next index is a new year are returned

source

Collapsing

Temporal.collapse — Function.

collapse(x::TS{V,T}, at::Function; fun::Function=last, args...)::TS{V,T} where {V,T}

Compute the function fun over period boundaries defined by at (e.g. eom, boq, etc.) and returng a time series of the results.

Keyword arguments (args...) are passed to the function fun.

source

julia> using Temporal, Statistics, Dates

julia> X = TS(randn(100, 4))
100x4 TS{Float64,Dates.Date}: 2018-11-27 to 2019-03-06

Index       A        B        C        D
2018-11-27  0.232    -0.0859  -1.0657  -0.5836
2018-11-28  -2.6893  0.5199   0.3309   0.4783
2018-11-29  -0.0906  -0.8511  -1.3011  0.3674
2018-11-30  -0.3303  -1.0763  -0.3605  1.8533
2018-12-01  0.9752   -0.691   0.0349   0.5687
2018-12-02  -1.1468  -1.8698  1.7241   1.2564
2018-12-03  0.2084   -0.4185  -0.2477  1.4239
⋮
2019-02-27  0.651    -0.0656  -0.2938  1.8783
2019-02-28  1.5439   0.4011   0.3198   0.8562
2019-03-01  -0.3631  0.3814   -0.3547  0.555
2019-03-02  -1.5259  0.231    -0.7529  0.7412
2019-03-03  0.4202   -0.591   -1.4832  -0.0115
2019-03-04  0.0632   0.2387   -0.7024  -0.186
2019-03-05  -2.0203  1.4098   -0.0987  -0.3954
2019-03-06  1.3841   1.5465   0.1379   0.3846

julia> collapse(X, eom, fun=mean)
4x4 TS{Float64,Dates.Date}: 2018-11-30 to 2019-02-28

Index       A        B        C        D
2018-11-30  -0.7196  -0.3733  -0.5991  0.5288
2018-12-31  -0.1079  -0.0919  0.1673   0.0273
2019-01-31  -0.3684  -0.1231  0.0059   0.207
2019-02-28  0.062    -0.1652  -0.1457  0.0021

julia> last_month = string(X.index[end])[1:7]
"2019-03"

julia> mean(X[last_month])
-0.04131138733537882

Summarizing

Temporal.apply — Function.

apply(x::TS{V}, dim::Int=1; fun=sum) where {V}

Apply function fun across dimension dim of a time series x.

If dim is 1, then apply fun to each row of x, returning a time series (TS) object of the results

If dim is 2, then apply fun to each column of x, returning an Array of the results

source

julia> using Temporal, Statistics

julia> X = TS(randn(100, 4))
100x4 TS{Float64,Dates.Date}: 2018-11-27 to 2019-03-06

Index       A        B        C        D
2018-11-27  0.7485   0.5368   0.5728   0.397
2018-11-28  0.112    0.7058   -1.5772  0.8621
2018-11-29  -1.4615  0.1374   1.286    0.1147
2018-11-30  0.5536   0.4988   -0.7091  -0.1476
2018-12-01  0.5644   0.4367   0.4958   0.7643
2018-12-02  -0.4324  2.3545   -1.0546  0.7174
2018-12-03  0.6566   -0.0116  0.8982   -0.5184
⋮
2019-02-27  -0.6301  -1.6958  0.0677   -0.9091
2019-02-28  0.8668   0.2142   -0.299   0.3549
2019-03-01  0.1709   0.072    -0.773   0.2653
2019-03-02  0.7243   0.1965   -0.3564  0.5918
2019-03-03  0.1506   -0.4336  0.6638   -0.204
2019-03-04  1.2053   0.0125   -0.3308  -0.9752
2019-03-05  0.7193   0.4944   0.328    -0.2258
2019-03-06  1.9405   0.347    1.1185   -0.0406

julia> apply(X, 1, fun=sum)
100x1 TS{Float64,Dates.Date}: 2018-11-27 to 2019-03-06

Index       Sum
2018-11-27  2.2551
2018-11-28  0.1026
2018-11-29  0.0767
2018-11-30  0.1957
2018-12-01  2.2612
2018-12-02  1.5848
2018-12-03  1.0248
⋮
2019-02-27  -3.1672
2019-02-28  1.137
2019-03-01  -0.2647
2019-03-02  1.1562
2019-03-03  0.1767
2019-03-04  -0.0882
2019-03-05  1.316
2019-03-06  3.3654

julia> apply(X, 2, fun=sum)
1×4 LinearAlgebra.Adjoint{Float64,Array{Float64,1}}:
 8.17999  4.72018  -10.1335  5.37657