Aggregation

Aggregation

Sampling

Weekdays

Temporal.mondays — Function.

mondays(x::TS)

Return a time series containing all observations occuring on Mondays of the given TS object.

source

Temporal.tuesdays — Function.

tuesdays(x::TS)

Return a time series containing all observations occuring on Tuesdays of the given TS object.

source

Temporal.wednesdays — Function.

wednesdays(x::TS)

Return a time series containing all observations occuring on Wednesdays of the given TS object.

source

Temporal.thursdays — Function.

thursdays(x::TS)

Return a time series containing all observations occuring on Thursdays of the given TS object.

source

Temporal.fridays — Function.

fridays(x::TS)

Return a time series containing all observations occuring on Fridays of the given TS object.

source

Temporal.saturdays — Function.

saturdays(x::TS)

Return a time series containing all observations occuring on Saturdays of the given TS object.

source

Temporal.sundays — Function.

sundays(x::TS)

Return a time series containing all observations occuring on Sundays of the given TS object.

source

Interval Boundaries

Weekly

Temporal.bow — Function.

bow(x::TS; cal::Bool=false)

Return a time series containing all observations occuring at the beginnings of the weeks of the input.

If cal is false, only observations occurring the last calendar day of the week are returned

If cal is true, all observation for which the previous index is a prior week are returned

source

Temporal.eow — Function.

eow(x::TS; cal::Bool=false)

Return a time series containing all observations occuring at the ends of the weeks of the input.

If cal is false, only observations occurring the last calendar day of the week are returned

If cal is true, all observation for which the next index is a new week are returned

source

Monthly

Temporal.bom — Function.

bom(x::TS; cal::Bool=false)

Return a time series containing all observations occuring at the beginnings of the months of the input.

If cal is false, only observations occurring the last calendar day of the month are returned

If cal is true, all observation for which the previous index is a prior month are returned

source

Temporal.eom — Function.

eom(x::TS; cal::Bool=false)

Return a time series containing all observations occuring at the ends of the months of the input.

If cal is false, only observations occurring the last calendar day of the month are returned

If cal is true, all observation for which the next index is a new month are returned

source

Quarterly

Temporal.boq — Function.

boq(x::TS; cal::Bool=false)

Return a time series containing all observations occuring at the beginnings of the quarters of the input.

If cal is false, only observations occurring the last calendar day of the quarter are returned

If cal is true, all observation for which the previous index is a prior quarter are returned

source

Temporal.eoq — Function.

eoq(x::TS; cal::Bool=false)

Return a time series containing all observations occuring at the ends of the quarters of the input.

If cal is false, only observations occurring the last calendar day of the quarter are returned

If cal is true, all observation for which the next index is a new quarter are returned

source

Yearly

Temporal.boy — Function.

boy(x::TS; cal::Bool=false)

Return a time series containing all observations occuring at the beginnings of the years of the input.

If cal is false, only observations occurring the last calendar day of the year are returned

If cal is true, all observation for which the previous index is a prior year are returned

source

Temporal.eoy — Function.

eoy(x::TS; cal::Bool=false)

Return a time series containing all observations occuring at the ends of the years of the input.

If cal is false, only observations occurring the last calendar day of the year are returned

If cal is true, all observation for which the next index is a new year are returned

source

Collapsing

Temporal.collapse — Function.

collapse(x::TS{V,T}, at::Function; fun::Function=last, args...)::TS{V,T} where {V,T}

Compute the function fun over period boundaries defined by at (e.g. eom, boq, etc.) and returng a time series of the results.

Keyword arguments (args...) are passed to the function fun.

source

julia> using Temporal, Statistics, Dates

julia> X = TS(randn(100, 4))
100x4 TS{Float64,Dates.Date}: 2018-11-06 to 2019-02-13

Index       A        B        C        D
2018-11-06  -0.8732  -0.9409  1.5053   -0.4485
2018-11-07  -1.0234  -0.1523  0.7448   -0.366
2018-11-08  -0.366   2.7439   1.7151   -0.3044
2018-11-09  0.8293   -1.0685  -0.5821  0.5282
2018-11-10  1.593    2.1003   1.4541   1.3565
2018-11-11  -0.8416  0.8367   0.4285   -0.247
2018-11-12  -0.3438  1.6708   -0.6894  -1.0421
⋮
2019-02-06  1.0681   0.1002   2.6078   1.2395
2019-02-07  0.5543   0.798    -1.3983  0.0143
2019-02-08  -1.2368  0.0527   -0.3828  0.5896
2019-02-09  0.4505   -0.1022  2.244    -0.2907
2019-02-10  -0.3502  0.04     -0.5196  0.1852
2019-02-11  1.6509   -0.7401  0.9774   -0.0245
2019-02-12  -0.5549  0.7785   -0.0653  -1.0789
2019-02-13  1.4529   -0.0869  0.7995   0.109

julia> collapse(X, eom, fun=mean)
3x4 TS{Float64,Dates.Date}: 2018-11-30 to 2019-01-31

Index       A        B        C        D
2018-11-30  0.0629   0.2776   0.0139   0.0061
2018-12-31  -0.0904  -0.0567  0.07     -0.1446
2019-01-31  -0.2496  -0.2316  -0.1129  -0.0285

julia> last_month = string(X.index[end])[1:7]
"2019-02"

julia> mean(X[last_month])
0.25289869535093534

Summarizing

Temporal.apply — Function.

apply(x::TS{V}, dim::Int=1; fun=sum) where {V}

Apply function fun across dimension dim of a time series x.

If dim is 1, then apply fun to each row of x, returning a time series (TS) object of the results

If dim is 2, then apply fun to each column of x, returning an Array of the results

source

julia> using Temporal, Statistics

julia> X = TS(randn(100, 4))
100x4 TS{Float64,Dates.Date}: 2018-11-06 to 2019-02-13

Index       A        B        C        D
2018-11-06  0.3606   -1.2188  0.5643   0.4344
2018-11-07  1.4603   0.3817   0.8671   0.6965
2018-11-08  -1.4269  0.0012   1.7696   -1.3493
2018-11-09  -0.5506  -1.5706  -0.8891  -0.2781
2018-11-10  -0.001   0.4211   1.1177   -2.0007
2018-11-11  0.6599   -0.6216  0.3846   2.1798
2018-11-12  0.5633   0.1645   0.4064   0.5574
⋮
2019-02-06  0.0623   -0.6496  -0.0224  0.4502
2019-02-07  -0.1447  -0.8053  0.9795   0.224
2019-02-08  0.8164   -0.2407  0.2861   1.1284
2019-02-09  -0.2623  -0.9975  -0.0286  0.1705
2019-02-10  -1.0583  -0.8915  -1.1284  -0.5084
2019-02-11  -3.5441  0.4301   0.5328   0.1593
2019-02-12  0.6024   1.6499   1.4245   0.862
2019-02-13  -3.03    2.7311   -0.6     1.0695

julia> apply(X, 1, fun=sum)
100x1 TS{Float64,Dates.Date}: 2018-11-06 to 2019-02-13

Index       Sum
2018-11-06  0.1405
2018-11-07  3.4056
2018-11-08  -1.0055
2018-11-09  -3.2884
2018-11-10  -0.4628
2018-11-11  2.6027
2018-11-12  1.6917
⋮
2019-02-06  -0.1595
2019-02-07  0.2536
2019-02-08  1.9902
2019-02-09  -1.1179
2019-02-10  -3.5866
2019-02-11  -2.422
2019-02-12  4.5388
2019-02-13  0.1706

julia> apply(X, 2, fun=sum)
1×4 LinearAlgebra.Adjoint{Float64,Array{Float64,1}}:
 3.23527  -9.23695  8.84733  -7.58343