Aggregation

Aggregation

Sampling

Weekdays

Temporal.mondays — Function.

mondays(x::TS)

Return a time series containing all observations occuring on Mondays of the given TS object.

source

Temporal.tuesdays — Function.

tuesdays(x::TS)

Return a time series containing all observations occuring on Tuesdays of the given TS object.

source

Temporal.wednesdays — Function.

wednesdays(x::TS)

Return a time series containing all observations occuring on Wednesdays of the given TS object.

source

Temporal.thursdays — Function.

thursdays(x::TS)

Return a time series containing all observations occuring on Thursdays of the given TS object.

source

Temporal.fridays — Function.

fridays(x::TS)

Return a time series containing all observations occuring on Fridays of the given TS object.

source

Temporal.saturdays — Function.

saturdays(x::TS)

Return a time series containing all observations occuring on Saturdays of the given TS object.

source

Temporal.sundays — Function.

sundays(x::TS)

Return a time series containing all observations occuring on Sundays of the given TS object.

source

Interval Boundaries

Weekly

Temporal.bow — Function.

bow(x::TS; cal::Bool=false)

Return a time series containing all observations occuring at the beginnings of the weeks of the input.

If cal is false, only observations occurring the last calendar day of the week are returned

If cal is true, all observation for which the previous index is a prior week are returned

source

Temporal.eow — Function.

eow(x::TS; cal::Bool=false)

Return a time series containing all observations occuring at the ends of the weeks of the input.

If cal is false, only observations occurring the last calendar day of the week are returned

If cal is true, all observation for which the next index is a new week are returned

source

Monthly

Temporal.bom — Function.

bom(x::TS; cal::Bool=false)

Return a time series containing all observations occuring at the beginnings of the months of the input.

If cal is false, only observations occurring the last calendar day of the month are returned

If cal is true, all observation for which the previous index is a prior month are returned

source

Temporal.eom — Function.

eom(x::TS; cal::Bool=false)

Return a time series containing all observations occuring at the ends of the months of the input.

If cal is false, only observations occurring the last calendar day of the month are returned

If cal is true, all observation for which the next index is a new month are returned

source

Quarterly

Temporal.boq — Function.

boq(x::TS; cal::Bool=false)

Return a time series containing all observations occuring at the beginnings of the quarters of the input.

If cal is false, only observations occurring the last calendar day of the quarter are returned

If cal is true, all observation for which the previous index is a prior quarter are returned

source

Temporal.eoq — Function.

eoq(x::TS; cal::Bool=false)

Return a time series containing all observations occuring at the ends of the quarters of the input.

If cal is false, only observations occurring the last calendar day of the quarter are returned

If cal is true, all observation for which the next index is a new quarter are returned

source

Yearly

Temporal.boy — Function.

boy(x::TS; cal::Bool=false)

Return a time series containing all observations occuring at the beginnings of the years of the input.

If cal is false, only observations occurring the last calendar day of the year are returned

If cal is true, all observation for which the previous index is a prior year are returned

source

Temporal.eoy — Function.

eoy(x::TS; cal::Bool=false)

Return a time series containing all observations occuring at the ends of the years of the input.

If cal is false, only observations occurring the last calendar day of the year are returned

If cal is true, all observation for which the next index is a new year are returned

source

Collapsing

Temporal.collapse — Function.

collapse(x::TS{V,T}, at::Function; fun::Function=last, args...)::TS{V,T} where {V,T}

Compute the function fun over period boundaries defined by at (e.g. eom, boq, etc.) and returng a time series of the results.

Keyword arguments (args...) are passed to the function fun.

source

julia> using Temporal, Statistics, Dates

julia> X = TS(randn(100, 4))
100x4 TS{Float64,Dates.Date}: 2018-11-04 to 2019-02-11

Index       A        B        C        D
2018-11-04  -0.2809  1.0091   1.0445   -1.1014
2018-11-05  -0.5161  -1.5921  -2.2596  1.071
2018-11-06  -0.1094  -0.052   1.0891   -0.8254
2018-11-07  -0.6475  -0.4377  -0.3597  -0.3814
2018-11-08  0.4095   -3.3373  -0.2041  -0.3318
2018-11-09  0.9984   -0.1975  -0.1065  -0.8242
2018-11-10  0.3254   0.6316   0.133    -1.364
⋮
2019-02-04  -1.8446  -2.6962  -0.1916  0.4046
2019-02-05  -0.0677  1.6227   0.7111   -0.2599
2019-02-06  0.3627   0.3407   0.8146   -0.1794
2019-02-07  0.7171   -0.9091  -0.3194  -0.147
2019-02-08  -0.4342  1.2037   0.4885   -1.455
2019-02-09  -1.8508  -0.8352  0.5212   -0.7938
2019-02-10  0.573    1.1384   -0.846   0.3792
2019-02-11  -0.2777  -0.5365  -0.2035  1.4226

julia> collapse(X, eom, fun=mean)
3x4 TS{Float64,Dates.Date}: 2018-11-30 to 2019-01-31

Index       A       B        C        D
2018-11-30  0.1731  -0.0413  0.1351   -0.2155
2018-12-31  0.2087  0.2676   0.134    0.1292
2019-01-31  0.3272  0.2893   -0.0727  0.041

julia> last_month = string(X.index[end])[1:7]
"2019-02"

julia> mean(X[last_month])
-0.09208042646181668

Summarizing

Temporal.apply — Function.

apply(x::TS{V}, dim::Int=1; fun=sum) where {V}

Apply function fun across dimension dim of a time series x.

If dim is 1, then apply fun to each row of x, returning a time series (TS) object of the results

If dim is 2, then apply fun to each column of x, returning an Array of the results

source

julia> using Temporal, Statistics

julia> X = TS(randn(100, 4))
100x4 TS{Float64,Dates.Date}: 2018-11-04 to 2019-02-11

Index       A        B        C        D
2018-11-04  0.3853   2.0427   0.1081   0.4202
2018-11-05  -0.0124  0.4132   0.5283   -0.4464
2018-11-06  0.1207   -0.4224  -0.1653  -0.3379
2018-11-07  -0.5742  1.1411   -0.595   0.6943
2018-11-08  -1.0261  -2.1879  -0.1388  0.3778
2018-11-09  0.3682   0.3883   -0.9423  -0.0778
2018-11-10  0.5833   1.0283   0.4414   -0.2077
⋮
2019-02-04  2.458    -0.683   -0.1466  0.01
2019-02-05  0.89     0.6919   -0.8576  -2.1248
2019-02-06  1.2886   0.3632   1.1399   -1.2606
2019-02-07  0.0137   -0.1396  -0.3321  0.7176
2019-02-08  -2.2275  1.5302   -1.3363  1.2127
2019-02-09  1.5683   -0.0485  -1.1301  -2.0799
2019-02-10  0.5542   1.3812   -1.1233  -1.0336
2019-02-11  -0.2677  -0.9953  1.0719   -0.7298

julia> apply(X, 1, fun=sum)
100x1 TS{Float64,Dates.Date}: 2018-11-04 to 2019-02-11

Index       Sum
2018-11-04  2.9563
2018-11-05  0.4828
2018-11-06  -0.805
2018-11-07  0.6662
2018-11-08  -2.9749
2018-11-09  -0.2635
2018-11-10  1.8453
⋮
2019-02-04  1.6385
2019-02-05  -1.4005
2019-02-06  1.5312
2019-02-07  0.2596
2019-02-08  -0.8209
2019-02-09  -1.6902
2019-02-10  -0.2214
2019-02-11  -0.9209

julia> apply(X, 2, fun=sum)
1×4 LinearAlgebra.Adjoint{Float64,Array{Float64,1}}:
 8.66294  -0.997722  -4.48264  -10.7809