Aggregation

Aggregation

Sampling

Weekdays

Temporal.mondays — Function.

mondays(x::TS)

Return a time series containing all observations occuring on Mondays of the given TS object.

source

Temporal.tuesdays — Function.

tuesdays(x::TS)

Return a time series containing all observations occuring on Tuesdays of the given TS object.

source

Temporal.wednesdays — Function.

wednesdays(x::TS)

Return a time series containing all observations occuring on Wednesdays of the given TS object.

source

Temporal.thursdays — Function.

thursdays(x::TS)

Return a time series containing all observations occuring on Thursdays of the given TS object.

source

Temporal.fridays — Function.

fridays(x::TS)

Return a time series containing all observations occuring on Fridays of the given TS object.

source

Temporal.saturdays — Function.

saturdays(x::TS)

Return a time series containing all observations occuring on Saturdays of the given TS object.

source

Temporal.sundays — Function.

sundays(x::TS)

Return a time series containing all observations occuring on Sundays of the given TS object.

source

Interval Boundaries

Weekly

Temporal.bow — Function.

bow(x::TS; cal::Bool=false)

Return a time series containing all observations occuring at the beginnings of the weeks of the input.

If cal is false, only observations occurring the last calendar day of the week are returned

If cal is true, all observation for which the previous index is a prior week are returned

source

Temporal.eow — Function.

eow(x::TS; cal::Bool=false)

Return a time series containing all observations occuring at the ends of the weeks of the input.

If cal is false, only observations occurring the last calendar day of the week are returned

If cal is true, all observation for which the next index is a new week are returned

source

Monthly

Temporal.bom — Function.

bom(x::TS; cal::Bool=false)

Return a time series containing all observations occuring at the beginnings of the months of the input.

If cal is false, only observations occurring the last calendar day of the month are returned

If cal is true, all observation for which the previous index is a prior month are returned

source

Temporal.eom — Function.

eom(x::TS; cal::Bool=false)

Return a time series containing all observations occuring at the ends of the months of the input.

If cal is false, only observations occurring the last calendar day of the month are returned

If cal is true, all observation for which the next index is a new month are returned

source

Quarterly

Temporal.boq — Function.

boq(x::TS; cal::Bool=false)

Return a time series containing all observations occuring at the beginnings of the quarters of the input.

If cal is false, only observations occurring the last calendar day of the quarter are returned

If cal is true, all observation for which the previous index is a prior quarter are returned

source

Temporal.eoq — Function.

eoq(x::TS; cal::Bool=false)

Return a time series containing all observations occuring at the ends of the quarters of the input.

If cal is false, only observations occurring the last calendar day of the quarter are returned

If cal is true, all observation for which the next index is a new quarter are returned

source

Yearly

Temporal.boy — Function.

boy(x::TS; cal::Bool=false)

Return a time series containing all observations occuring at the beginnings of the years of the input.

If cal is false, only observations occurring the last calendar day of the year are returned

If cal is true, all observation for which the previous index is a prior year are returned

source

Temporal.eoy — Function.

eoy(x::TS; cal::Bool=false)

Return a time series containing all observations occuring at the ends of the years of the input.

If cal is false, only observations occurring the last calendar day of the year are returned

If cal is true, all observation for which the next index is a new year are returned

source

Collapsing

Temporal.collapse — Function.

collapse(x::TS{V,T}, at::Function; fun::Function=last, args...)::TS{V,T} where {V,T}

Compute the function fun over period boundaries defined by at (e.g. eom, boq, etc.) and returng a time series of the results.

Keyword arguments (args...) are passed to the function fun.

source

julia> using Temporal, Statistics, Dates

julia> X = TS(randn(100, 4))
100x4 TS{Float64,Dates.Date}: 2018-11-06 to 2019-02-13

Index       A        B        C        D
2018-11-06  0.5873   -0.4861  1.2665   0.0434
2018-11-07  -1.1884  0.095    -0.8617  -0.9545
2018-11-08  -0.7781  0.2549   -0.2923  1.5284
2018-11-09  0.5413   1.3809   -1.8088  -1.6819
2018-11-10  -1.301   -0.9467  0.4292   -0.9876
2018-11-11  -0.6832  -0.3349  -0.381   0.5451
2018-11-12  -0.65    0.8212   0.0703   -0.4174
⋮
2019-02-06  1.9345   0.3372   0.1444   0.1085
2019-02-07  1.8315   -0.7694  -0.0164  0.6803
2019-02-08  -0.3795  0.9767   -2.4897  0.5291
2019-02-09  0.0606   -0.3161  1.3218   0.9788
2019-02-10  1.6057   0.1812   0.5018   0.3228
2019-02-11  -0.6271  -0.3816  -0.3073  0.7779
2019-02-12  -0.115   -0.0103  0.1481   0.5509
2019-02-13  0.4001   -0.8338  0.6098   -0.5694

julia> collapse(X, eom, fun=mean)
3x4 TS{Float64,Dates.Date}: 2018-11-30 to 2019-01-31

Index       A        B        C        D
2018-11-30  -0.4739  -0.1402  -0.2192  -0.2119
2018-12-31  -0.2569  -0.188   -0.159   0.132
2019-01-31  0.2017   -0.0275  -0.0031  0.1337

julia> last_month = string(X.index[end])[1:7]
"2019-02"

julia> mean(X[last_month])
-0.015551728808229474

Summarizing

Temporal.apply — Function.

apply(x::TS{V}, dim::Int=1; fun=sum) where {V}

Apply function fun across dimension dim of a time series x.

If dim is 1, then apply fun to each row of x, returning a time series (TS) object of the results

If dim is 2, then apply fun to each column of x, returning an Array of the results

source

julia> using Temporal, Statistics

julia> X = TS(randn(100, 4))
100x4 TS{Float64,Dates.Date}: 2018-11-06 to 2019-02-13

Index       A        B        C        D
2018-11-06  1.709    -1.7253  0.5297   -0.8166
2018-11-07  0.5825   1.1673   2.1147   -1.5256
2018-11-08  1.0235   -0.6125  -2.3172  -0.0543
2018-11-09  1.1005   0.3353   2.0274   -0.4348
2018-11-10  1.3212   0.5894   0.7075   0.9505
2018-11-11  -0.827   1.3801   -1.6406  -1.2544
2018-11-12  0.0962   -0.5569  -0.3263  1.8688
⋮
2019-02-06  0.6194   -1.9077  1.1574   -0.9523
2019-02-07  0.6877   0.5805   -1.016   0.0956
2019-02-08  1.99     -0.4797  1.2611   0.32
2019-02-09  0.2519   0.1685   -0.5698  -1.5205
2019-02-10  0.5663   -2.4228  0.5852   1.1792
2019-02-11  -0.1184  -0.4241  -1.788   0.3077
2019-02-12  0.207    -0.4896  -0.8894  0.1776
2019-02-13  0.7208   1.4062   -0.467   -0.0184

julia> apply(X, 1, fun=sum)
100x1 TS{Float64,Dates.Date}: 2018-11-06 to 2019-02-13

Index       Sum
2018-11-06  -0.3033
2018-11-07  2.3389
2018-11-08  -1.9604
2018-11-09  3.0285
2018-11-10  3.5685
2018-11-11  -2.342
2018-11-12  1.0818
⋮
2019-02-06  -1.0831
2019-02-07  0.3478
2019-02-08  3.0913
2019-02-09  -1.6698
2019-02-10  -0.0921
2019-02-11  -2.0227
2019-02-12  -0.9944
2019-02-13  1.6416

julia> apply(X, 2, fun=sum)
1×4 LinearAlgebra.Adjoint{Float64,Array{Float64,1}}:
 10.4014  -2.84054  -6.4556  6.34041