Aggregation

Aggregation

Sampling

Weekdays

Temporal.mondays — Function.

mondays(x::TS)

Return a time series containing all observations occuring on Mondays of the given TS object.

source

Temporal.tuesdays — Function.

tuesdays(x::TS)

Return a time series containing all observations occuring on Tuesdays of the given TS object.

source

Temporal.wednesdays — Function.

wednesdays(x::TS)

Return a time series containing all observations occuring on Wednesdays of the given TS object.

source

Temporal.thursdays — Function.

thursdays(x::TS)

Return a time series containing all observations occuring on Thursdays of the given TS object.

source

Temporal.fridays — Function.

fridays(x::TS)

Return a time series containing all observations occuring on Fridays of the given TS object.

source

Temporal.saturdays — Function.

saturdays(x::TS)

Return a time series containing all observations occuring on Saturdays of the given TS object.

source

Temporal.sundays — Function.

sundays(x::TS)

Return a time series containing all observations occuring on Sundays of the given TS object.

source

Interval Boundaries

Weekly

Temporal.bow — Function.

bow(x::TS; cal::Bool=false)

Return a time series containing all observations occuring at the beginnings of the weeks of the input.

If cal is false, only observations occurring the last calendar day of the week are returned

If cal is true, all observation for which the previous index is a prior week are returned

source

Temporal.eow — Function.

eow(x::TS; cal::Bool=false)

Return a time series containing all observations occuring at the ends of the weeks of the input.

If cal is false, only observations occurring the last calendar day of the week are returned

If cal is true, all observation for which the next index is a new week are returned

source

Monthly

Temporal.bom — Function.

bom(x::TS; cal::Bool=false)

Return a time series containing all observations occuring at the beginnings of the months of the input.

If cal is false, only observations occurring the last calendar day of the month are returned

If cal is true, all observation for which the previous index is a prior month are returned

source

Temporal.eom — Function.

eom(x::TS; cal::Bool=false)

Return a time series containing all observations occuring at the ends of the months of the input.

If cal is false, only observations occurring the last calendar day of the month are returned

If cal is true, all observation for which the next index is a new month are returned

source

Quarterly

Temporal.boq — Function.

boq(x::TS; cal::Bool=false)

Return a time series containing all observations occuring at the beginnings of the quarters of the input.

If cal is false, only observations occurring the last calendar day of the quarter are returned

If cal is true, all observation for which the previous index is a prior quarter are returned

source

Temporal.eoq — Function.

eoq(x::TS; cal::Bool=false)

Return a time series containing all observations occuring at the ends of the quarters of the input.

If cal is false, only observations occurring the last calendar day of the quarter are returned

If cal is true, all observation for which the next index is a new quarter are returned

source

Yearly

Temporal.boy — Function.

boy(x::TS; cal::Bool=false)

Return a time series containing all observations occuring at the beginnings of the years of the input.

If cal is false, only observations occurring the last calendar day of the year are returned

If cal is true, all observation for which the previous index is a prior year are returned

source

Temporal.eoy — Function.

eoy(x::TS; cal::Bool=false)

Return a time series containing all observations occuring at the ends of the years of the input.

If cal is false, only observations occurring the last calendar day of the year are returned

If cal is true, all observation for which the next index is a new year are returned

source

Collapsing

Temporal.collapse — Function.

collapse(x::TS{V,T}, at::Function; fun::Function=last, args...)::TS{V,T} where {V,T}

Compute the function fun over period boundaries defined by at (e.g. eom, boq, etc.) and returng a time series of the results.

Keyword arguments (args...) are passed to the function fun.

source

julia> using Temporal, Statistics, Dates

julia> X = TS(randn(100, 4))
100x4 TS{Float64,Dates.Date}: 2019-07-24 to 2019-10-31

Index       A        B        C        D
2019-07-24  -2.3018  -0.2828  0.4875   -0.246
2019-07-25  0.232    1.2037   -1.8754  1.497
2019-07-26  -1.1751  2.1054   0.3261   -1.5753
2019-07-27  1.2386   -1.7215  0.6213   0.039
2019-07-28  1.2486   -0.8634  1.0374   -0.3833
2019-07-29  -1.0043  -0.3877  0.7699   -0.0497
2019-07-30  -0.614   0.5707   -1.2834  -0.4234
⋮
2019-10-24  -1.4874  0.8961   -0.7731  0.0394
2019-10-25  -0.6248  -0.3519  0.5384   -1.4135
2019-10-26  0.426    0.012    -0.0375  0.1976
2019-10-27  0.0807   -0.3356  -0.1162  -1.8235
2019-10-28  -0.0953  0.8098   0.1789   0.2415
2019-10-29  -0.2331  -1.0167  0.3641   0.5183
2019-10-30  -1.8526  -0.0866  0.1864   -1.0404
2019-10-31  -0.102   1.0886   0.3754   1.0626

julia> collapse(X, eom, fun=mean)
3x4 TS{Float64,Dates.Date}: 2019-07-31 to 2019-09-30

Index       A        B        C        D
2019-07-31  -0.2615  -0.0366  0.1248   -0.0837
2019-08-31  -0.3009  0.2209   -0.0875  0.1175
2019-09-30  0.0996   0.2017   -0.2594  0.1218

julia> last_month = string(X.index[end])[1:7]
"2019-10"

julia> mean(X[last_month])
0.09482673362821867

Summarizing

Temporal.apply — Function.

apply(x::TS{V}, dim::Int=1; fun=sum) where {V}

Apply function fun across dimension dim of a time series x.

If dim is 1, then apply fun to each row of x, returning a time series (TS) object of the results

If dim is 2, then apply fun to each column of x, returning an Array of the results

source

julia> using Temporal, Statistics

julia> X = TS(randn(100, 4))
100x4 TS{Float64,Dates.Date}: 2019-07-24 to 2019-10-31

Index       A        B        C        D
2019-07-24  1.3778   -0.1293  0.7941   0.1109
2019-07-25  -0.1386  0.123    0.2897   -0.564
2019-07-26  1.4485   -0.4569  -0.0129  -0.4919
2019-07-27  0.342    -0.2555  -0.4758  -1.0367
2019-07-28  0.7639   -0.0346  0.765    0.8831
2019-07-29  0.6874   0.7513   -1.5539  -0.1651
2019-07-30  0.8322   -0.1887  1.3765   0.8741
⋮
2019-10-24  1.0363   1.1077   1.3918   -1.1015
2019-10-25  0.2999   0.1202   -2.5723  -1.1503
2019-10-26  0.4951   -1.0033  1.116    -2.2547
2019-10-27  0.0847   1.0247   0.3704   -0.1555
2019-10-28  -1.8072  -0.9992  0.7867   -0.9603
2019-10-29  0.5776   -0.8929  -0.2154  0.6576
2019-10-30  0.0316   -1.5257  0.1965   0.451
2019-10-31  0.9361   -0.7277  -0.143   0.4833

julia> apply(X, 1, fun=sum)
100x1 TS{Float64,Dates.Date}: 2019-07-24 to 2019-10-31

Index       Sum
2019-07-24  2.1535
2019-07-25  -0.2899
2019-07-26  0.4867
2019-07-27  -1.426
2019-07-28  2.3773
2019-07-29  -0.2804
2019-07-30  2.894
⋮
2019-10-24  2.4343
2019-10-25  -3.3024
2019-10-26  -1.6468
2019-10-27  1.3243
2019-10-28  -2.98
2019-10-29  0.127
2019-10-30  -0.8466
2019-10-31  0.5487

julia> apply(X, 2, fun=sum)
1×4 LinearAlgebra.Adjoint{Float64,Array{Float64,1}}:
 10.7832  -11.5962  -19.33  2.80257