Aggregation

Aggregation

Sampling

Weekdays

Temporal.mondays — Function.

mondays(x::TS)

Return a time series containing all observations occuring on Mondays of the given TS object.

source

Temporal.tuesdays — Function.

tuesdays(x::TS)

Return a time series containing all observations occuring on Tuesdays of the given TS object.

source

Temporal.wednesdays — Function.

wednesdays(x::TS)

Return a time series containing all observations occuring on Wednesdays of the given TS object.

source

Temporal.thursdays — Function.

thursdays(x::TS)

Return a time series containing all observations occuring on Thursdays of the given TS object.

source

Temporal.fridays — Function.

fridays(x::TS)

Return a time series containing all observations occuring on Fridays of the given TS object.

source

Temporal.saturdays — Function.

saturdays(x::TS)

Return a time series containing all observations occuring on Saturdays of the given TS object.

source

Temporal.sundays — Function.

sundays(x::TS)

Return a time series containing all observations occuring on Sundays of the given TS object.

source

Interval Boundaries

Weekly

Temporal.bow — Function.

bow(x::TS; cal::Bool=false)

Return a time series containing all observations occuring at the beginnings of the weeks of the input.

If cal is false, only observations occurring the last calendar day of the week are returned

If cal is true, all observation for which the previous index is a prior week are returned

source

Temporal.eow — Function.

eow(x::TS; cal::Bool=false)

Return a time series containing all observations occuring at the ends of the weeks of the input.

If cal is false, only observations occurring the last calendar day of the week are returned

If cal is true, all observation for which the next index is a new week are returned

source

Monthly

Temporal.bom — Function.

bom(x::TS; cal::Bool=false)

Return a time series containing all observations occuring at the beginnings of the months of the input.

If cal is false, only observations occurring the last calendar day of the month are returned

If cal is true, all observation for which the previous index is a prior month are returned

source

Temporal.eom — Function.

eom(x::TS; cal::Bool=false)

Return a time series containing all observations occuring at the ends of the months of the input.

If cal is false, only observations occurring the last calendar day of the month are returned

If cal is true, all observation for which the next index is a new month are returned

source

Quarterly

Temporal.boq — Function.

boq(x::TS; cal::Bool=false)

Return a time series containing all observations occuring at the beginnings of the quarters of the input.

If cal is false, only observations occurring the last calendar day of the quarter are returned

If cal is true, all observation for which the previous index is a prior quarter are returned

source

Temporal.eoq — Function.

eoq(x::TS; cal::Bool=false)

Return a time series containing all observations occuring at the ends of the quarters of the input.

If cal is false, only observations occurring the last calendar day of the quarter are returned

If cal is true, all observation for which the next index is a new quarter are returned

source

Yearly

Temporal.boy — Function.

boy(x::TS; cal::Bool=false)

Return a time series containing all observations occuring at the beginnings of the years of the input.

If cal is false, only observations occurring the last calendar day of the year are returned

If cal is true, all observation for which the previous index is a prior year are returned

source

Temporal.eoy — Function.

eoy(x::TS; cal::Bool=false)

Return a time series containing all observations occuring at the ends of the years of the input.

If cal is false, only observations occurring the last calendar day of the year are returned

If cal is true, all observation for which the next index is a new year are returned

source

Collapsing

Temporal.collapse — Function.

collapse(x::TS{V,T}, at::Function; fun::Function=last, args...)::TS{V,T} where {V,T}

Compute the function fun over period boundaries defined by at (e.g. eom, boq, etc.) and returng a time series of the results.

Keyword arguments (args...) are passed to the function fun.

source

julia> using Temporal, Statistics, Dates

julia> X = TS(randn(100, 4))
100x4 TS{Float64,Dates.Date}: 2018-11-27 to 2019-03-06

Index       A        B        C        D
2018-11-27  0.6892   -0.8157  0.9092   -0.057
2018-11-28  -1.1986  -0.0658  -0.7102  -1.7321
2018-11-29  -0.0302  -1.4353  0.0259   -0.8192
2018-11-30  -1.0916  0.0088   1.2349   -1.2884
2018-12-01  -1.7438  -0.5997  -0.4076  0.2871
2018-12-02  -0.6087  0.3578   0.5884   -1.0364
2018-12-03  1.8525   0.2691   0.0089   -0.8449
⋮
2019-02-27  0.629    0.8967   -0.1753  -2.0792
2019-02-28  -1.1811  0.5254   1.4421   -0.069
2019-03-01  -0.9822  -0.4131  -0.57    2.0463
2019-03-02  -0.1924  1.4449   -0.7887  0.4908
2019-03-03  -2.1353  0.3814   -0.1333  2.3521
2019-03-04  1.2238   0.5244   1.4612   0.6597
2019-03-05  -0.5179  -0.5183  -1.4998  -0.6771
2019-03-06  0.4733   1.4169   -0.1252  -2.3057

julia> collapse(X, eom, fun=mean)
4x4 TS{Float64,Dates.Date}: 2018-11-30 to 2019-02-28

Index       A        B        C        D
2018-11-30  -0.4078  -0.577   0.3649   -0.9742
2018-12-31  -0.2393  -0.1585  -0.2482  0.0806
2019-01-31  -0.2907  -0.0593  -0.0657  -0.2845
2019-02-28  0.203    -0.1223  -0.1992  0.1521

julia> last_month = string(X.index[end])[1:7]
"2019-03"

julia> mean(X[last_month])
0.06732449874741143

Summarizing

Temporal.apply — Function.

apply(x::TS{V}, dim::Int=1; fun=sum) where {V}

Apply function fun across dimension dim of a time series x.

If dim is 1, then apply fun to each row of x, returning a time series (TS) object of the results

If dim is 2, then apply fun to each column of x, returning an Array of the results

source

julia> using Temporal, Statistics

julia> X = TS(randn(100, 4))
100x4 TS{Float64,Dates.Date}: 2018-11-27 to 2019-03-06

Index       A        B        C        D
2018-11-27  -0.738   1.0791   0.7145   -0.028
2018-11-28  0.5416   1.3368   0.7567   -0.6349
2018-11-29  -0.9448  -0.4565  1.5069   0.5877
2018-11-30  -0.2143  0.5532   -0.8375  -1.6427
2018-12-01  -0.5844  0.2455   0.8058   -1.1391
2018-12-02  -1.9826  -0.7526  0.7004   0.3549
2018-12-03  -1.6709  -0.0051  -0.608   1.519
⋮
2019-02-27  -0.182   0.6969   1.6685   1.5751
2019-02-28  0.9778   -1.4779  0.221    0.8345
2019-03-01  -0.3596  0.0747   1.1977   -0.3053
2019-03-02  0.0804   -0.9451  1.54     -0.0832
2019-03-03  -0.5704  -1.5468  0.7116   0.9108
2019-03-04  0.9224   1.5277   -0.1026  0.8264
2019-03-05  0.0059   0.7334   -1.0678  -0.3652
2019-03-06  1.7465   -0.1186  0.8632   -2.2526

julia> apply(X, 1, fun=sum)
100x1 TS{Float64,Dates.Date}: 2018-11-27 to 2019-03-06

Index       Sum
2018-11-27  1.0276
2018-11-28  2.0003
2018-11-29  0.6933
2018-11-30  -2.1413
2018-12-01  -0.6722
2018-12-02  -1.6798
2018-12-03  -0.765
⋮
2019-02-27  3.7586
2019-02-28  0.5555
2019-03-01  0.6075
2019-03-02  0.5921
2019-03-03  -0.4948
2019-03-04  3.1739
2019-03-05  -0.6938
2019-03-06  0.2385

julia> apply(X, 2, fun=sum)
1×4 LinearAlgebra.Adjoint{Float64,Array{Float64,1}}:
 2.14155  5.78286  6.84679  1.97491