Aggregation

Aggregation

Sampling

Weekdays

Temporal.mondays — Function.

mondays(x::TS)

Return a time series containing all observations occuring on Mondays of the given TS object.

source

Temporal.tuesdays — Function.

tuesdays(x::TS)

Return a time series containing all observations occuring on Tuesdays of the given TS object.

source

Temporal.wednesdays — Function.

wednesdays(x::TS)

Return a time series containing all observations occuring on Wednesdays of the given TS object.

source

Temporal.thursdays — Function.

thursdays(x::TS)

Return a time series containing all observations occuring on Thursdays of the given TS object.

source

Temporal.fridays — Function.

fridays(x::TS)

Return a time series containing all observations occuring on Fridays of the given TS object.

source

Temporal.saturdays — Function.

saturdays(x::TS)

Return a time series containing all observations occuring on Saturdays of the given TS object.

source

Temporal.sundays — Function.

sundays(x::TS)

Return a time series containing all observations occuring on Sundays of the given TS object.

source

Interval Boundaries

Weekly

Temporal.bow — Function.

bow(x::TS; cal::Bool=false)

Return a time series containing all observations occuring at the beginnings of the weeks of the input.

If cal is false, only observations occurring the last calendar day of the week are returned

If cal is true, all observation for which the previous index is a prior week are returned

source

Temporal.eow — Function.

eow(x::TS; cal::Bool=false)

Return a time series containing all observations occuring at the ends of the weeks of the input.

If cal is false, only observations occurring the last calendar day of the week are returned

If cal is true, all observation for which the next index is a new week are returned

source

Monthly

Temporal.bom — Function.

bom(x::TS; cal::Bool=false)

Return a time series containing all observations occuring at the beginnings of the months of the input.

If cal is false, only observations occurring the last calendar day of the month are returned

If cal is true, all observation for which the previous index is a prior month are returned

source

Temporal.eom — Function.

eom(x::TS; cal::Bool=false)

Return a time series containing all observations occuring at the ends of the months of the input.

If cal is false, only observations occurring the last calendar day of the month are returned

If cal is true, all observation for which the next index is a new month are returned

source

Quarterly

Temporal.boq — Function.

boq(x::TS; cal::Bool=false)

Return a time series containing all observations occuring at the beginnings of the quarters of the input.

If cal is false, only observations occurring the last calendar day of the quarter are returned

If cal is true, all observation for which the previous index is a prior quarter are returned

source

Temporal.eoq — Function.

eoq(x::TS; cal::Bool=false)

Return a time series containing all observations occuring at the ends of the quarters of the input.

If cal is false, only observations occurring the last calendar day of the quarter are returned

If cal is true, all observation for which the next index is a new quarter are returned

source

Yearly

Temporal.boy — Function.

boy(x::TS; cal::Bool=false)

Return a time series containing all observations occuring at the beginnings of the years of the input.

If cal is false, only observations occurring the last calendar day of the year are returned

If cal is true, all observation for which the previous index is a prior year are returned

source

Temporal.eoy — Function.

eoy(x::TS; cal::Bool=false)

Return a time series containing all observations occuring at the ends of the years of the input.

If cal is false, only observations occurring the last calendar day of the year are returned

If cal is true, all observation for which the next index is a new year are returned

source

Collapsing

Temporal.collapse — Function.

collapse(x::TS{V,T}, at::Function; fun::Function=last, args...)::TS{V,T} where {V,T}

Compute the function fun over period boundaries defined by at (e.g. eom, boq, etc.) and returng a time series of the results.

Keyword arguments (args...) are passed to the function fun.

source

julia> using Temporal, Statistics, Dates

julia> X = TS(randn(100, 4))
100x4 TS{Float64,Dates.Date}: 2018-11-06 to 2019-02-13

Index       A        B        C        D
2018-11-06  -1.0877  1.1192   -0.0336  -0.2622
2018-11-07  1.327    0.443    0.4515   -1.7116
2018-11-08  0.6583   -1.5832  -0.1281  -0.2399
2018-11-09  0.7655   -2.0433  0.3275   0.5625
2018-11-10  1.26     -0.5568  0.4617   0.4664
2018-11-11  -0.2415  -1.3631  -1.8452  1.1616
2018-11-12  -0.8764  -0.7752  -0.5459  -0.576
⋮
2019-02-06  -0.3801  -1.6498  2.3212   1.8464
2019-02-07  2.1928   0.6241   0.0605   0.0555
2019-02-08  -1.1104  0.8696   -0.9197  -0.4593
2019-02-09  0.8591   1.8355   -0.0312  0.0006
2019-02-10  0.6151   1.0747   1.0583   1.5799
2019-02-11  -0.5108  -0.0357  1.0441   1.1757
2019-02-12  -0.3041  -0.575   0.564    -0.4066
2019-02-13  -0.1956  0.953    -0.9176  -0.203

julia> collapse(X, eom, fun=mean)
3x4 TS{Float64,Dates.Date}: 2018-11-30 to 2019-01-31

Index       A        B        C        D
2018-11-30  0.265    -0.1111  -0.1079  0.083
2018-12-31  -0.0549  0.0158   0.2674   0.2254
2019-01-31  0.0166   0.2681   0.2172   0.1071

julia> last_month = string(X.index[end])[1:7]
"2019-02"

julia> mean(X[last_month])
0.26974655517157026

Summarizing

Temporal.apply — Function.

apply(x::TS{V}, dim::Int=1; fun=sum) where {V}

Apply function fun across dimension dim of a time series x.

If dim is 1, then apply fun to each row of x, returning a time series (TS) object of the results

If dim is 2, then apply fun to each column of x, returning an Array of the results

source

julia> using Temporal, Statistics

julia> X = TS(randn(100, 4))
100x4 TS{Float64,Dates.Date}: 2018-11-06 to 2019-02-13

Index       A        B        C        D
2018-11-06  0.0519   1.3262   0.8649   -1.4723
2018-11-07  -0.617   0.9331   -1.2081  1.598
2018-11-08  -1.5866  -1.112   -0.9125  -1.2729
2018-11-09  -0.4291  -0.7041  0.9838   -0.9549
2018-11-10  0.9955   -0.3628  -2.8232  0.3326
2018-11-11  -1.6353  -0.1014  -1.0499  -1.6576
2018-11-12  -0.2824  2.0873   0.5268   0.324
⋮
2019-02-06  -0.5244  -0.8661  -1.166   0.7051
2019-02-07  0.7247   -0.088   -0.4975  0.4862
2019-02-08  0.3768   -0.612   -1.5328  -1.6178
2019-02-09  0.2721   -0.4393  1.9      -0.7413
2019-02-10  1.2085   0.2255   -1.2863  0.0048
2019-02-11  -0.96    1.7775   -1.0651  -1.1135
2019-02-12  -0.4782  1.7592   -0.1376  0.167
2019-02-13  1.937    1.4642   -0.7314  -0.7054

julia> apply(X, 1, fun=sum)
100x1 TS{Float64,Dates.Date}: 2018-11-06 to 2019-02-13

Index       Sum
2018-11-06  0.7707
2018-11-07  0.706
2018-11-08  -4.884
2018-11-09  -1.1042
2018-11-10  -1.8579
2018-11-11  -4.4441
2018-11-12  2.6556
⋮
2019-02-06  -1.8514
2019-02-07  0.6254
2019-02-08  -3.3859
2019-02-09  0.9915
2019-02-10  0.1525
2019-02-11  -1.361
2019-02-12  1.3104
2019-02-13  1.9644

julia> apply(X, 2, fun=sum)
1×4 LinearAlgebra.Adjoint{Float64,Array{Float64,1}}:
 -3.98009  -9.2153  -8.25184  -13.4137