Aggregation

Aggregation

Sampling

Weekdays

Temporal.mondays — Function.

mondays(x::TS)

Return a time series containing all observations occuring on Mondays of the given TS object.

source

Temporal.tuesdays — Function.

tuesdays(x::TS)

Return a time series containing all observations occuring on Tuesdays of the given TS object.

source

Temporal.wednesdays — Function.

wednesdays(x::TS)

Return a time series containing all observations occuring on Wednesdays of the given TS object.

source

Temporal.thursdays — Function.

thursdays(x::TS)

Return a time series containing all observations occuring on Thursdays of the given TS object.

source

Temporal.fridays — Function.

fridays(x::TS)

Return a time series containing all observations occuring on Fridays of the given TS object.

source

Temporal.saturdays — Function.

saturdays(x::TS)

Return a time series containing all observations occuring on Saturdays of the given TS object.

source

Temporal.sundays — Function.

sundays(x::TS)

Return a time series containing all observations occuring on Sundays of the given TS object.

source

Interval Boundaries

Weekly

Temporal.bow — Function.

bow(x::TS; cal::Bool=false)

Return a time series containing all observations occuring at the beginnings of the weeks of the input.

If cal is false, only observations occurring the last calendar day of the week are returned

If cal is true, all observation for which the previous index is a prior week are returned

source

Temporal.eow — Function.

eow(x::TS; cal::Bool=false)

Return a time series containing all observations occuring at the ends of the weeks of the input.

If cal is false, only observations occurring the last calendar day of the week are returned

If cal is true, all observation for which the next index is a new week are returned

source

Monthly

Temporal.bom — Function.

bom(x::TS; cal::Bool=false)

Return a time series containing all observations occuring at the beginnings of the months of the input.

If cal is false, only observations occurring the last calendar day of the month are returned

If cal is true, all observation for which the previous index is a prior month are returned

source

Temporal.eom — Function.

eom(x::TS; cal::Bool=false)

Return a time series containing all observations occuring at the ends of the months of the input.

If cal is false, only observations occurring the last calendar day of the month are returned

If cal is true, all observation for which the next index is a new month are returned

source

Quarterly

Temporal.boq — Function.

boq(x::TS; cal::Bool=false)

Return a time series containing all observations occuring at the beginnings of the quarters of the input.

If cal is false, only observations occurring the last calendar day of the quarter are returned

If cal is true, all observation for which the previous index is a prior quarter are returned

source

Temporal.eoq — Function.

eoq(x::TS; cal::Bool=false)

Return a time series containing all observations occuring at the ends of the quarters of the input.

If cal is false, only observations occurring the last calendar day of the quarter are returned

If cal is true, all observation for which the next index is a new quarter are returned

source

Yearly

Temporal.boy — Function.

boy(x::TS; cal::Bool=false)

Return a time series containing all observations occuring at the beginnings of the years of the input.

If cal is false, only observations occurring the last calendar day of the year are returned

If cal is true, all observation for which the previous index is a prior year are returned

source

Temporal.eoy — Function.

eoy(x::TS; cal::Bool=false)

Return a time series containing all observations occuring at the ends of the years of the input.

If cal is false, only observations occurring the last calendar day of the year are returned

If cal is true, all observation for which the next index is a new year are returned

source

Collapsing

Temporal.collapse — Function.

collapse(x::TS{V,T}, at::Function; fun::Function=last, args...)::TS{V,T} where {V,T}

Compute the function fun over period boundaries defined by at (e.g. eom, boq, etc.) and returng a time series of the results.

Keyword arguments (args...) are passed to the function fun.

source

julia> using Temporal, Statistics, Dates

julia> X = TS(randn(100, 4))
100x4 TS{Float64,Dates.Date}: 2018-12-09 to 2019-03-18

Index       A        B        C        D
2018-12-09  1.0111   0.2838   1.2668   -1.6254
2018-12-10  0.7164   -0.5353  -0.4308  -0.8121
2018-12-11  0.3945   -0.8081  0.3122   0.9898
2018-12-12  -1.894   -1.529   -0.2186  1.2007
2018-12-13  0.2572   0.0063   -0.617   -0.8631
2018-12-14  0.716    0.5219   -0.129   -2.4696
2018-12-15  0.0014   -1.0881  -0.7419  -1.0195
⋮
2019-03-11  -0.4423  0.2564   1.351    0.9879
2019-03-12  -0.0541  -0.1051  0.3179   -0.9601
2019-03-13  0.4745   0.3695   0.0712   0.3337
2019-03-14  -1.0735  -0.5392  0.9662   -3.3228
2019-03-15  1.6305   0.106    0.0487   -0.1123
2019-03-16  -0.3129  -0.7351  0.0064   -1.0215
2019-03-17  -1.7801  -0.0339  0.2343   1.0083
2019-03-18  -0.9576  -0.5619  0.3665   0.7806

julia> collapse(X, eom, fun=mean)
3x4 TS{Float64,Dates.Date}: 2018-12-31 to 2019-02-28

Index       A        B        C        D
2018-12-31  -0.1797  0.0156   -0.0108  -0.1178
2019-01-31  -0.1708  0.0596   0.0364   0.0429
2019-02-28  0.1562   -0.0096  -0.1507  0.1877

julia> last_month = string(X.index[end])[1:7]
"2019-03"

julia> mean(X[last_month])
0.035238890185251635

Summarizing

Temporal.apply — Function.

apply(x::TS{V}, dim::Int=1; fun=sum) where {V}

Apply function fun across dimension dim of a time series x.

If dim is 1, then apply fun to each row of x, returning a time series (TS) object of the results

If dim is 2, then apply fun to each column of x, returning an Array of the results

source

julia> using Temporal, Statistics

julia> X = TS(randn(100, 4))
100x4 TS{Float64,Dates.Date}: 2018-12-09 to 2019-03-18

Index       A        B        C        D
2018-12-09  -0.6325  1.5702   -0.3169  -1.6358
2018-12-10  1.0781   -1.7128  0.3218   0.2583
2018-12-11  -0.5837  0.3962   1.4505   0.5907
2018-12-12  0.8223   -0.2441  -1.1727  -0.9085
2018-12-13  -0.8337  1.2818   0.01     2.8162
2018-12-14  -0.1496  -0.6178  -0.5698  0.7061
2018-12-15  0.8954   0.517    -0.237   0.8928
⋮
2019-03-11  1.4687   0.7916   0.5566   0.0328
2019-03-12  -2.3058  0.3208   0.573    -0.8525
2019-03-13  -0.3268  0.1037   0.6157   -1.4963
2019-03-14  -0.4627  0.9209   0.2298   -1.3816
2019-03-15  -0.7573  -0.7876  0.02     0.0064
2019-03-16  -0.4116  -0.9602  -0.7464  -2.1483
2019-03-17  -1.4044  0.0578   -1.2     2.9365
2019-03-18  1.1423   -0.0239  -0.7982  0.3688

julia> apply(X, 1, fun=sum)
100x1 TS{Float64,Dates.Date}: 2018-12-09 to 2019-03-18

Index       Sum
2018-12-09  -1.015
2018-12-10  -0.0547
2018-12-11  1.8537
2018-12-12  -1.503
2018-12-13  3.2743
2018-12-14  -0.6312
2018-12-15  2.0682
⋮
2019-03-11  2.8496
2019-03-12  -2.2646
2019-03-13  -1.1037
2019-03-14  -0.6936
2019-03-15  -1.5184
2019-03-16  -4.2665
2019-03-17  0.3899
2019-03-18  0.689

julia> apply(X, 2, fun=sum)
1×4 LinearAlgebra.Adjoint{Float64,Array{Float64,1}}:
 -12.2637  -4.17824  -16.2438  4.22533