Aggregation

Aggregation

Sampling

Weekdays

Temporal.mondays — Function.

mondays(x::TS)

Return a time series containing all observations occuring on Mondays of the given TS object.

source

Temporal.tuesdays — Function.

tuesdays(x::TS)

Return a time series containing all observations occuring on Tuesdays of the given TS object.

source

Temporal.wednesdays — Function.

wednesdays(x::TS)

Return a time series containing all observations occuring on Wednesdays of the given TS object.

source

Temporal.thursdays — Function.

thursdays(x::TS)

Return a time series containing all observations occuring on Thursdays of the given TS object.

source

Temporal.fridays — Function.

fridays(x::TS)

Return a time series containing all observations occuring on Fridays of the given TS object.

source

Temporal.saturdays — Function.

saturdays(x::TS)

Return a time series containing all observations occuring on Saturdays of the given TS object.

source

Temporal.sundays — Function.

sundays(x::TS)

Return a time series containing all observations occuring on Sundays of the given TS object.

source

Interval Boundaries

Weekly

Temporal.bow — Function.

bow(x::TS; cal::Bool=false)

Return a time series containing all observations occuring at the beginnings of the weeks of the input.

If cal is false, only observations occurring the last calendar day of the week are returned

If cal is true, all observation for which the previous index is a prior week are returned

source

Temporal.eow — Function.

eow(x::TS; cal::Bool=false)

Return a time series containing all observations occuring at the ends of the weeks of the input.

If cal is false, only observations occurring the last calendar day of the week are returned

If cal is true, all observation for which the next index is a new week are returned

source

Monthly

Temporal.bom — Function.

bom(x::TS; cal::Bool=false)

Return a time series containing all observations occuring at the beginnings of the months of the input.

If cal is false, only observations occurring the last calendar day of the month are returned

If cal is true, all observation for which the previous index is a prior month are returned

source

Temporal.eom — Function.

eom(x::TS; cal::Bool=false)

Return a time series containing all observations occuring at the ends of the months of the input.

If cal is false, only observations occurring the last calendar day of the month are returned

If cal is true, all observation for which the next index is a new month are returned

source

Quarterly

Temporal.boq — Function.

boq(x::TS; cal::Bool=false)

Return a time series containing all observations occuring at the beginnings of the quarters of the input.

If cal is false, only observations occurring the last calendar day of the quarter are returned

If cal is true, all observation for which the previous index is a prior quarter are returned

source

Temporal.eoq — Function.

eoq(x::TS; cal::Bool=false)

Return a time series containing all observations occuring at the ends of the quarters of the input.

If cal is false, only observations occurring the last calendar day of the quarter are returned

If cal is true, all observation for which the next index is a new quarter are returned

source

Yearly

Temporal.boy — Function.

boy(x::TS; cal::Bool=false)

Return a time series containing all observations occuring at the beginnings of the years of the input.

If cal is false, only observations occurring the last calendar day of the year are returned

If cal is true, all observation for which the previous index is a prior year are returned

source

Temporal.eoy — Function.

eoy(x::TS; cal::Bool=false)

Return a time series containing all observations occuring at the ends of the years of the input.

If cal is false, only observations occurring the last calendar day of the year are returned

If cal is true, all observation for which the next index is a new year are returned

source

Collapsing

Temporal.collapse — Function.

collapse(x::TS{V,T}, at::Function; fun::Function=last, args...)::TS{V,T} where {V,T}

Compute the function fun over period boundaries defined by at (e.g. eom, boq, etc.) and returng a time series of the results.

Keyword arguments (args...) are passed to the function fun.

source

julia> using Temporal, Statistics, Dates

julia> X = TS(randn(100, 4))
100x4 TS{Float64,Dates.Date}: 2018-11-06 to 2019-02-13

Index       A        B        C        D
2018-11-06  0.9247   -0.1011  -0.6217  0.9323
2018-11-07  1.2036   -0.3881  -1.4866  1.7032
2018-11-08  0.8145   -0.6455  -0.9634  -0.8325
2018-11-09  1.1524   0.3341   1.9313   -0.5672
2018-11-10  -1.7134  -1.0342  -1.7696  1.9095
2018-11-11  -0.891   -2.0766  0.193    -0.1374
2018-11-12  1.1713   0.8865   -0.3579  -0.6771
⋮
2019-02-06  -0.0317  0.6761   -1.7151  0.1326
2019-02-07  -0.4443  2.4976   -1.141   1.0485
2019-02-08  0.8383   -0.7062  0.9065   0.6306
2019-02-09  -0.1805  -0.2438  0.1703   -0.8664
2019-02-10  0.2437   0.7607   -0.2609  0.1296
2019-02-11  0.7794   -1.0543  0.3285   -0.1835
2019-02-12  -0.1971  -1.3493  -0.3433  1.4639
2019-02-13  0.6534   1.4361   1.251    -0.0952

julia> collapse(X, eom, fun=mean)
3x4 TS{Float64,Dates.Date}: 2018-11-30 to 2019-01-31

Index       A       B        C        D
2018-11-30  0.0029  -0.0227  -0.3272  0.3547
2018-12-31  0.1047  0.1418   0.0925   -0.0144
2019-01-31  0.1542  -0.0869  0.4007   0.0417

julia> last_month = string(X.index[end])[1:7]
"2019-02"

julia> mean(X[last_month])
0.1508564748111179

Summarizing

Temporal.apply — Function.

apply(x::TS{V}, dim::Int=1; fun=sum) where {V}

Apply function fun across dimension dim of a time series x.

If dim is 1, then apply fun to each row of x, returning a time series (TS) object of the results

If dim is 2, then apply fun to each column of x, returning an Array of the results

source

julia> using Temporal, Statistics

julia> X = TS(randn(100, 4))
100x4 TS{Float64,Dates.Date}: 2018-11-06 to 2019-02-13

Index       A        B        C        D
2018-11-06  1.8132   1.1031   0.1996   -0.1711
2018-11-07  -0.2946  0.0853   -0.316   -0.4474
2018-11-08  0.1055   0.2935   0.5593   1.1939
2018-11-09  -0.9514  -0.8276  -1.1189  -0.3054
2018-11-10  1.0496   1.9896   0.4565   0.9046
2018-11-11  0.8779   0.5056   2.0537   0.9471
2018-11-12  1.2665   -0.6338  1.3821   0.1292
⋮
2019-02-06  0.9673   -1.0302  0.7878   0.8872
2019-02-07  -0.3139  -1.3865  -0.0309  0.3596
2019-02-08  1.7711   -0.6551  -0.0863  -1.5153
2019-02-09  0.7761   0.5879   -0.0667  -0.4531
2019-02-10  -0.135   0.4725   -0.3293  -0.2167
2019-02-11  -0.9235  0.5852   -0.6625  0.6765
2019-02-12  -0.0734  -0.71    1.0998   -1.4563
2019-02-13  0.6345   -0.2514  0.358    1.318

julia> apply(X, 1, fun=sum)
100x1 TS{Float64,Dates.Date}: 2018-11-06 to 2019-02-13

Index       Sum
2018-11-06  2.9448
2018-11-07  -0.9727
2018-11-08  2.1522
2018-11-09  -3.2033
2018-11-10  4.4003
2018-11-11  4.3843
2018-11-12  2.1439
⋮
2019-02-06  1.6121
2019-02-07  -1.3717
2019-02-08  -0.4857
2019-02-09  0.8441
2019-02-10  -0.2085
2019-02-11  -0.3244
2019-02-12  -1.1398
2019-02-13  2.0591

julia> apply(X, 2, fun=sum)
1×4 LinearAlgebra.Adjoint{Float64,Array{Float64,1}}:
 7.43539  -8.79226  17.7318  -5.20507