Aggregation

Aggregation

Sampling

Weekdays

Temporal.mondays — Function.

mondays(x::TS)

Return a time series containing all observations occuring on Mondays of the given TS object.

source

Temporal.tuesdays — Function.

tuesdays(x::TS)

Return a time series containing all observations occuring on Tuesdays of the given TS object.

source

Temporal.wednesdays — Function.

wednesdays(x::TS)

Return a time series containing all observations occuring on Wednesdays of the given TS object.

source

Temporal.thursdays — Function.

thursdays(x::TS)

Return a time series containing all observations occuring on Thursdays of the given TS object.

source

Temporal.fridays — Function.

fridays(x::TS)

Return a time series containing all observations occuring on Fridays of the given TS object.

source

Temporal.saturdays — Function.

saturdays(x::TS)

Return a time series containing all observations occuring on Saturdays of the given TS object.

source

Temporal.sundays — Function.

sundays(x::TS)

Return a time series containing all observations occuring on Sundays of the given TS object.

source

Interval Boundaries

Weekly

Temporal.bow — Function.

bow(x::TS; cal::Bool=false)

Return a time series containing all observations occuring at the beginnings of the weeks of the input.

If cal is false, only observations occurring the last calendar day of the week are returned

If cal is true, all observation for which the previous index is a prior week are returned

source

Temporal.eow — Function.

eow(x::TS; cal::Bool=false)

Return a time series containing all observations occuring at the ends of the weeks of the input.

If cal is false, only observations occurring the last calendar day of the week are returned

If cal is true, all observation for which the next index is a new week are returned

source

Monthly

Temporal.bom — Function.

bom(x::TS; cal::Bool=false)

Return a time series containing all observations occuring at the beginnings of the months of the input.

If cal is false, only observations occurring the last calendar day of the month are returned

If cal is true, all observation for which the previous index is a prior month are returned

source

Temporal.eom — Function.

eom(x::TS; cal::Bool=false)

Return a time series containing all observations occuring at the ends of the months of the input.

If cal is false, only observations occurring the last calendar day of the month are returned

If cal is true, all observation for which the next index is a new month are returned

source

Quarterly

Temporal.boq — Function.

boq(x::TS; cal::Bool=false)

Return a time series containing all observations occuring at the beginnings of the quarters of the input.

If cal is false, only observations occurring the last calendar day of the quarter are returned

If cal is true, all observation for which the previous index is a prior quarter are returned

source

Temporal.eoq — Function.

eoq(x::TS; cal::Bool=false)

Return a time series containing all observations occuring at the ends of the quarters of the input.

If cal is false, only observations occurring the last calendar day of the quarter are returned

If cal is true, all observation for which the next index is a new quarter are returned

source

Yearly

Temporal.boy — Function.

boy(x::TS; cal::Bool=false)

Return a time series containing all observations occuring at the beginnings of the years of the input.

If cal is false, only observations occurring the last calendar day of the year are returned

If cal is true, all observation for which the previous index is a prior year are returned

source

Temporal.eoy — Function.

eoy(x::TS; cal::Bool=false)

Return a time series containing all observations occuring at the ends of the years of the input.

If cal is false, only observations occurring the last calendar day of the year are returned

If cal is true, all observation for which the next index is a new year are returned

source

Collapsing

Temporal.collapse — Function.

collapse(x::TS{V,T}, at::Function; fun::Function=last, args...)::TS{V,T} where {V,T}

Compute the function fun over period boundaries defined by at (e.g. eom, boq, etc.) and returng a time series of the results.

Keyword arguments (args...) are passed to the function fun.

source

julia> using Temporal, Statistics, Dates

julia> X = TS(randn(100, 4))
100x4 TS{Float64,Dates.Date}: 2018-11-05 to 2019-02-12

Index       A        B        C        D
2018-11-05  0.425    1.8676   -0.911   0.4706
2018-11-06  -0.2011  0.8107   -1.9547  0.8939
2018-11-07  0.3455   -0.7605  1.8331   -0.1139
2018-11-08  1.3245   -0.0185  0.8697   1.1839
2018-11-09  0.6985   1.4229   -0.3005  -1.4196
2018-11-10  0.1597   -1.7494  1.3806   0.1748
2018-11-11  0.465    -0.279   1.1049   -0.4641
⋮
2019-02-05  0.7088   0.8651   -1.0145  0.4161
2019-02-06  1.2714   -0.846   0.1427   -0.8964
2019-02-07  0.1061   0.5388   -1.5646  -0.9795
2019-02-08  -0.3197  -1.0285  0.2784   0.1205
2019-02-09  -0.7768  0.6926   -1.1746  -0.3646
2019-02-10  -0.6999  -0.9179  1.3971   0.1004
2019-02-11  -2.2161  -0.726   0.6158   -0.5098
2019-02-12  0.4752   0.9951   0.2618   -0.5689

julia> collapse(X, eom, fun=mean)
3x4 TS{Float64,Dates.Date}: 2018-11-30 to 2019-01-31

Index       A        B       C        D
2018-11-30  0.3627   0.2549  0.1819   0.0687
2018-12-31  0.2311   0.2066  0.0731   -0.0493
2019-01-31  -0.1487  0.1687  -0.1598  -0.0253

julia> last_month = string(X.index[end])[1:7]
"2019-02"

julia> mean(X[last_month])
-0.01381143447971587

Summarizing

Temporal.apply — Function.

apply(x::TS{V}, dim::Int=1; fun=sum) where {V}

Apply function fun across dimension dim of a time series x.

If dim is 1, then apply fun to each row of x, returning a time series (TS) object of the results

If dim is 2, then apply fun to each column of x, returning an Array of the results

source

julia> using Temporal, Statistics

julia> X = TS(randn(100, 4))
100x4 TS{Float64,Dates.Date}: 2018-11-05 to 2019-02-12

Index       A        B        C        D
2018-11-05  -0.2582  0.8213   1.2244   0.4108
2018-11-06  -0.6026  0.4626   0.0997   -0.3249
2018-11-07  -0.3574  0.4709   -0.3503  -1.5952
2018-11-08  0.3888   1.2658   -0.4452  -1.2501
2018-11-09  0.5162   -0.3423  0.9567   1.1203
2018-11-10  0.1238   1.3061   -0.6378  1.2393
2018-11-11  -1.7797  -1.9653  1.4144   -0.6091
⋮
2019-02-05  1.4046   0.432    -0.9487  -0.6863
2019-02-06  1.4457   0.4568   0.1345   -0.7176
2019-02-07  0.399    1.609    1.8408   -0.425
2019-02-08  -1.9272  -0.3155  0.21     0.2127
2019-02-09  0.6933   -2.175   1.1493   1.5534
2019-02-10  -0.3628  -2.3905  -0.1996  0.1398
2019-02-11  -0.3714  1.6485   -0.8746  0.1833
2019-02-12  1.8171   -0.4172  -0.3941  -0.9822

julia> apply(X, 1, fun=sum)
100x1 TS{Float64,Dates.Date}: 2018-11-05 to 2019-02-12

Index       Sum
2018-11-05  2.1983
2018-11-06  -0.3653
2018-11-07  -1.8319
2018-11-08  -0.0407
2018-11-09  2.2509
2018-11-10  2.0314
2018-11-11  -2.9397
⋮
2019-02-05  0.2016
2019-02-06  1.3195
2019-02-07  3.4238
2019-02-08  -1.82
2019-02-09  1.221
2019-02-10  -2.8131
2019-02-11  0.5858
2019-02-12  0.0238

julia> apply(X, 2, fun=sum)
1×4 LinearAlgebra.Adjoint{Float64,Array{Float64,1}}:
 -3.94519  0.428786  7.3502  -17.4123