Aggregation

Sampling

Weekdays

Temporal.mondays — Function

mondays(x::TS)

Return a time series containing all observations occuring on Mondays of the given TS object.

Temporal.tuesdays — Function

tuesdays(x::TS)

Return a time series containing all observations occuring on Tuesdays of the given TS object.

Temporal.wednesdays — Function

wednesdays(x::TS)

Return a time series containing all observations occuring on Wednesdays of the given TS object.

Temporal.thursdays — Function

thursdays(x::TS)

Return a time series containing all observations occuring on Thursdays of the given TS object.

Temporal.fridays — Function

fridays(x::TS)

Return a time series containing all observations occuring on Fridays of the given TS object.

Temporal.saturdays — Function

saturdays(x::TS)

Return a time series containing all observations occuring on Saturdays of the given TS object.

Temporal.sundays — Function

sundays(x::TS)

Return a time series containing all observations occuring on Sundays of the given TS object.

Interval Boundaries

Weekly

Temporal.bow — Function

bow(x::TS; cal::Bool=false)

Return a time series containing all observations occuring at the beginnings of the weeks of the input.

If cal is false, only observations occurring the last calendar day of the week are returned

If cal is true, all observation for which the previous index is a prior week are returned

Temporal.eow — Function

eow(x::TS; cal::Bool=false)

Return a time series containing all observations occuring at the ends of the weeks of the input.

If cal is false, only observations occurring the last calendar day of the week are returned

If cal is true, all observation for which the next index is a new week are returned

Monthly

Temporal.bom — Function

bom(x::TS; cal::Bool=false)

Return a time series containing all observations occuring at the beginnings of the months of the input.

If cal is false, only observations occurring the last calendar day of the month are returned

If cal is true, all observation for which the previous index is a prior month are returned

Temporal.eom — Function

eom(x::TS; cal::Bool=false)

Return a time series containing all observations occuring at the ends of the months of the input.

If cal is false, only observations occurring the last calendar day of the month are returned

If cal is true, all observation for which the next index is a new month are returned

Quarterly

Temporal.boq — Function

boq(x::TS; cal::Bool=false)

Return a time series containing all observations occuring at the beginnings of the quarters of the input.

If cal is false, only observations occurring the last calendar day of the quarter are returned

If cal is true, all observation for which the previous index is a prior quarter are returned

Temporal.eoq — Function

eoq(x::TS; cal::Bool=false)

Return a time series containing all observations occuring at the ends of the quarters of the input.

If cal is false, only observations occurring the last calendar day of the quarter are returned

If cal is true, all observation for which the next index is a new quarter are returned

Yearly

Temporal.boy — Function

boy(x::TS; cal::Bool=false)

Return a time series containing all observations occuring at the beginnings of the years of the input.

If cal is false, only observations occurring the last calendar day of the year are returned

If cal is true, all observation for which the previous index is a prior year are returned

Temporal.eoy — Function

eoy(x::TS; cal::Bool=false)

Return a time series containing all observations occuring at the ends of the years of the input.

If cal is false, only observations occurring the last calendar day of the year are returned

If cal is true, all observation for which the next index is a new year are returned

Collapsing

Temporal.collapse — Function

collapse(x::TS, at::Function; fun::Function=last, args...)::TS

Compute the function fun over period boundaries defined by at (e.g. eom, boq, etc.) and returng a time series of the results.

Keyword arguments (args...) are passed to the function fun.

julia> using Temporal, Statistics, Dates

julia> X = TS(randn(100, 4))
100x4 TS{Float64,Dates.Date}: 2019-10-09 to 2020-01-16

Index       A        B        C        D
2019-10-09  -0.7302  0.2601   0.4975   -0.5286
2019-10-10  0.6983   1.0646   0.6224   -0.5284
2019-10-11  -0.7792  0.829    0.7014   -0.9417
2019-10-12  0.2502   -1.9908  1.193    -1.8558
2019-10-13  -0.3931  0.7053   -0.8707  0.3862
2019-10-14  1.0358   -1.2234  1.1985   0.7152
2019-10-15  -0.2894  -1.0319  -1.231   -0.9524
⋮
2020-01-09  0.2952   -0.6848  -0.8419  -0.6214
2020-01-10  -1.6163  -1.4597  -0.2946  0.8694
2020-01-11  -0.6942  0.0839   -0.9361  -1.3343
2020-01-12  0.6498   2.5684   0.4752   -0.4972
2020-01-13  0.0148   0.9143   0.2683   1.2115
2020-01-14  2.2422   -0.9324  0.6211   0.7767
2020-01-15  2.629    -0.8005  1.1533   -1.0565
2020-01-16  0.5788   0.49     0.0718   0.6368

julia> collapse(X, eom, fun=mean)
3x4 TS{Float64,Dates.Date}: 2019-10-31 to 2019-12-31

Index       A        B        C        D
2019-10-31  -0.1326  -0.2095  0.1652   -0.1263
2019-11-30  0.1438   0.2357   -0.2109  0.0881
2019-12-31  0.27     0.0928   0.0365   0.0151

julia> last_month = string(X.index[end])[1:7]
"2020-01"

julia> mean(X[last_month])
0.2151089552061221

Summarizing

Temporal.apply — Function

apply(x::TS, dim::Int=1; fun::Function=sum)

Apply function fun across dimension dim of a time series x.

If dim is 1, then apply fun to each row of x, returning a time series (TS) object of the results

If dim is 2, then apply fun to each column of x, returning an Array of the results

julia> using Temporal, Statistics

julia> X = TS(randn(100, 4))
100x4 TS{Float64,Dates.Date}: 2019-10-09 to 2020-01-16

Index       A        B        C        D
2019-10-09  -0.0106  -1.11    0.8115   -0.9157
2019-10-10  1.5216   -1.1414  1.0558   -0.4153
2019-10-11  -0.5895  0.346    1.1849   0.2054
2019-10-12  -0.3383  -0.551   -0.3833  -0.732
2019-10-13  -1.8687  3.3919   0.7779   -0.0416
2019-10-14  -1.2293  -0.248   -2.2012  -1.6443
2019-10-15  0.8946   -0.1842  -1.6318  0.3735
⋮
2020-01-09  -0.1703  0.6038   -1.4672  0.9032
2020-01-10  0.5112   1.0764   -0.8032  1.085
2020-01-11  -0.3998  -0.2764  -1.4942  0.3238
2020-01-12  0.4184   1.0034   -0.0818  0.2902
2020-01-13  -1.4466  -0.4011  -0.2231  0.2227
2020-01-14  0.2752   0.184    0.6191   0.1031
2020-01-15  1.381    0.6889   -0.6406  0.34
2020-01-16  -0.1414  0.5668   0.6199   -0.4664

julia> apply(X, 1, fun=sum)
100x1 TS{Float64,Dates.Date}: 2019-10-09 to 2020-01-16

Index       Sum
2019-10-09  -1.2248
2019-10-10  1.0206
2019-10-11  1.1468
2019-10-12  -2.0045
2019-10-13  2.2595
2019-10-14  -5.3229
2019-10-15  -0.5479
⋮
2020-01-09  -0.1305
2020-01-10  1.8693
2020-01-11  -1.8465
2020-01-12  1.6303
2020-01-13  -1.848
2020-01-14  1.1815
2020-01-15  1.7693
2020-01-16  0.5789

julia> apply(X, 2, fun=sum)
1×4 LinearAlgebra.Adjoint{Float64,Array{Float64,1}}:
 -1.55642  9.57397  -1.66947  2.31548