Aggregation

Aggregation

Sampling

Weekdays

Temporal.mondays — Function.

mondays(x::TS)

Return a time series containing all observations occuring on Mondays of the given TS object.

source

Temporal.tuesdays — Function.

tuesdays(x::TS)

Return a time series containing all observations occuring on Tuesdays of the given TS object.

source

Temporal.wednesdays — Function.

wednesdays(x::TS)

Return a time series containing all observations occuring on Wednesdays of the given TS object.

source

Temporal.thursdays — Function.

thursdays(x::TS)

Return a time series containing all observations occuring on Thursdays of the given TS object.

source

Temporal.fridays — Function.

fridays(x::TS)

Return a time series containing all observations occuring on Fridays of the given TS object.

source

Temporal.saturdays — Function.

saturdays(x::TS)

Return a time series containing all observations occuring on Saturdays of the given TS object.

source

Temporal.sundays — Function.

sundays(x::TS)

Return a time series containing all observations occuring on Sundays of the given TS object.

source

Interval Boundaries

Weekly

Temporal.bow — Function.

bow(x::TS; cal::Bool=false)

Return a time series containing all observations occuring at the beginnings of the weeks of the input.

If cal is false, only observations occurring the last calendar day of the week are returned

If cal is true, all observation for which the previous index is a prior week are returned

source

Temporal.eow — Function.

eow(x::TS; cal::Bool=false)

Return a time series containing all observations occuring at the ends of the weeks of the input.

If cal is false, only observations occurring the last calendar day of the week are returned

If cal is true, all observation for which the next index is a new week are returned

source

Monthly

Temporal.bom — Function.

bom(x::TS; cal::Bool=false)

Return a time series containing all observations occuring at the beginnings of the months of the input.

If cal is false, only observations occurring the last calendar day of the month are returned

If cal is true, all observation for which the previous index is a prior month are returned

source

Temporal.eom — Function.

eom(x::TS; cal::Bool=false)

Return a time series containing all observations occuring at the ends of the months of the input.

If cal is false, only observations occurring the last calendar day of the month are returned

If cal is true, all observation for which the next index is a new month are returned

source

Quarterly

Temporal.boq — Function.

boq(x::TS; cal::Bool=false)

Return a time series containing all observations occuring at the beginnings of the quarters of the input.

If cal is false, only observations occurring the last calendar day of the quarter are returned

If cal is true, all observation for which the previous index is a prior quarter are returned

source

Temporal.eoq — Function.

eoq(x::TS; cal::Bool=false)

Return a time series containing all observations occuring at the ends of the quarters of the input.

If cal is false, only observations occurring the last calendar day of the quarter are returned

If cal is true, all observation for which the next index is a new quarter are returned

source

Yearly

Temporal.boy — Function.

boy(x::TS; cal::Bool=false)

Return a time series containing all observations occuring at the beginnings of the years of the input.

If cal is false, only observations occurring the last calendar day of the year are returned

If cal is true, all observation for which the previous index is a prior year are returned

source

Temporal.eoy — Function.

eoy(x::TS; cal::Bool=false)

Return a time series containing all observations occuring at the ends of the years of the input.

If cal is false, only observations occurring the last calendar day of the year are returned

If cal is true, all observation for which the next index is a new year are returned

source

Collapsing

Temporal.collapse — Function.

collapse(x::TS{V,T}, at::Function; fun::Function=last, args...)::TS{V,T} where {V,T}

Compute the function fun over period boundaries defined by at (e.g. eom, boq, etc.) and returng a time series of the results.

Keyword arguments (args...) are passed to the function fun.

source

julia> using Temporal, Statistics, Dates

julia> X = TS(randn(100, 4))
100x4 TS{Float64,Dates.Date}: 2018-11-12 to 2019-02-19

Index       A        B        C        D
2018-11-12  -0.0456  0.1473   -0.1514  -0.2594
2018-11-13  0.3537   0.8406   0.7183   -1.419
2018-11-14  0.0995   1.9058   0.6748   -0.5448
2018-11-15  1.0102   -1.9273  0.4454   0.3112
2018-11-16  0.8911   1.4314   -0.1982  -0.7571
2018-11-17  0.8709   -1.9453  0.1205   -0.4764
2018-11-18  -1.2494  -0.4777  0.5043   0.598
⋮
2019-02-12  -0.0029  -0.5527  -1.1558  0.462
2019-02-13  1.9182   -0.9346  0.3552   -1.3001
2019-02-14  -0.5714  1.21     0.5531   0.4087
2019-02-15  -0.9153  -0.8261  1.4658   -0.2939
2019-02-16  0.8003   -0.2088  -0.3346  -1.2505
2019-02-17  -0.2356  -0.6002  -1.8346  0.7683
2019-02-18  0.9213   -1.1454  1.4038   -0.4475
2019-02-19  -0.4081  0.0071   1.0337   0.5486

julia> collapse(X, eom, fun=mean)
3x4 TS{Float64,Dates.Date}: 2018-11-30 to 2019-01-31

Index       A        B       C        D
2018-11-30  0.2407   -0.459  0.2602   0.1894
2018-12-31  -0.3555  0.4505  -0.2619  -0.0426
2019-01-31  -0.1579  0.0409  -0.1726  0.1729

julia> last_month = string(X.index[end])[1:7]
"2019-02"

julia> mean(X[last_month])
0.04660389368924249

Summarizing

Temporal.apply — Function.

apply(x::TS{V}, dim::Int=1; fun=sum) where {V}

Apply function fun across dimension dim of a time series x.

If dim is 1, then apply fun to each row of x, returning a time series (TS) object of the results

If dim is 2, then apply fun to each column of x, returning an Array of the results

source

julia> using Temporal, Statistics

julia> X = TS(randn(100, 4))
100x4 TS{Float64,Dates.Date}: 2018-11-12 to 2019-02-19

Index       A        B        C        D
2018-11-12  0.0229   -0.1428  0.5411   -0.2542
2018-11-13  -0.3514  0.9252   0.0851   -0.4472
2018-11-14  0.6381   -1.727   0.262    0.4508
2018-11-15  -0.4555  1.7569   1.0693   -2.0925
2018-11-16  0.5372   -1.916   2.1096   -0.1049
2018-11-17  -1.8901  -0.3423  2.1042   0.3018
2018-11-18  -0.3516  -0.1912  -1.6388  -0.0273
⋮
2019-02-12  1.8139   -1.373   -1.5426  0.3481
2019-02-13  -1.1108  -1.989   -0.0253  0.8179
2019-02-14  0.5317   0.0292   -0.4336  0.2217
2019-02-15  0.7854   0.5806   0.1425   -0.1493
2019-02-16  -1.0413  -0.8995  -0.6525  1.2515
2019-02-17  0.6529   -0.611   0.6258   -1.1878
2019-02-18  0.2045   -0.1256  0.5358   -1.1927
2019-02-19  0.4954   -0.5709  -1.2679  -2.2807

julia> apply(X, 1, fun=sum)
100x1 TS{Float64,Dates.Date}: 2018-11-12 to 2019-02-19

Index       Sum
2018-11-12  0.167
2018-11-13  0.2117
2018-11-14  -0.3761
2018-11-15  0.2782
2018-11-16  0.6259
2018-11-17  0.1736
2018-11-18  -2.209
⋮
2019-02-12  -0.7537
2019-02-13  -2.3072
2019-02-14  0.349
2019-02-15  1.3591
2019-02-16  -1.3418
2019-02-17  -0.5201
2019-02-18  -0.5781
2019-02-19  -3.624

julia> apply(X, 2, fun=sum)
1×4 LinearAlgebra.Adjoint{Float64,Array{Float64,1}}:
 1.67248  2.26979  4.14349  -9.86479