Aggregation

Aggregation

Sampling

Weekdays

Temporal.mondays — Function.

mondays(x::TS)

Return a time series containing all observations occuring on Mondays of the given TS object.

source

Temporal.tuesdays — Function.

tuesdays(x::TS)

Return a time series containing all observations occuring on Tuesdays of the given TS object.

source

Temporal.wednesdays — Function.

wednesdays(x::TS)

Return a time series containing all observations occuring on Wednesdays of the given TS object.

source

Temporal.thursdays — Function.

thursdays(x::TS)

Return a time series containing all observations occuring on Thursdays of the given TS object.

source

Temporal.fridays — Function.

fridays(x::TS)

Return a time series containing all observations occuring on Fridays of the given TS object.

source

Temporal.saturdays — Function.

saturdays(x::TS)

Return a time series containing all observations occuring on Saturdays of the given TS object.

source

Temporal.sundays — Function.

sundays(x::TS)

Return a time series containing all observations occuring on Sundays of the given TS object.

source

Interval Boundaries

Weekly

Temporal.bow — Function.

bow(x::TS; cal::Bool=false)

Return a time series containing all observations occuring at the beginnings of the weeks of the input.

If cal is false, only observations occurring the last calendar day of the week are returned

If cal is true, all observation for which the previous index is a prior week are returned

source

Temporal.eow — Function.

eow(x::TS; cal::Bool=false)

Return a time series containing all observations occuring at the ends of the weeks of the input.

If cal is false, only observations occurring the last calendar day of the week are returned

If cal is true, all observation for which the next index is a new week are returned

source

Monthly

Temporal.bom — Function.

bom(x::TS; cal::Bool=false)

Return a time series containing all observations occuring at the beginnings of the months of the input.

If cal is false, only observations occurring the last calendar day of the month are returned

If cal is true, all observation for which the previous index is a prior month are returned

source

Temporal.eom — Function.

eom(x::TS; cal::Bool=false)

Return a time series containing all observations occuring at the ends of the months of the input.

If cal is false, only observations occurring the last calendar day of the month are returned

If cal is true, all observation for which the next index is a new month are returned

source

Quarterly

Temporal.boq — Function.

boq(x::TS; cal::Bool=false)

Return a time series containing all observations occuring at the beginnings of the quarters of the input.

If cal is false, only observations occurring the last calendar day of the quarter are returned

If cal is true, all observation for which the previous index is a prior quarter are returned

source

Temporal.eoq — Function.

eoq(x::TS; cal::Bool=false)

Return a time series containing all observations occuring at the ends of the quarters of the input.

If cal is false, only observations occurring the last calendar day of the quarter are returned

If cal is true, all observation for which the next index is a new quarter are returned

source

Yearly

Temporal.boy — Function.

boy(x::TS; cal::Bool=false)

Return a time series containing all observations occuring at the beginnings of the years of the input.

If cal is false, only observations occurring the last calendar day of the year are returned

If cal is true, all observation for which the previous index is a prior year are returned

source

Temporal.eoy — Function.

eoy(x::TS; cal::Bool=false)

Return a time series containing all observations occuring at the ends of the years of the input.

If cal is false, only observations occurring the last calendar day of the year are returned

If cal is true, all observation for which the next index is a new year are returned

source

Collapsing

Temporal.collapse — Function.

collapse(x::TS{V,T}, at::Function; fun::Function=last, args...)::TS{V,T} where {V,T}

Compute the function fun over period boundaries defined by at (e.g. eom, boq, etc.) and returng a time series of the results.

Keyword arguments (args...) are passed to the function fun.

source

julia> using Temporal, Statistics, Dates

julia> X = TS(randn(100, 4))
100x4 TS{Float64,Dates.Date}: 2019-07-24 to 2019-10-31

Index       A        B        C        D
2019-07-24  -0.2976  0.8885   0.5891   0.7586
2019-07-25  0.3544   -1.5273  -0.2736  -0.8375
2019-07-26  0.8454   1.2192   -1.2488  0.53
2019-07-27  -1.5916  0.7345   0.6978   -0.6318
2019-07-28  1.6876   -1.5845  -0.4042  0.0215
2019-07-29  -0.6653  -0.9818  1.4495   1.5495
2019-07-30  -0.2351  0.3356   0.1389   0.0302
⋮
2019-10-24  0.6188   -2.3665  0.797    -0.7059
2019-10-25  0.8423   -0.6137  0.1539   0.9912
2019-10-26  -1.4262  1.1098   -0.1     -1.0857
2019-10-27  -0.5709  -2.5471  0.608    1.6041
2019-10-28  0.4724   -0.221   1.6809   1.7936
2019-10-29  0.1856   0.3181   -0.0528  -0.2578
2019-10-30  0.9419   -0.0343  -0.4256  1.4848
2019-10-31  -0.1996  0.8228   1.3253   1.3294

julia> collapse(X, eom, fun=mean)
3x4 TS{Float64,Dates.Date}: 2019-07-31 to 2019-09-30

Index       A       B        C        D
2019-07-31  0.1718  -0.1844  -0.1066  0.1862
2019-08-31  0.3499  -0.1195  -0.1134  0.1022
2019-09-30  0.0273  0.1306   -0.0012  0.053

julia> last_month = string(X.index[end])[1:7]
"2019-10"

julia> mean(X[last_month])
0.05185788295820775

Summarizing

Temporal.apply — Function.

apply(x::TS{V}, dim::Int=1; fun=sum) where {V}

Apply function fun across dimension dim of a time series x.

If dim is 1, then apply fun to each row of x, returning a time series (TS) object of the results

If dim is 2, then apply fun to each column of x, returning an Array of the results

source

julia> using Temporal, Statistics

julia> X = TS(randn(100, 4))
100x4 TS{Float64,Dates.Date}: 2019-07-24 to 2019-10-31

Index       A        B        C        D
2019-07-24  0.7092   -1.8634  -0.0829  0.4949
2019-07-25  0.4604   0.8538   -1.445   -0.2344
2019-07-26  0.5796   -0.7909  2.3475   1.0899
2019-07-27  0.6703   0.1545   0.2629   0.1591
2019-07-28  -0.4509  0.8954   -0.7402  -0.1712
2019-07-29  -0.419   -0.9283  0.9626   1.468
2019-07-30  -0.5446  0.0168   -1.3182  -1.2339
⋮
2019-10-24  0.6061   0.8989   -0.9143  0.7308
2019-10-25  -1.2597  -0.2223  0.9144   -1.6353
2019-10-26  1.5785   -1.3752  -0.4603  -0.0059
2019-10-27  0.564    0.9714   -0.0235  1.5719
2019-10-28  -0.0393  -0.363   1.1089   -0.5388
2019-10-29  1.7704   -0.8532  0.894    -0.9828
2019-10-30  -0.354   0.5045   -0.4724  -0.7719
2019-10-31  0.5005   0.301    2.0446   -0.7645

julia> apply(X, 1, fun=sum)
100x1 TS{Float64,Dates.Date}: 2019-07-24 to 2019-10-31

Index       Sum
2019-07-24  -0.7421
2019-07-25  -0.3651
2019-07-26  3.2261
2019-07-27  1.2468
2019-07-28  -0.4669
2019-07-29  1.0833
2019-07-30  -3.0799
⋮
2019-10-24  1.3215
2019-10-25  -2.203
2019-10-26  -0.2629
2019-10-27  3.0838
2019-10-28  0.1678
2019-10-29  0.8285
2019-10-30  -1.0938
2019-10-31  2.0816

julia> apply(X, 2, fun=sum)
1×4 LinearAlgebra.Adjoint{Float64,Array{Float64,1}}:
 1.3531  -13.7915  18.4283  -9.33233