Aggregation

Sampling

Weekdays

Temporal.mondays — Function

mondays(x::TS)

Return a time series containing all observations occuring on Mondays of the given TS object.

Temporal.tuesdays — Function

tuesdays(x::TS)

Return a time series containing all observations occuring on Tuesdays of the given TS object.

Temporal.wednesdays — Function

wednesdays(x::TS)

Return a time series containing all observations occuring on Wednesdays of the given TS object.

Temporal.thursdays — Function

thursdays(x::TS)

Return a time series containing all observations occuring on Thursdays of the given TS object.

Temporal.fridays — Function

fridays(x::TS)

Return a time series containing all observations occuring on Fridays of the given TS object.

Temporal.saturdays — Function

saturdays(x::TS)

Return a time series containing all observations occuring on Saturdays of the given TS object.

Temporal.sundays — Function

sundays(x::TS)

Return a time series containing all observations occuring on Sundays of the given TS object.

Interval Boundaries

Weekly

Temporal.bow — Function

bow(x::TS; cal::Bool=false)

Return a time series containing all observations occuring at the beginnings of the weeks of the input.

If cal is false, only observations occurring the last calendar day of the week are returned

If cal is true, all observation for which the previous index is a prior week are returned

Temporal.eow — Function

eow(x::TS; cal::Bool=false)

Return a time series containing all observations occuring at the ends of the weeks of the input.

If cal is false, only observations occurring the last calendar day of the week are returned

If cal is true, all observation for which the next index is a new week are returned

Monthly

Temporal.bom — Function

bom(x::TS; cal::Bool=false)

Return a time series containing all observations occuring at the beginnings of the months of the input.

If cal is false, only observations occurring the last calendar day of the month are returned

If cal is true, all observation for which the previous index is a prior month are returned

Temporal.eom — Function

eom(x::TS; cal::Bool=false)

Return a time series containing all observations occuring at the ends of the months of the input.

If cal is false, only observations occurring the last calendar day of the month are returned

If cal is true, all observation for which the next index is a new month are returned

Quarterly

Temporal.boq — Function

boq(x::TS; cal::Bool=false)

Return a time series containing all observations occuring at the beginnings of the quarters of the input.

If cal is false, only observations occurring the last calendar day of the quarter are returned

If cal is true, all observation for which the previous index is a prior quarter are returned

Temporal.eoq — Function

eoq(x::TS; cal::Bool=false)

Return a time series containing all observations occuring at the ends of the quarters of the input.

If cal is false, only observations occurring the last calendar day of the quarter are returned

If cal is true, all observation for which the next index is a new quarter are returned

Yearly

Temporal.boy — Function

boy(x::TS; cal::Bool=false)

Return a time series containing all observations occuring at the beginnings of the years of the input.

If cal is false, only observations occurring the last calendar day of the year are returned

If cal is true, all observation for which the previous index is a prior year are returned

Temporal.eoy — Function

eoy(x::TS; cal::Bool=false)

Return a time series containing all observations occuring at the ends of the years of the input.

If cal is false, only observations occurring the last calendar day of the year are returned

If cal is true, all observation for which the next index is a new year are returned

Collapsing

Temporal.collapse — Function

collapse(x::TS{V,T}, at::Function; fun::Function=last, args...)::TS{V,T} where {V,T}

Compute the function fun over period boundaries defined by at (e.g. eom, boq, etc.) and returng a time series of the results.

Keyword arguments (args...) are passed to the function fun.

julia> using Temporal, Statistics, Dates

julia> X = TS(randn(100, 4))
100x4 TS{Float64,Dates.Date}: 2019-09-10 to 2019-12-18

Index       A        B        C        D
2019-09-10  1.408    -1.1322  0.1089   -0.0098
2019-09-11  1.0517   0.6333   1.5263   0.6739
2019-09-12  0.1032   1.421    -0.7046  0.0464
2019-09-13  0.4984   0.524    0.8709   -0.2199
2019-09-14  -1.6193  1.2555   0.8713   1.1116
2019-09-15  1.1955   0.0143   -0.4649  0.8897
2019-09-16  -0.9354  0.231    -1.2353  -0.4033
⋮
2019-12-11  0.3284   -0.9906  -1.2946  1.512
2019-12-12  -0.002   -0.6225  0.9681   1.7235
2019-12-13  0.1856   0.527    0.3269   -0.4176
2019-12-14  1.3645   0.8526   -0.9456  -0.4699
2019-12-15  -0.3791  0.0631   0.4853   0.4071
2019-12-16  0.0852   -0.9686  -2.3618  0.6792
2019-12-17  -0.05    0.4691   1.3157   -1.94
2019-12-18  0.7054   -0.632   1.3701   1.8384

julia> collapse(X, eom, fun=mean)
3x4 TS{Float64,Dates.Date}: 2019-09-30 to 2019-11-30

Index       A        B       C       D
2019-09-30  -0.0836  0.1743  0.1112  0.1229
2019-10-31  0.0128   0.0026  0.1137  0.2576
2019-11-30  0.073    0.2493  0.2943  0.0536

julia> last_month = string(X.index[end])[1:7]
"2019-12"

julia> mean(X[last_month])
0.09379122604983177

Summarizing

Temporal.apply — Function

apply(x::TS{V}, dim::Int=1; fun=sum) where {V}

Apply function fun across dimension dim of a time series x.

If dim is 1, then apply fun to each row of x, returning a time series (TS) object of the results

If dim is 2, then apply fun to each column of x, returning an Array of the results

julia> using Temporal, Statistics

julia> X = TS(randn(100, 4))
100x4 TS{Float64,Dates.Date}: 2019-09-10 to 2019-12-18

Index       A        B        C        D
2019-09-10  -0.6707  0.5257   0.6643   -0.3159
2019-09-11  -1.677   -0.3016  -0.3742  0.7225
2019-09-12  -1.0425  0.1674   -1.2667  -0.4097
2019-09-13  -0.0256  0.2018   -0.8482  -0.9491
2019-09-14  0.3472   0.3016   -0.7894  0.7275
2019-09-15  0.2477   0.6251   -0.7763  1.0452
2019-09-16  0.0671   -2.2906  -0.0327  -0.4777
⋮
2019-12-11  1.5178   0.0626   -1.228   -0.6447
2019-12-12  -0.4675  -0.6753  -1.7641  0.5244
2019-12-13  1.3075   1.9789   1.2396   1.1915
2019-12-14  1.6201   0.1419   0.1243   0.755
2019-12-15  -0.7521  -1.0128  1.165    -1.2235
2019-12-16  0.7035   -0.41    -1.3183  1.4711
2019-12-17  -0.3593  -1.702   -0.9452  -1.5811
2019-12-18  0.6713   -0.2173  -0.6259  0.722

julia> apply(X, 1, fun=sum)
100x1 TS{Float64,Dates.Date}: 2019-09-10 to 2019-12-18

Index       Sum
2019-09-10  0.2034
2019-09-11  -1.6302
2019-09-12  -2.5515
2019-09-13  -1.6211
2019-09-14  0.5868
2019-09-15  1.1416
2019-09-16  -2.734
⋮
2019-12-11  -0.2923
2019-12-12  -2.3824
2019-12-13  5.7175
2019-12-14  2.6413
2019-12-15  -1.8234
2019-12-16  0.4462
2019-12-17  -4.5876
2019-12-18  0.55

julia> apply(X, 2, fun=sum)
1×4 LinearAlgebra.Adjoint{Float64,Array{Float64,1}}:
 -10.0379  -3.47233  -5.46262  11.5445