Aggregation

Aggregation

Sampling

Weekdays

Temporal.mondays — Function.

mondays(x::TS)

Return a time series containing all observations occuring on Mondays of the given TS object.

source

Temporal.tuesdays — Function.

tuesdays(x::TS)

Return a time series containing all observations occuring on Tuesdays of the given TS object.

source

Temporal.wednesdays — Function.

wednesdays(x::TS)

Return a time series containing all observations occuring on Wednesdays of the given TS object.

source

Temporal.thursdays — Function.

thursdays(x::TS)

Return a time series containing all observations occuring on Thursdays of the given TS object.

source

Temporal.fridays — Function.

fridays(x::TS)

Return a time series containing all observations occuring on Fridays of the given TS object.

source

Temporal.saturdays — Function.

saturdays(x::TS)

Return a time series containing all observations occuring on Saturdays of the given TS object.

source

Temporal.sundays — Function.

sundays(x::TS)

Return a time series containing all observations occuring on Sundays of the given TS object.

source

Interval Boundaries

Weekly

Temporal.bow — Function.

bow(x::TS; cal::Bool=false)

Return a time series containing all observations occuring at the beginnings of the weeks of the input.

If cal is false, only observations occurring the last calendar day of the week are returned

If cal is true, all observation for which the previous index is a prior week are returned

source

Temporal.eow — Function.

eow(x::TS; cal::Bool=false)

Return a time series containing all observations occuring at the ends of the weeks of the input.

If cal is false, only observations occurring the last calendar day of the week are returned

If cal is true, all observation for which the next index is a new week are returned

source

Monthly

Temporal.bom — Function.

bom(x::TS; cal::Bool=false)

Return a time series containing all observations occuring at the beginnings of the months of the input.

If cal is false, only observations occurring the last calendar day of the month are returned

If cal is true, all observation for which the previous index is a prior month are returned

source

Temporal.eom — Function.

eom(x::TS; cal::Bool=false)

Return a time series containing all observations occuring at the ends of the months of the input.

If cal is false, only observations occurring the last calendar day of the month are returned

If cal is true, all observation for which the next index is a new month are returned

source

Quarterly

Temporal.boq — Function.

boq(x::TS; cal::Bool=false)

Return a time series containing all observations occuring at the beginnings of the quarters of the input.

If cal is false, only observations occurring the last calendar day of the quarter are returned

If cal is true, all observation for which the previous index is a prior quarter are returned

source

Temporal.eoq — Function.

eoq(x::TS; cal::Bool=false)

Return a time series containing all observations occuring at the ends of the quarters of the input.

If cal is false, only observations occurring the last calendar day of the quarter are returned

If cal is true, all observation for which the next index is a new quarter are returned

source

Yearly

Temporal.boy — Function.

boy(x::TS; cal::Bool=false)

Return a time series containing all observations occuring at the beginnings of the years of the input.

If cal is false, only observations occurring the last calendar day of the year are returned

If cal is true, all observation for which the previous index is a prior year are returned

source

Temporal.eoy — Function.

eoy(x::TS; cal::Bool=false)

Return a time series containing all observations occuring at the ends of the years of the input.

If cal is false, only observations occurring the last calendar day of the year are returned

If cal is true, all observation for which the next index is a new year are returned

source

Collapsing

Temporal.collapse — Function.

collapse(x::TS{V,T}, at::Function; fun::Function=last, args...)::TS{V,T} where {V,T}

Compute the function fun over period boundaries defined by at (e.g. eom, boq, etc.) and returng a time series of the results.

Keyword arguments (args...) are passed to the function fun.

source

julia> using Temporal, Statistics, Dates

julia> X = TS(randn(100, 4))
100x4 TS{Float64,Dates.Date}: 2019-07-23 to 2019-10-30

Index       A        B        C        D
2019-07-23  -0.5509  0.095    -1.421   1.3695
2019-07-24  -0.0569  2.7511   -0.4091  -1.3007
2019-07-25  -0.6035  0.1486   -1.2923  -0.349
2019-07-26  1.2617   0.1029   0.3026   0.0342
2019-07-27  1.6185   0.5041   1.6718   0.9679
2019-07-28  -0.4459  -0.1911  1.2449   -0.249
2019-07-29  1.9683   0.7204   0.9269   -0.2509
⋮
2019-10-23  -1.1189  0.4519   -0.5006  -1.9032
2019-10-24  -0.891   -0.7899  -0.6549  0.8984
2019-10-25  -1.1505  0.3291   -1.2022  0.1318
2019-10-26  1.3417   2.3487   -0.0701  -0.125
2019-10-27  -1.0212  0.6983   0.0824   0.0186
2019-10-28  0.1018   -0.7025  0.4053   0.1385
2019-10-29  1.5746   0.0215   0.0552   0.1182
2019-10-30  -0.9133  -1.2229  -1.269   0.4768

julia> collapse(X, eom, fun=mean)
3x4 TS{Float64,Dates.Date}: 2019-07-31 to 2019-09-30

Index       A       B       C        D
2019-07-31  0.3065  0.4161  0.283    0.0665
2019-08-31  0.0326  0.0919  -0.0005  0.0501
2019-09-30  0.1792  0.1275  -0.2358  0.1651

julia> last_month = string(X.index[end])[1:7]
"2019-10"

julia> mean(X[last_month])
0.10574873022557216

Summarizing

Temporal.apply — Function.

apply(x::TS{V}, dim::Int=1; fun=sum) where {V}

Apply function fun across dimension dim of a time series x.

If dim is 1, then apply fun to each row of x, returning a time series (TS) object of the results

If dim is 2, then apply fun to each column of x, returning an Array of the results

source

julia> using Temporal, Statistics

julia> X = TS(randn(100, 4))
100x4 TS{Float64,Dates.Date}: 2019-07-23 to 2019-10-30

Index       A        B        C        D
2019-07-23  0.4633   1.6389   0.8227   -0.0634
2019-07-24  0.2651   -0.2364  -0.416   -1.893
2019-07-25  0.4413   -0.3061  0.3012   -0.7591
2019-07-26  -0.7326  1.1259   1.3111   1.6936
2019-07-27  -0.1366  -0.0717  -0.718   -0.5293
2019-07-28  1.3451   1.5804   1.7171   -0.2612
2019-07-29  0.9881   0.1943   -0.9478  0.209
⋮
2019-10-23  0.1619   0.1974   1.4525   -1.0804
2019-10-24  0.1628   0.1272   1.6878   -0.3708
2019-10-25  0.1526   -0.3894  -0.9368  -0.1612
2019-10-26  -3.8389  -1.6481  -0.5087  -0.8262
2019-10-27  0.1138   0.7279   0.1034   -0.8971
2019-10-28  0.1403   1.0486   -0.7381  -0.0349
2019-10-29  0.0545   -0.5084  -1.5247  0.5783
2019-10-30  2.0086   2.328    -0.2699  0.4138

julia> apply(X, 1, fun=sum)
100x1 TS{Float64,Dates.Date}: 2019-07-23 to 2019-10-30

Index       Sum
2019-07-23  2.8615
2019-07-24  -2.2803
2019-07-25  -0.3227
2019-07-26  3.3979
2019-07-27  -1.4556
2019-07-28  4.3814
2019-07-29  0.4435
⋮
2019-10-23  0.7313
2019-10-24  1.607
2019-10-25  -1.3348
2019-10-26  -6.8218
2019-10-27  0.048
2019-10-28  0.4159
2019-10-29  -1.4003
2019-10-30  4.4806

julia> apply(X, 2, fun=sum)
1×4 LinearAlgebra.Adjoint{Float64,Array{Float64,1}}:
 -16.7785  10.3719  14.3539  -12.1597