Aggregation

Aggregation

Sampling

Weekdays

Temporal.mondays — Function.

mondays(x::TS)

Return a time series containing all observations occuring on Mondays of the given TS object.

source

Temporal.tuesdays — Function.

tuesdays(x::TS)

Return a time series containing all observations occuring on Tuesdays of the given TS object.

source

Temporal.wednesdays — Function.

wednesdays(x::TS)

Return a time series containing all observations occuring on Wednesdays of the given TS object.

source

Temporal.thursdays — Function.

thursdays(x::TS)

Return a time series containing all observations occuring on Thursdays of the given TS object.

source

Temporal.fridays — Function.

fridays(x::TS)

Return a time series containing all observations occuring on Fridays of the given TS object.

source

Temporal.saturdays — Function.

saturdays(x::TS)

Return a time series containing all observations occuring on Saturdays of the given TS object.

source

Temporal.sundays — Function.

sundays(x::TS)

Return a time series containing all observations occuring on Sundays of the given TS object.

source

Interval Boundaries

Weekly

Temporal.bow — Function.

bow(x::TS; cal::Bool=false)

Return a time series containing all observations occuring at the beginnings of the weeks of the input.

If cal is false, only observations occurring the last calendar day of the week are returned

If cal is true, all observation for which the previous index is a prior week are returned

source

Temporal.eow — Function.

eow(x::TS; cal::Bool=false)

Return a time series containing all observations occuring at the ends of the weeks of the input.

If cal is false, only observations occurring the last calendar day of the week are returned

If cal is true, all observation for which the next index is a new week are returned

source

Monthly

Temporal.bom — Function.

bom(x::TS; cal::Bool=false)

Return a time series containing all observations occuring at the beginnings of the months of the input.

If cal is false, only observations occurring the last calendar day of the month are returned

If cal is true, all observation for which the previous index is a prior month are returned

source

Temporal.eom — Function.

eom(x::TS; cal::Bool=false)

Return a time series containing all observations occuring at the ends of the months of the input.

If cal is false, only observations occurring the last calendar day of the month are returned

If cal is true, all observation for which the next index is a new month are returned

source

Quarterly

Temporal.boq — Function.

boq(x::TS; cal::Bool=false)

Return a time series containing all observations occuring at the beginnings of the quarters of the input.

If cal is false, only observations occurring the last calendar day of the quarter are returned

If cal is true, all observation for which the previous index is a prior quarter are returned

source

Temporal.eoq — Function.

eoq(x::TS; cal::Bool=false)

Return a time series containing all observations occuring at the ends of the quarters of the input.

If cal is false, only observations occurring the last calendar day of the quarter are returned

If cal is true, all observation for which the next index is a new quarter are returned

source

Yearly

Temporal.boy — Function.

boy(x::TS; cal::Bool=false)

Return a time series containing all observations occuring at the beginnings of the years of the input.

If cal is false, only observations occurring the last calendar day of the year are returned

If cal is true, all observation for which the previous index is a prior year are returned

source

Temporal.eoy — Function.

eoy(x::TS; cal::Bool=false)

Return a time series containing all observations occuring at the ends of the years of the input.

If cal is false, only observations occurring the last calendar day of the year are returned

If cal is true, all observation for which the next index is a new year are returned

source

Collapsing

Temporal.collapse — Function.

collapse(x::TS{V,T}, at::Function; fun::Function=last, args...)::TS{V,T} where {V,T}

Compute the function fun over period boundaries defined by at (e.g. eom, boq, etc.) and returng a time series of the results.

Keyword arguments (args...) are passed to the function fun.

source

julia> using Temporal, Statistics, Dates

julia> X = TS(randn(100, 4))
100x4 TS{Float64,Dates.Date}: 2018-11-02 to 2019-02-09

Index       A        B        C        D
2018-11-02  1.5165   -1.7651  -0.8835  0.7949
2018-11-03  0.2055   0.0468   -0.7589  -1.1442
2018-11-04  0.5007   0.627    -0.8609  -0.7873
2018-11-05  -1.8808  -1.4636  -0.2289  0.901
2018-11-06  -0.8768  -0.787   0.8736   0.8685
2018-11-07  -0.0374  -0.0922  -0.7234  -0.9683
2018-11-08  0.3722   0.4717   0.1048   0.0675
⋮
2019-02-02  1.2992   -0.5791  0.8944   0.0198
2019-02-03  -0.829   -0.7153  0.676    0.976
2019-02-04  0.2549   0.451    0.1968   0.6942
2019-02-05  0.1262   -0.9669  0.0176   1.0449
2019-02-06  2.7519   -0.2641  1.1673   -0.9773
2019-02-07  1.138    0.577    0.9352   -0.1136
2019-02-08  2.3612   -0.1377  0.9603   -0.1339
2019-02-09  0.0351   0.2735   0.6319   0.3084

julia> collapse(X, eom, fun=mean)
3x4 TS{Float64,Dates.Date}: 2018-11-30 to 2019-01-31

Index       A        B        C        D
2018-11-30  0.0454   -0.1076  -0.1703  -0.0282
2018-12-31  -0.0181  0.2384   0.1697   0.0498
2019-01-31  0.2083   -0.021   0.0183   -0.2074

julia> last_month = string(X.index[end])[1:7]
"2019-02"

julia> mean(X[last_month])
0.1994197818099902

Summarizing

Temporal.apply — Function.

apply(x::TS{V}, dim::Int=1; fun=sum) where {V}

Apply function fun across dimension dim of a time series x.

If dim is 1, then apply fun to each row of x, returning a time series (TS) object of the results

If dim is 2, then apply fun to each column of x, returning an Array of the results

source

julia> using Temporal, Statistics

julia> X = TS(randn(100, 4))
100x4 TS{Float64,Dates.Date}: 2018-11-02 to 2019-02-09

Index       A        B        C        D
2018-11-02  0.993    0.679    -0.9067  -0.0368
2018-11-03  -0.5494  1.6599   -1.1514  0.2322
2018-11-04  0.4486   -0.5012  -0.0051  -0.0915
2018-11-05  -1.0494  -0.4288  -1.1852  -0.5807
2018-11-06  -0.0341  -1.0676  0.1694   1.1273
2018-11-07  -0.3454  0.9883   1.4655   -1.3557
2018-11-08  0.9158   0.824    -1.1436  2.1734
⋮
2019-02-02  0.1323   0.0601   -1.0277  0.9196
2019-02-03  0.7005   1.8294   -0.6011  -0.7096
2019-02-04  -0.7339  0.9449   -1.6856  1.1347
2019-02-05  2.4355   -0.4747  0.2161   0.0259
2019-02-06  -1.3965  -0.006   -0.601   -0.7635
2019-02-07  0.6663   1.3514   1.027    1.1583
2019-02-08  0.2744   0.3394   -0.6233  -0.286
2019-02-09  0.1263   0.2296   0.6694   0.5291

julia> apply(X, 1, fun=sum)
100x1 TS{Float64,Dates.Date}: 2018-11-02 to 2019-02-09

Index       Sum
2018-11-02  0.7286
2018-11-03  0.1913
2018-11-04  -0.1492
2018-11-05  -3.2441
2018-11-06  0.195
2018-11-07  0.7526
2018-11-08  2.7696
⋮
2019-02-02  0.0843
2019-02-03  1.2192
2019-02-04  -0.34
2019-02-05  2.2028
2019-02-06  -2.7671
2019-02-07  4.203
2019-02-08  -0.2956
2019-02-09  1.5544

julia> apply(X, 2, fun=sum)
1×4 LinearAlgebra.Adjoint{Float64,Array{Float64,1}}:
 7.87496  19.352  -1.33885  4.45394