Aggregation

Aggregation

Sampling

Weekdays

Temporal.mondays — Function.

mondays(x::TS)

Return a time series containing all observations occuring on Mondays of the given TS object.

source

Temporal.tuesdays — Function.

tuesdays(x::TS)

Return a time series containing all observations occuring on Tuesdays of the given TS object.

source

Temporal.wednesdays — Function.

wednesdays(x::TS)

Return a time series containing all observations occuring on Wednesdays of the given TS object.

source

Temporal.thursdays — Function.

thursdays(x::TS)

Return a time series containing all observations occuring on Thursdays of the given TS object.

source

Temporal.fridays — Function.

fridays(x::TS)

Return a time series containing all observations occuring on Fridays of the given TS object.

source

Temporal.saturdays — Function.

saturdays(x::TS)

Return a time series containing all observations occuring on Saturdays of the given TS object.

source

Temporal.sundays — Function.

sundays(x::TS)

Return a time series containing all observations occuring on Sundays of the given TS object.

source

Interval Boundaries

Weekly

Temporal.bow — Function.

bow(x::TS; cal::Bool=false)

Return a time series containing all observations occuring at the beginnings of the weeks of the input.

If cal is false, only observations occurring the last calendar day of the week are returned

If cal is true, all observation for which the previous index is a prior week are returned

source

Temporal.eow — Function.

eow(x::TS; cal::Bool=false)

Return a time series containing all observations occuring at the ends of the weeks of the input.

If cal is false, only observations occurring the last calendar day of the week are returned

If cal is true, all observation for which the next index is a new week are returned

source

Monthly

Temporal.bom — Function.

bom(x::TS; cal::Bool=false)

Return a time series containing all observations occuring at the beginnings of the months of the input.

If cal is false, only observations occurring the last calendar day of the month are returned

If cal is true, all observation for which the previous index is a prior month are returned

source

Temporal.eom — Function.

eom(x::TS; cal::Bool=false)

Return a time series containing all observations occuring at the ends of the months of the input.

If cal is false, only observations occurring the last calendar day of the month are returned

If cal is true, all observation for which the next index is a new month are returned

source

Quarterly

Temporal.boq — Function.

boq(x::TS; cal::Bool=false)

Return a time series containing all observations occuring at the beginnings of the quarters of the input.

If cal is false, only observations occurring the last calendar day of the quarter are returned

If cal is true, all observation for which the previous index is a prior quarter are returned

source

Temporal.eoq — Function.

eoq(x::TS; cal::Bool=false)

Return a time series containing all observations occuring at the ends of the quarters of the input.

If cal is false, only observations occurring the last calendar day of the quarter are returned

If cal is true, all observation for which the next index is a new quarter are returned

source

Yearly

Temporal.boy — Function.

boy(x::TS; cal::Bool=false)

Return a time series containing all observations occuring at the beginnings of the years of the input.

If cal is false, only observations occurring the last calendar day of the year are returned

If cal is true, all observation for which the previous index is a prior year are returned

source

Temporal.eoy — Function.

eoy(x::TS; cal::Bool=false)

Return a time series containing all observations occuring at the ends of the years of the input.

If cal is false, only observations occurring the last calendar day of the year are returned

If cal is true, all observation for which the next index is a new year are returned

source

Collapsing

Temporal.collapse — Function.

collapse(x::TS{V,T}, at::Function; fun::Function=last, args...)::TS{V,T} where {V,T}

Compute the function fun over period boundaries defined by at (e.g. eom, boq, etc.) and returng a time series of the results.

Keyword arguments (args...) are passed to the function fun.

source

julia> using Temporal, Statistics, Dates

julia> X = TS(randn(100, 4))
100x4 TS{Float64,Dates.Date}: 2018-11-01 to 2019-02-08

Index       A        B        C        D
2018-11-01  -1.0165  1.022    -0.5231  0.1878
2018-11-02  0.8601   -0.3492  -1.3081  -0.4736
2018-11-03  -0.7302  -1.6041  -0.7334  -0.8251
2018-11-04  0.8528   0.6133   0.5812   1.9763
2018-11-05  0.3123   -0.7661  -0.164   -0.2995
2018-11-06  0.6162   0.563    0.3435   0.5009
2018-11-07  -0.1897  -1.271   0.8134   -0.0152
⋮
2019-02-01  -0.858   0.4933   -1.0187  -0.2305
2019-02-02  -1.7413  0.2446   -2.1624  -0.5032
2019-02-03  1.4519   -0.3756  -0.6857  0.5931
2019-02-04  -0.0409  -0.3563  0.227    0.4203
2019-02-05  -0.0723  -0.6659  -0.9097  -1.4016
2019-02-06  -0.7046  -0.4716  -0.2605  0.8197
2019-02-07  -0.2249  -0.7938  0.6176   -1.0686
2019-02-08  -3.0523  -0.3526  -2.3604  0.7451

julia> collapse(X, eom, fun=mean)
3x4 TS{Float64,Dates.Date}: 2018-11-30 to 2019-01-31

Index       A        B        C        D
2018-11-30  -0.0526  -0.1693  -0.2415  -0.0288
2018-12-31  -0.1279  0.1272   0.0646   0.0208
2019-01-31  -0.0251  0.0896   -0.0356  0.1701

julia> last_month = string(X.index[end])[1:7]
"2019-02"

julia> mean(X[last_month])
-0.4593372548053258

Summarizing

Temporal.apply — Function.

apply(x::TS{V}, dim::Int=1; fun=sum) where {V}

Apply function fun across dimension dim of a time series x.

If dim is 1, then apply fun to each row of x, returning a time series (TS) object of the results

If dim is 2, then apply fun to each column of x, returning an Array of the results

source

julia> using Temporal, Statistics

julia> X = TS(randn(100, 4))
100x4 TS{Float64,Dates.Date}: 2018-11-01 to 2019-02-08

Index       A        B        C        D
2018-11-01  -0.4148  0.2286   -0.9618  -2.1965
2018-11-02  0.5592   -1.0212  -0.4131  0.7788
2018-11-03  3.1369   1.6215   -0.1402  -1.4501
2018-11-04  -0.6725  0.5239   1.9993   -0.2349
2018-11-05  -0.6293  0.1108   0.1658   -0.0202
2018-11-06  -1.3459  -1.0367  -0.3215  0.3022
2018-11-07  0.5058   0.5209   -0.7147  -0.5999
⋮
2019-02-01  0.3604   -0.8424  0.612    0.5587
2019-02-02  0.3861   0.2534   -0.3557  -0.3298
2019-02-03  1.0354   -0.1126  0.0542   -0.2634
2019-02-04  1.0662   -0.0358  -1.752   -1.0506
2019-02-05  -0.96    1.1873   -1.829   0.4279
2019-02-06  0.0206   -0.3375  0.9504   0.3261
2019-02-07  -0.1446  -0.3896  -0.0656  0.2435
2019-02-08  0.4467   -0.5929  0.7391   0.0874

julia> apply(X, 1, fun=sum)
100x1 TS{Float64,Dates.Date}: 2018-11-01 to 2019-02-08

Index       Sum
2018-11-01  -3.3445
2018-11-02  -0.0962
2018-11-03  3.1681
2018-11-04  1.6159
2018-11-05  -0.373
2018-11-06  -2.4019
2018-11-07  -0.2879
⋮
2019-02-01  0.6888
2019-02-02  -0.0461
2019-02-03  0.7136
2019-02-04  -1.7722
2019-02-05  -1.1738
2019-02-06  0.9597
2019-02-07  -0.3563
2019-02-08  0.6802

julia> apply(X, 2, fun=sum)
1×4 LinearAlgebra.Adjoint{Float64,Array{Float64,1}}:
 5.12261  9.31271  -12.0825  -6.43123