Combining

Joins

Outer Joins

One can perform a full outer join on the time indexes of two TS objects $x$ and $y$ in the following ways:

merge(x, y)

ojoin(x, y)

[x y]

hcat(x, y)

Where there are dates in the index of one that do not exist in the other, values will be filled with NaN objects. As the missing functionality matures in Julia's base syntax, it will eventually replace NaN in this context, since unfortunately NaN is only applicable for Float64 element types.

Base.merge — Function.

merge(x::TS,y::TS;join::Char='o')::TS

Merge two time series objects together by index with an optionally specified join type parameter.

...

Arguments

x::TS: Left side of the join.

y::TS: Right side of the join.

Optional args:

join::Char='o'::TS: Specifies the logic used to perform the merge, and may take on the values 'o' (outer join), 'i' (inner join), 'l' (left join), or 'r' (right join). Defaults to outer join, whose result is the same as hcat(x, y) or [x y].

...

source

Temporal.ojoin — Function.

ojoin(x::TS,y::TS)::TS

Outer join two TS objects by index.

Equivalent to x OUTER JOIN y ON x.index = y.index.

...

Arguments

x::TS: Left side of the join.

y::TS: Right side of the join.

...

source

Example

julia> x = TS(rand(252))
252x1 TS{Float64,Dates.Date}: 2018-06-28 to 2019-03-06

Index       A
2018-06-28  0.0285
2018-06-29  0.8357
2018-06-30  0.5288
2018-07-01  0.0663
2018-07-02  0.2481
2018-07-03  0.6471
2018-07-04  0.7264
⋮
2019-02-27  0.738
2019-02-28  0.7158
2019-03-01  0.0261
2019-03-02  0.4518
2019-03-03  0.9489
2019-03-04  0.1504
2019-03-05  0.0426
2019-03-06  0.5213

julia> y = TS(rand(252), x.index .- Month(6))
252x1 TS{Float64,Dates.Date}: 2017-12-28 to 2018-09-06

Index       A
2017-12-28  0.0428
2017-12-29  0.2207
2017-12-30  0.7375
2018-01-01  0.2433
2018-01-02  0.8534
2018-01-03  0.3097
2018-01-04  0.7552
⋮
2018-08-27  0.2992
2018-08-28  0.3082
2018-09-01  0.1871
2018-09-02  0.1911
2018-09-03  0.3117
2018-09-04  0.0956
2018-09-05  0.29
2018-09-06  0.8438

julia> [x y]
431x2 TS{Float64,Dates.Date}: 2017-12-28 to 2019-03-06

Index       A       A
2017-12-28  NaN     0.0428
2017-12-29  NaN     0.2207
2017-12-30  NaN     0.7375
2018-01-01  NaN     0.2433
2018-01-02  NaN     0.8534
2018-01-03  NaN     0.3097
2018-01-04  NaN     0.7552
⋮
2019-02-27  0.738   NaN
2019-02-28  0.7158  NaN
2019-03-01  0.0261  NaN
2019-03-02  0.4518  NaN
2019-03-03  0.9489  NaN
2019-03-04  0.1504  NaN
2019-03-05  0.0426  NaN
2019-03-06  0.5213  NaN

Inner Joins

You can do inner joins on TS objects using the ijoin function, which will remove any observations corresponding to time steps where at least one of the joined objects is missing a row. This will basically keep only the rows where the time index of the left side and the right side intersect.

Temporal.ijoin — Function.

ijoin(x::TS,y::TS)::TS

Inner join two TS objects by index.

Equivalent to x INNER JOIN y on x.index = y.index.

...

Arguments

x::TS: Left side of the join.

y::TS: Right side of the join.

...

source

Example

julia> x = TS(rand(252))
252x1 TS{Float64,Dates.Date}: 2018-06-28 to 2019-03-06

Index       A
2018-06-28  0.8692
2018-06-29  0.6486
2018-06-30  0.8929
2018-07-01  0.2427
2018-07-02  0.814
2018-07-03  0.3022
2018-07-04  0.3149
⋮
2019-02-27  0.135
2019-02-28  0.776
2019-03-01  0.5377
2019-03-02  0.0262
2019-03-03  0.02
2019-03-04  0.3667
2019-03-05  0.6966
2019-03-06  0.3845

julia> y = TS(rand(252), x.index .- Month(6))
252x1 TS{Float64,Dates.Date}: 2017-12-28 to 2018-09-06

Index       A
2017-12-28  0.4178
2017-12-29  0.5072
2017-12-30  0.1453
2018-01-01  0.2263
2018-01-02  0.1227
2018-01-03  0.0197
2018-01-04  0.3397
⋮
2018-08-27  0.4369
2018-08-28  0.0292
2018-09-01  0.5308
2018-09-02  0.8843
2018-09-03  0.6871
2018-09-04  0.2423
2018-09-05  0.1565
2018-09-06  0.2517

julia> ijoin(x, y)
ERROR: ArgumentError: number of rows of each array must match (got (68, 69))

Left/Right Joins

Left and right joins are performed similarly to inner joins and the typical SQL join queries using the index field each object as the joining key.

Left Join: keep all observations of the left side of the join, fill the right side with NaN's where missing the corresponding time index

Right Join: keep all observations of the right side of the join, fill the left side with NaN's where missing the corresponding time index

Temporal.ljoin — Function.

ljoin(x::TS, y::TS)::TS

Left join two TS objects by index.

Equivalent to x LEFT JOIN y ON x.index = y.index.

...

Arguments

x::TS: Left side of the join.

y::TS: Right side of the join.

...

source

Temporal.rjoin — Function.

rjoin(x::TS, y::TS)::TS

Right join two TS objects by index.

Equivalent to x RIGHT JOIN y ON x.index = y.index.

...

Arguments

x::TS: Left side of the join.

y::TS: Right side of the join.

...

source

Example

julia> x = TS(rand(252))
252x1 TS{Float64,Dates.Date}: 2018-06-28 to 2019-03-06

Index       A
2018-06-28  0.3472
2018-06-29  0.885
2018-06-30  0.1026
2018-07-01  0.4163
2018-07-02  0.4041
2018-07-03  0.8674
2018-07-04  0.0196
⋮
2019-02-27  0.1076
2019-02-28  0.6533
2019-03-01  0.603
2019-03-02  0.8294
2019-03-03  0.8086
2019-03-04  0.1958
2019-03-05  0.1796
2019-03-06  0.7861

julia> y = TS(rand(252), x.index .- Month(6))
252x1 TS{Float64,Dates.Date}: 2017-12-28 to 2018-09-06

Index       A
2017-12-28  0.5267
2017-12-29  0.7289
2017-12-30  0.2739
2018-01-01  0.3401
2018-01-02  0.9526
2018-01-03  0.4583
2018-01-04  0.5393
⋮
2018-08-27  0.781
2018-08-28  0.4747
2018-09-01  0.0519
2018-09-02  0.3595
2018-09-03  0.5636
2018-09-04  0.4064
2018-09-05  0.629
2018-09-06  0.9698

julia> ljoin(x, y)
252x2 TS{Float64,Dates.Date}: 2018-06-28 to 2019-03-06

Index       A       A
2018-06-28  0.3472  0.6111
2018-06-29  0.885   0.4548
2018-06-30  0.1026  0.9923
2018-07-01  0.4163  0.3381
2018-07-02  0.4041  0.3335
2018-07-03  0.8674  0.7436
2018-07-04  0.0196  0.8984
⋮
2019-02-27  0.1076  NaN
2019-02-28  0.6533  NaN
2019-03-01  0.603   NaN
2019-03-02  0.8294  NaN
2019-03-03  0.8086  NaN
2019-03-04  0.1958  NaN
2019-03-05  0.1796  NaN
2019-03-06  0.7861  NaN

julia> rjoin(x, y)
247x2 TS{Float64,Dates.Date}: 2017-12-28 to 2018-09-06

Index       A       A
2017-12-28  NaN     0.5267
2017-12-29  NaN     0.7289
2017-12-30  NaN     0.2739
2018-01-01  NaN     0.3401
2018-01-02  NaN     0.9526
2018-01-03  NaN     0.4583
2018-01-04  NaN     0.5393
⋮
2018-08-27  0.7277  0.8947
2018-08-28  0.5022  0.0293
2018-09-01  0.2582  0.3151
2018-09-02  0.7633  0.0996
2018-09-03  0.5602  0.6144
2018-09-04  0.0138  0.781
2018-09-05  0.1053  0.4747
2018-09-06  0.016   0.0519