Combining

Joins

Outer Joins

One can perform a full outer join on the time indexes of two TS objects $x$ and $y$ in the following ways:

merge(x, y)

ojoin(x, y)

[x y]

hcat(x, y)

Where there are dates in the index of one that do not exist in the other, values will be filled with NaN objects. As the missing functionality matures in Julia's base syntax, it will eventually replace NaN in this context, since unfortunately NaN is only applicable for Float64 element types.

Base.merge — Function.

merge(x::TS,y::TS;join::Char='o')::TS

Merge two time series objects together by index with an optionally specified join type parameter.

...

Arguments

x::TS: Left side of the join.

y::TS: Right side of the join.

Optional args:

join::Char='o'::TS: Specifies the logic used to perform the merge, and may take on the values 'o' (outer join), 'i' (inner join), 'l' (left join), or 'r' (right join). Defaults to outer join, whose result is the same as hcat(x, y) or [x y].

...

source

Temporal.ojoin — Function.

ojoin(x::TS,y::TS)::TS

Outer join two TS objects by index.

Equivalent to x OUTER JOIN y ON x.index = y.index.

...

Arguments

x::TS: Left side of the join.

y::TS: Right side of the join.

...

source

Example

julia> x = TS(rand(252))
252x1 TS{Float64,Dates.Date}: 2018-07-10 to 2019-03-18

Index       A
2018-07-10  0.3675
2018-07-11  0.7024
2018-07-12  0.721
2018-07-13  0.027
2018-07-14  0.7736
2018-07-15  0.0802
2018-07-16  0.1775
⋮
2019-03-11  0.8018
2019-03-12  0.0049
2019-03-13  0.2569
2019-03-14  0.3557
2019-03-15  0.1204
2019-03-16  0.912
2019-03-17  0.099
2019-03-18  0.5588

julia> y = TS(rand(252), x.index .- Month(6))
252x1 TS{Float64,Dates.Date}: 2018-01-10 to 2018-09-18

Index       A
2018-01-10  0.1101
2018-01-11  0.8657
2018-01-12  0.6509
2018-01-13  0.6505
2018-01-14  0.5666
2018-01-15  0.4002
2018-01-16  0.9948
⋮
2018-09-11  0.4497
2018-09-12  0.7276
2018-09-13  0.8818
2018-09-14  0.617
2018-09-15  0.1436
2018-09-16  0.7725
2018-09-17  0.9823
2018-09-18  0.7645

julia> [x y]
431x2 TS{Float64,Dates.Date}: 2018-01-10 to 2019-03-18

Index       A       A
2018-01-10  NaN     0.1101
2018-01-11  NaN     0.8657
2018-01-12  NaN     0.6509
2018-01-13  NaN     0.6505
2018-01-14  NaN     0.5666
2018-01-15  NaN     0.4002
2018-01-16  NaN     0.9948
⋮
2019-03-11  0.8018  NaN
2019-03-12  0.0049  NaN
2019-03-13  0.2569  NaN
2019-03-14  0.3557  NaN
2019-03-15  0.1204  NaN
2019-03-16  0.912   NaN
2019-03-17  0.099   NaN
2019-03-18  0.5588  NaN

Inner Joins

You can do inner joins on TS objects using the ijoin function, which will remove any observations corresponding to time steps where at least one of the joined objects is missing a row. This will basically keep only the rows where the time index of the left side and the right side intersect.

Temporal.ijoin — Function.

ijoin(x::TS,y::TS)::TS

Inner join two TS objects by index.

Equivalent to x INNER JOIN y on x.index = y.index.

...

Arguments

x::TS: Left side of the join.

y::TS: Right side of the join.

...

source

Example

julia> x = TS(rand(252))
252x1 TS{Float64,Dates.Date}: 2018-07-10 to 2019-03-18

Index       A
2018-07-10  0.2808
2018-07-11  0.7197
2018-07-12  0.039
2018-07-13  0.6484
2018-07-14  0.1664
2018-07-15  0.9607
2018-07-16  0.6072
⋮
2019-03-11  0.9366
2019-03-12  0.0655
2019-03-13  0.839
2019-03-14  0.245
2019-03-15  0.3648
2019-03-16  0.0758
2019-03-17  0.7685
2019-03-18  0.6144

julia> y = TS(rand(252), x.index .- Month(6))
252x1 TS{Float64,Dates.Date}: 2018-01-10 to 2018-09-18

Index       A
2018-01-10  0.8515
2018-01-11  0.3752
2018-01-12  0.2554
2018-01-13  0.3984
2018-01-14  0.5772
2018-01-15  0.2383
2018-01-16  0.3797
⋮
2018-09-11  0.9479
2018-09-12  0.5261
2018-09-13  0.0505
2018-09-14  0.6402
2018-09-15  0.5819
2018-09-16  0.9917
2018-09-17  0.6703
2018-09-18  0.9267

julia> ijoin(x, y)
68x2 TS{Float64,Dates.Date}: 2018-07-10 to 2018-09-18

Index       A       A
2018-07-10  0.2808  0.99
2018-07-11  0.7197  0.5131
2018-07-12  0.039   0.5433
2018-07-13  0.6484  0.0819
2018-07-14  0.1664  0.3609
2018-07-15  0.9607  0.3833
2018-07-16  0.6072  0.0947
⋮
2018-09-11  0.0441  0.9479
2018-09-12  0.6174  0.5261
2018-09-13  0.2135  0.0505
2018-09-14  0.1156  0.6402
2018-09-15  0.9793  0.5819
2018-09-16  0.6228  0.9917
2018-09-17  0.9606  0.6703
2018-09-18  0.5489  0.9267

Left/Right Joins

Left and right joins are performed similarly to inner joins and the typical SQL join queries using the index field each object as the joining key.

Left Join: keep all observations of the left side of the join, fill the right side with NaN's where missing the corresponding time index

Right Join: keep all observations of the right side of the join, fill the left side with NaN's where missing the corresponding time index

Temporal.ljoin — Function.

ljoin(x::TS, y::TS)::TS

Left join two TS objects by index.

Equivalent to x LEFT JOIN y ON x.index = y.index.

...

Arguments

x::TS: Left side of the join.

y::TS: Right side of the join.

...

source

Temporal.rjoin — Function.

rjoin(x::TS, y::TS)::TS

Right join two TS objects by index.

Equivalent to x RIGHT JOIN y ON x.index = y.index.

...

Arguments

x::TS: Left side of the join.

y::TS: Right side of the join.

...

source

Example

julia> x = TS(rand(252))
252x1 TS{Float64,Dates.Date}: 2018-07-10 to 2019-03-18

Index       A
2018-07-10  0.3969
2018-07-11  0.5743
2018-07-12  0.7367
2018-07-13  0.4139
2018-07-14  0.6446
2018-07-15  0.7493
2018-07-16  0.4045
⋮
2019-03-11  0.8509
2019-03-12  0.6834
2019-03-13  0.0357
2019-03-14  0.3383
2019-03-15  0.1362
2019-03-16  0.6111
2019-03-17  0.5581
2019-03-18  0.6471

julia> y = TS(rand(252), x.index .- Month(6))
252x1 TS{Float64,Dates.Date}: 2018-01-10 to 2018-09-18

Index       A
2018-01-10  0.1665
2018-01-11  0.0231
2018-01-12  0.6026
2018-01-13  0.1693
2018-01-14  0.9653
2018-01-15  0.6987
2018-01-16  0.2042
⋮
2018-09-11  0.274
2018-09-12  0.1166
2018-09-13  0.4035
2018-09-14  0.7894
2018-09-15  0.9959
2018-09-16  0.9972
2018-09-17  0.5895
2018-09-18  0.164

julia> ljoin(x, y)
252x2 TS{Float64,Dates.Date}: 2018-07-10 to 2019-03-18

Index       A       A
2018-07-10  0.3969  0.6258
2018-07-11  0.5743  0.4161
2018-07-12  0.7367  0.9817
2018-07-13  0.4139  0.4028
2018-07-14  0.6446  0.0505
2018-07-15  0.7493  0.955
2018-07-16  0.4045  0.0752
⋮
2019-03-11  0.8509  NaN
2019-03-12  0.6834  NaN
2019-03-13  0.0357  NaN
2019-03-14  0.3383  NaN
2019-03-15  0.1362  NaN
2019-03-16  0.6111  NaN
2019-03-17  0.5581  NaN
2019-03-18  0.6471  NaN

julia> rjoin(x, y)
247x2 TS{Float64,Dates.Date}: 2018-01-10 to 2018-09-18

Index       A       A
2018-01-10  NaN     0.1665
2018-01-11  NaN     0.0231
2018-01-12  NaN     0.6026
2018-01-13  NaN     0.1693
2018-01-14  NaN     0.9653
2018-01-15  NaN     0.6987
2018-01-16  NaN     0.2042
⋮
2018-09-11  0.4526  0.8579
2018-09-12  0.6014  0.4841
2018-09-13  0.2657  0.5286
2018-09-14  0.0582  0.521
2018-09-15  0.6949  0.5944
2018-09-16  0.1377  0.274
2018-09-17  0.7822  0.1166
2018-09-18  0.2721  0.4035