Joins

Joins

Outer Joins

One can perform a full outer join on the time indexes of two TS objects $x$ and $y$ in the following ways:

merge(x, y)

ojoin(x, y)

[x y]

hcat(x, y)

Where there are dates in the index of one that do not exist in the other, values will be filled with NaN objects. As the missing functionality matures in Julia's base syntax, it will eventually replace NaN in this context, since unfortunately NaN is only applicable for Float64 element types.

Base.merge — Function.

merge(x::TS,y::TS;join::Char='o')::TS

Merge two time series objects together by index with an optionally specified join type parameter.

...

Arguments

x::TS: Left side of the join.

y::TS: Right side of the join.

Optional args:

join::Char='o'::TS: Specifies the logic used to perform the merge, and may take on the values 'o' (outer join), 'i' (inner join), 'l' (left join), or 'r' (right join). Defaults to outer join, whose result is the same as hcat(x, y) or [x y].

...

source

Temporal.ojoin — Function.

ojoin(x::TS,y::TS)::TS

Outer join two TS objects by index.

Equivalent to x OUTER JOIN y ON x.index = y.index.

...

Arguments

x::TS: Left side of the join.

y::TS: Right side of the join.

...

source

Example

julia> x = TS(rand(252))
252x1 TS{Float64,Dates.Date}: 2018-06-07 to 2019-02-13

Index       A
2018-06-07  0.4315
2018-06-08  0.0574
2018-06-09  0.6588
2018-06-10  0.6434
2018-06-11  0.4174
2018-06-12  0.4376
2018-06-13  0.0597
⋮
2019-02-06  0.3692
2019-02-07  0.8531
2019-02-08  0.9396
2019-02-09  0.4369
2019-02-10  0.0522
2019-02-11  0.2157
2019-02-12  0.643
2019-02-13  0.4576

julia> y = TS(rand(252), x.index .- Month(6))
252x1 TS{Float64,Dates.Date}: 2017-12-07 to 2018-08-13

Index       A
2017-12-07  0.2136
2017-12-08  0.0814
2017-12-09  0.9142
2017-12-10  0.0831
2017-12-11  0.5431
2017-12-12  0.4403
2017-12-13  0.2003
⋮
2018-08-06  0.061
2018-08-07  0.6385
2018-08-08  0.0717
2018-08-09  0.7429
2018-08-10  0.046
2018-08-11  0.3092
2018-08-12  0.3796
2018-08-13  0.1765

julia> [x y]
431x2 TS{Float64,Dates.Date}: 2017-12-07 to 2019-02-13

Index       A       A
2017-12-07  NaN     0.2136
2017-12-08  NaN     0.0814
2017-12-09  NaN     0.9142
2017-12-10  NaN     0.0831
2017-12-11  NaN     0.5431
2017-12-12  NaN     0.4403
2017-12-13  NaN     0.2003
⋮
2019-02-06  0.3692  NaN
2019-02-07  0.8531  NaN
2019-02-08  0.9396  NaN
2019-02-09  0.4369  NaN
2019-02-10  0.0522  NaN
2019-02-11  0.2157  NaN
2019-02-12  0.643   NaN
2019-02-13  0.4576  NaN

Inner Joins

You can do inner joins on TS objects using the ijoin function, which will remove any observations corresponding to time steps where at least one of the joined objects is missing a row. This will basically keep only the rows where the time index of the left side and the right side intersect.

Temporal.ijoin — Function.

ijoin(x::TS,y::TS)::TS

Inner join two TS objects by index.

Equivalent to x INNER JOIN y on x.index = y.index.

...

Arguments

x::TS: Left side of the join.

y::TS: Right side of the join.

...

source

Example

julia> x = TS(rand(252))
252x1 TS{Float64,Dates.Date}: 2018-06-07 to 2019-02-13

Index       A
2018-06-07  0.017
2018-06-08  0.8486
2018-06-09  0.7905
2018-06-10  0.1288
2018-06-11  0.5746
2018-06-12  0.7391
2018-06-13  0.0378
⋮
2019-02-06  0.8237
2019-02-07  0.5866
2019-02-08  0.0488
2019-02-09  0.7453
2019-02-10  0.1212
2019-02-11  0.6714
2019-02-12  0.9414
2019-02-13  0.0423

julia> y = TS(rand(252), x.index .- Month(6))
252x1 TS{Float64,Dates.Date}: 2017-12-07 to 2018-08-13

Index       A
2017-12-07  0.5281
2017-12-08  0.8389
2017-12-09  0.9844
2017-12-10  0.8229
2017-12-11  0.5362
2017-12-12  0.1661
2017-12-13  0.1024
⋮
2018-08-06  0.1596
2018-08-07  0.9073
2018-08-08  0.0068
2018-08-09  0.0194
2018-08-10  0.1144
2018-08-11  0.5185
2018-08-12  0.3767
2018-08-13  0.8051

julia> ijoin(x, y)
ERROR: ArgumentError: number of rows of each array must match (got (68, 69))

Left/Right Joins

Left and right joins are performed similarly to inner joins and the typical SQL join queries using the index field each object as the joining key.

Left Join: keep all observations of the left side of the join, fill the right side with NaN's where missing the corresponding time index

Right Join: keep all observations of the right side of the join, fill the left side with NaN's where missing the corresponding time index

Temporal.ljoin — Function.

ljoin(x::TS, y::TS)::TS

Left join two TS objects by index.

Equivalent to x LEFT JOIN y ON x.index = y.index.

...

Arguments

x::TS: Left side of the join.

y::TS: Right side of the join.

...

source

Temporal.rjoin — Function.

rjoin(x::TS, y::TS)::TS

Right join two TS objects by index.

Equivalent to x RIGHT JOIN y ON x.index = y.index.

...

Arguments

x::TS: Left side of the join.

y::TS: Right side of the join.

...

source

Example

julia> x = TS(rand(252))
252x1 TS{Float64,Dates.Date}: 2018-06-07 to 2019-02-13

Index       A
2018-06-07  0.7894
2018-06-08  0.3156
2018-06-09  0.1223
2018-06-10  0.1144
2018-06-11  0.1875
2018-06-12  0.7093
2018-06-13  0.8874
⋮
2019-02-06  0.3151
2019-02-07  0.1448
2019-02-08  0.211
2019-02-09  0.4227
2019-02-10  0.7381
2019-02-11  0.8669
2019-02-12  0.1337
2019-02-13  0.6035

julia> y = TS(rand(252), x.index .- Month(6))
252x1 TS{Float64,Dates.Date}: 2017-12-07 to 2018-08-13

Index       A
2017-12-07  0.4741
2017-12-08  0.1987
2017-12-09  0.1753
2017-12-10  0.5006
2017-12-11  0.9868
2017-12-12  0.242
2017-12-13  0.5103
⋮
2018-08-06  0.2557
2018-08-07  0.4044
2018-08-08  0.591
2018-08-09  0.1746
2018-08-10  0.75
2018-08-11  0.1798
2018-08-12  0.8354
2018-08-13  0.3146

julia> ljoin(x, y)
252x2 TS{Float64,Dates.Date}: 2018-06-07 to 2019-02-13

Index       A       A
2018-06-07  0.7894  0.3351
2018-06-08  0.3156  0.908
2018-06-09  0.1223  0.1188
2018-06-10  0.1144  0.7358
2018-06-11  0.1875  0.0706
2018-06-12  0.7093  0.5805
2018-06-13  0.8874  0.4618
⋮
2019-02-06  0.3151  NaN
2019-02-07  0.1448  NaN
2019-02-08  0.211   NaN
2019-02-09  0.4227  NaN
2019-02-10  0.7381  NaN
2019-02-11  0.8669  NaN
2019-02-12  0.1337  NaN
2019-02-13  0.6035  NaN

julia> rjoin(x, y)
247x2 TS{Float64,Dates.Date}: 2017-12-07 to 2018-08-13

Index       A       A
2017-12-07  NaN     0.4741
2017-12-08  NaN     0.1987
2017-12-09  NaN     0.1753
2017-12-10  NaN     0.5006
2017-12-11  NaN     0.9868
2017-12-12  NaN     0.242
2017-12-13  NaN     0.5103
⋮
2018-08-06  0.4848  0.781
2018-08-07  0.4411  0.9848
2018-08-08  0.3591  0.0272
2018-08-09  0.0621  0.9151
2018-08-10  0.4601  0.8972
2018-08-11  0.2925  0.2557
2018-08-12  0.1113  0.4044
2018-08-13  0.6767  0.591