Combining

Joins

Outer Joins

One can perform a full outer join on the time indexes of two TS objects $x$ and $y$ in the following ways:

merge(x, y)

ojoin(x, y)

[x y]

hcat(x, y)

Where there are dates in the index of one that do not exist in the other, values will be filled with NaN objects. As the missing functionality matures in Julia's base syntax, it will eventually replace NaN in this context, since unfortunately NaN is only applicable for Float64 element types.

Base.merge — Function.

merge(x::TS,y::TS;join::Char='o')::TS

Merge two time series objects together by index with an optionally specified join type parameter.

...

Arguments

x::TS: Left side of the join.

y::TS: Right side of the join.

Optional args:

join::Char='o'::TS: Specifies the logic used to perform the merge, and may take on the values 'o' (outer join), 'i' (inner join), 'l' (left join), or 'r' (right join). Defaults to outer join, whose result is the same as hcat(x, y) or [x y].

...

source

Temporal.ojoin — Function.

ojoin(x::TS,y::TS)::TS

Outer join two TS objects by index.

Equivalent to x OUTER JOIN y ON x.index = y.index.

...

Arguments

x::TS: Left side of the join.

y::TS: Right side of the join.

...

source

Example

julia> x = TS(rand(252))
252x1 TS{Float64,Dates.Date}: 2018-06-13 to 2019-02-19

Index       A
2018-06-13  0.6427
2018-06-14  0.3718
2018-06-15  0.0603
2018-06-16  0.2524
2018-06-17  0.3154
2018-06-18  0.9633
2018-06-19  0.7154
⋮
2019-02-12  0.1524
2019-02-13  0.9263
2019-02-14  0.9767
2019-02-15  0.9745
2019-02-16  0.2796
2019-02-17  0.2554
2019-02-18  0.0906
2019-02-19  0.8039

julia> y = TS(rand(252), x.index .- Month(6))
252x1 TS{Float64,Dates.Date}: 2017-12-13 to 2018-08-19

Index       A
2017-12-13  0.7077
2017-12-14  0.0405
2017-12-15  0.2341
2017-12-16  0.0792
2017-12-17  0.4612
2017-12-18  0.5214
2017-12-19  0.784
⋮
2018-08-12  0.955
2018-08-13  0.2347
2018-08-14  0.2202
2018-08-15  0.6393
2018-08-16  0.4245
2018-08-17  0.8622
2018-08-18  0.0755
2018-08-19  0.308

julia> [x y]
431x2 TS{Float64,Dates.Date}: 2017-12-13 to 2019-02-19

Index       A       A
2017-12-13  NaN     0.7077
2017-12-14  NaN     0.0405
2017-12-15  NaN     0.2341
2017-12-16  NaN     0.0792
2017-12-17  NaN     0.4612
2017-12-18  NaN     0.5214
2017-12-19  NaN     0.784
⋮
2019-02-12  0.1524  NaN
2019-02-13  0.9263  NaN
2019-02-14  0.9767  NaN
2019-02-15  0.9745  NaN
2019-02-16  0.2796  NaN
2019-02-17  0.2554  NaN
2019-02-18  0.0906  NaN
2019-02-19  0.8039  NaN

Inner Joins

You can do inner joins on TS objects using the ijoin function, which will remove any observations corresponding to time steps where at least one of the joined objects is missing a row. This will basically keep only the rows where the time index of the left side and the right side intersect.

Temporal.ijoin — Function.

ijoin(x::TS,y::TS)::TS

Inner join two TS objects by index.

Equivalent to x INNER JOIN y on x.index = y.index.

...

Arguments

x::TS: Left side of the join.

y::TS: Right side of the join.

...

source

Example

julia> x = TS(rand(252))
252x1 TS{Float64,Dates.Date}: 2018-06-13 to 2019-02-19

Index       A
2018-06-13  0.5789
2018-06-14  0.4046
2018-06-15  0.4586
2018-06-16  0.8481
2018-06-17  0.6177
2018-06-18  0.6347
2018-06-19  0.1659
⋮
2019-02-12  0.8018
2019-02-13  0.491
2019-02-14  0.565
2019-02-15  0.5254
2019-02-16  0.5223
2019-02-17  0.3257
2019-02-18  0.519
2019-02-19  0.1448

julia> y = TS(rand(252), x.index .- Month(6))
252x1 TS{Float64,Dates.Date}: 2017-12-13 to 2018-08-19

Index       A
2017-12-13  0.5757
2017-12-14  0.0409
2017-12-15  0.1728
2017-12-16  0.4552
2017-12-17  0.1543
2017-12-18  0.2491
2017-12-19  0.4353
⋮
2018-08-12  0.0538
2018-08-13  0.8063
2018-08-14  0.482
2018-08-15  0.8255
2018-08-16  0.8744
2018-08-17  0.6819
2018-08-18  0.1348
2018-08-19  0.3437

julia> ijoin(x, y)
ERROR: ArgumentError: number of rows of each array must match (got (68, 69))

Left/Right Joins

Left and right joins are performed similarly to inner joins and the typical SQL join queries using the index field each object as the joining key.

Left Join: keep all observations of the left side of the join, fill the right side with NaN's where missing the corresponding time index

Right Join: keep all observations of the right side of the join, fill the left side with NaN's where missing the corresponding time index

Temporal.ljoin — Function.

ljoin(x::TS, y::TS)::TS

Left join two TS objects by index.

Equivalent to x LEFT JOIN y ON x.index = y.index.

...

Arguments

x::TS: Left side of the join.

y::TS: Right side of the join.

...

source

Temporal.rjoin — Function.

rjoin(x::TS, y::TS)::TS

Right join two TS objects by index.

Equivalent to x RIGHT JOIN y ON x.index = y.index.

...

Arguments

x::TS: Left side of the join.

y::TS: Right side of the join.

...

source

Example

julia> x = TS(rand(252))
252x1 TS{Float64,Dates.Date}: 2018-06-13 to 2019-02-19

Index       A
2018-06-13  0.498
2018-06-14  0.5148
2018-06-15  0.0836
2018-06-16  0.3151
2018-06-17  0.4542
2018-06-18  0.1374
2018-06-19  0.7789
⋮
2019-02-12  0.3796
2019-02-13  0.048
2019-02-14  0.793
2019-02-15  0.4529
2019-02-16  0.5453
2019-02-17  0.1
2019-02-18  0.656
2019-02-19  0.1389

julia> y = TS(rand(252), x.index .- Month(6))
252x1 TS{Float64,Dates.Date}: 2017-12-13 to 2018-08-19

Index       A
2017-12-13  0.1776
2017-12-14  0.3402
2017-12-15  0.5305
2017-12-16  0.1081
2017-12-17  0.1725
2017-12-18  0.0206
2017-12-19  0.7945
⋮
2018-08-12  0.2717
2018-08-13  0.9902
2018-08-14  0.2946
2018-08-15  0.1304
2018-08-16  0.6662
2018-08-17  0.4193
2018-08-18  0.5683
2018-08-19  0.1762

julia> ljoin(x, y)
252x2 TS{Float64,Dates.Date}: 2018-06-13 to 2019-02-19

Index       A       A
2018-06-13  0.498   0.9472
2018-06-14  0.5148  0.8049
2018-06-15  0.0836  0.2469
2018-06-16  0.3151  0.1633
2018-06-17  0.4542  0.1978
2018-06-18  0.1374  0.1708
2018-06-19  0.7789  0.7472
⋮
2019-02-12  0.3796  NaN
2019-02-13  0.048   NaN
2019-02-14  0.793   NaN
2019-02-15  0.4529  NaN
2019-02-16  0.5453  NaN
2019-02-17  0.1     NaN
2019-02-18  0.656   NaN
2019-02-19  0.1389  NaN

julia> rjoin(x, y)
247x2 TS{Float64,Dates.Date}: 2017-12-13 to 2018-08-19

Index       A       A
2017-12-13  NaN     0.1776
2017-12-14  NaN     0.3402
2017-12-15  NaN     0.5305
2017-12-16  NaN     0.1081
2017-12-17  NaN     0.1725
2017-12-18  NaN     0.0206
2017-12-19  NaN     0.7945
⋮
2018-08-12  0.3086  0.2929
2018-08-13  0.5037  0.7678
2018-08-14  0.2947  0.637
2018-08-15  0.589   0.7364
2018-08-16  0.3118  0.5823
2018-08-17  0.4115  0.2717
2018-08-18  0.6826  0.9902
2018-08-19  0.9924  0.2946