Combining

Joins

Outer Joins

One can perform a full outer join on the time indexes of two TS objects $x$ and $y$ in the following ways:

merge(x, y)

ojoin(x, y)

[x y]

hcat(x, y)

Where there are dates in the index of one that do not exist in the other, values will be filled with NaN objects. As the missing functionality matures in Julia's base syntax, it will eventually replace NaN in this context, since unfortunately NaN is only applicable for Float64 element types.

Base.merge — Function.

merge(x::TS,y::TS;join::Char='o')::TS

Merge two time series objects together by index with an optionally specified join type parameter.

...

Arguments

x::TS: Left side of the join.

y::TS: Right side of the join.

Optional args:

join::Char='o'::TS: Specifies the logic used to perform the merge, and may take on the values 'o' (outer join), 'i' (inner join), 'l' (left join), or 'r' (right join). Defaults to outer join, whose result is the same as hcat(x, y) or [x y].

...

source

Temporal.ojoin — Function.

ojoin(x::TS,y::TS)::TS

Outer join two TS objects by index.

Equivalent to x OUTER JOIN y ON x.index = y.index.

...

Arguments

x::TS: Left side of the join.

y::TS: Right side of the join.

...

source

Example

julia> x = TS(rand(252))
252x1 TS{Float64,Dates.Date}: 2018-06-07 to 2019-02-13

Index       A
2018-06-07  0.5881
2018-06-08  0.4607
2018-06-09  0.2691
2018-06-10  0.5992
2018-06-11  0.6887
2018-06-12  0.2341
2018-06-13  0.9216
⋮
2019-02-06  0.9986
2019-02-07  0.7811
2019-02-08  0.2516
2019-02-09  0.9807
2019-02-10  0.5711
2019-02-11  0.8737
2019-02-12  0.1
2019-02-13  0.1555

julia> y = TS(rand(252), x.index .- Month(6))
252x1 TS{Float64,Dates.Date}: 2017-12-07 to 2018-08-13

Index       A
2017-12-07  0.9794
2017-12-08  0.8943
2017-12-09  0.1485
2017-12-10  0.9095
2017-12-11  0.6194
2017-12-12  0.3819
2017-12-13  0.5328
⋮
2018-08-06  0.244
2018-08-07  0.3868
2018-08-08  0.1865
2018-08-09  0.4327
2018-08-10  0.434
2018-08-11  0.7735
2018-08-12  0.2865
2018-08-13  0.3237

julia> [x y]
431x2 TS{Float64,Dates.Date}: 2017-12-07 to 2019-02-13

Index       A       A
2017-12-07  NaN     0.9794
2017-12-08  NaN     0.8943
2017-12-09  NaN     0.1485
2017-12-10  NaN     0.9095
2017-12-11  NaN     0.6194
2017-12-12  NaN     0.3819
2017-12-13  NaN     0.5328
⋮
2019-02-06  0.9986  NaN
2019-02-07  0.7811  NaN
2019-02-08  0.2516  NaN
2019-02-09  0.9807  NaN
2019-02-10  0.5711  NaN
2019-02-11  0.8737  NaN
2019-02-12  0.1     NaN
2019-02-13  0.1555  NaN

Inner Joins

You can do inner joins on TS objects using the ijoin function, which will remove any observations corresponding to time steps where at least one of the joined objects is missing a row. This will basically keep only the rows where the time index of the left side and the right side intersect.

Temporal.ijoin — Function.

ijoin(x::TS,y::TS)::TS

Inner join two TS objects by index.

Equivalent to x INNER JOIN y on x.index = y.index.

...

Arguments

x::TS: Left side of the join.

y::TS: Right side of the join.

...

source

Example

julia> x = TS(rand(252))
252x1 TS{Float64,Dates.Date}: 2018-06-07 to 2019-02-13

Index       A
2018-06-07  0.9864
2018-06-08  0.7455
2018-06-09  0.0018
2018-06-10  0.293
2018-06-11  0.0262
2018-06-12  0.1116
2018-06-13  0.1321
⋮
2019-02-06  0.1569
2019-02-07  0.9361
2019-02-08  0.3352
2019-02-09  0.7569
2019-02-10  0.6547
2019-02-11  0.5953
2019-02-12  0.9913
2019-02-13  0.8585

julia> y = TS(rand(252), x.index .- Month(6))
252x1 TS{Float64,Dates.Date}: 2017-12-07 to 2018-08-13

Index       A
2017-12-07  0.3052
2017-12-08  0.0211
2017-12-09  0.9976
2017-12-10  0.1652
2017-12-11  0.0609
2017-12-12  0.5586
2017-12-13  0.7781
⋮
2018-08-06  0.4464
2018-08-07  0.7702
2018-08-08  0.642
2018-08-09  0.1375
2018-08-10  0.6473
2018-08-11  0.3102
2018-08-12  0.1023
2018-08-13  0.4895

julia> ijoin(x, y)
ERROR: ArgumentError: number of rows of each array must match (got (68, 69))

Left/Right Joins

Left and right joins are performed similarly to inner joins and the typical SQL join queries using the index field each object as the joining key.

Left Join: keep all observations of the left side of the join, fill the right side with NaN's where missing the corresponding time index

Right Join: keep all observations of the right side of the join, fill the left side with NaN's where missing the corresponding time index

Temporal.ljoin — Function.

ljoin(x::TS, y::TS)::TS

Left join two TS objects by index.

Equivalent to x LEFT JOIN y ON x.index = y.index.

...

Arguments

x::TS: Left side of the join.

y::TS: Right side of the join.

...

source

Temporal.rjoin — Function.

rjoin(x::TS, y::TS)::TS

Right join two TS objects by index.

Equivalent to x RIGHT JOIN y ON x.index = y.index.

...

Arguments

x::TS: Left side of the join.

y::TS: Right side of the join.

...

source

Example

julia> x = TS(rand(252))
252x1 TS{Float64,Dates.Date}: 2018-06-07 to 2019-02-13

Index       A
2018-06-07  0.6387
2018-06-08  0.1725
2018-06-09  0.7799
2018-06-10  0.9465
2018-06-11  0.6442
2018-06-12  0.2956
2018-06-13  0.4306
⋮
2019-02-06  0.0922
2019-02-07  0.2896
2019-02-08  0.7657
2019-02-09  0.3571
2019-02-10  0.3426
2019-02-11  0.8987
2019-02-12  0.5782
2019-02-13  0.1525

julia> y = TS(rand(252), x.index .- Month(6))
252x1 TS{Float64,Dates.Date}: 2017-12-07 to 2018-08-13

Index       A
2017-12-07  0.6627
2017-12-08  0.3502
2017-12-09  0.412
2017-12-10  0.3185
2017-12-11  0.1577
2017-12-12  0.5328
2017-12-13  0.4082
⋮
2018-08-06  0.543
2018-08-07  0.9269
2018-08-08  0.3528
2018-08-09  0.3683
2018-08-10  0.6014
2018-08-11  0.0949
2018-08-12  0.9848
2018-08-13  0.4084

julia> ljoin(x, y)
252x2 TS{Float64,Dates.Date}: 2018-06-07 to 2019-02-13

Index       A       A
2018-06-07  0.6387  0.5735
2018-06-08  0.1725  0.995
2018-06-09  0.7799  0.6028
2018-06-10  0.9465  0.0389
2018-06-11  0.6442  0.112
2018-06-12  0.2956  0.1023
2018-06-13  0.4306  0.1351
⋮
2019-02-06  0.0922  NaN
2019-02-07  0.2896  NaN
2019-02-08  0.7657  NaN
2019-02-09  0.3571  NaN
2019-02-10  0.3426  NaN
2019-02-11  0.8987  NaN
2019-02-12  0.5782  NaN
2019-02-13  0.1525  NaN

julia> rjoin(x, y)
247x2 TS{Float64,Dates.Date}: 2017-12-07 to 2018-08-13

Index       A       A
2017-12-07  NaN     0.6627
2017-12-08  NaN     0.3502
2017-12-09  NaN     0.412
2017-12-10  NaN     0.3185
2017-12-11  NaN     0.1577
2017-12-12  NaN     0.5328
2017-12-13  NaN     0.4082
⋮
2018-08-06  0.7927  0.8422
2018-08-07  0.2827  0.6389
2018-08-08  0.1649  0.8454
2018-08-09  0.2408  0.8888
2018-08-10  0.7502  0.8565
2018-08-11  0.6014  0.543
2018-08-12  0.7895  0.9269
2018-08-13  0.1213  0.3528