Joins

Joins

Outer Joins

One can perform a full outer join on the time indexes of two TS objects $x$ and $y$ in the following ways:

merge(x, y)

ojoin(x, y)

[x y]

hcat(x, y)

Where there are dates in the index of one that do not exist in the other, values will be filled with NaN objects. As the missing functionality matures in Julia's base syntax, it will eventually replace NaN in this context, since unfortunately NaN is only applicable for Float64 element types.

Base.merge — Function.

merge(x::TS,y::TS;join::Char='o')::TS

Merge two time series objects together by index with an optionally specified join type parameter.

...

Arguments

x::TS: Left side of the join.

y::TS: Right side of the join.

Optional args:

join::Char='o'::TS: Specifies the logic used to perform the merge, and may take on the values 'o' (outer join), 'i' (inner join), 'l' (left join), or 'r' (right join). Defaults to outer join, whose result is the same as hcat(x, y) or [x y].

...

source

Temporal.ojoin — Function.

ojoin(x::TS,y::TS)::TS

Outer join two TS objects by index.

Equivalent to x OUTER JOIN y ON x.index = y.index.

...

Arguments

x::TS: Left side of the join.

y::TS: Right side of the join.

...

source

Example

julia> x = TS(rand(252))
252x1 TS{Float64,Dates.Date}: 2018-06-07 to 2019-02-13

Index       A
2018-06-07  0.9548
2018-06-08  0.4713
2018-06-09  0.0953
2018-06-10  0.6349
2018-06-11  0.2663
2018-06-12  0.4196
2018-06-13  0.0453
⋮
2019-02-06  0.3518
2019-02-07  0.0344
2019-02-08  0.3026
2019-02-09  0.6058
2019-02-10  0.3346
2019-02-11  0.2896
2019-02-12  0.4804
2019-02-13  0.717

julia> y = TS(rand(252), x.index .- Month(6))
252x1 TS{Float64,Dates.Date}: 2017-12-07 to 2018-08-13

Index       A
2017-12-07  0.3937
2017-12-08  0.7336
2017-12-09  0.8618
2017-12-10  0.5125
2017-12-11  0.4632
2017-12-12  0.2459
2017-12-13  0.896
⋮
2018-08-06  0.0218
2018-08-07  0.2321
2018-08-08  0.9346
2018-08-09  0.789
2018-08-10  0.9093
2018-08-11  0.329
2018-08-12  0.1904
2018-08-13  0.4918

julia> [x y]
431x2 TS{Float64,Dates.Date}: 2017-12-07 to 2019-02-13

Index       A       A
2017-12-07  NaN     0.3937
2017-12-08  NaN     0.7336
2017-12-09  NaN     0.8618
2017-12-10  NaN     0.5125
2017-12-11  NaN     0.4632
2017-12-12  NaN     0.2459
2017-12-13  NaN     0.896
⋮
2019-02-06  0.3518  NaN
2019-02-07  0.0344  NaN
2019-02-08  0.3026  NaN
2019-02-09  0.6058  NaN
2019-02-10  0.3346  NaN
2019-02-11  0.2896  NaN
2019-02-12  0.4804  NaN
2019-02-13  0.717   NaN

Inner Joins

You can do inner joins on TS objects using the ijoin function, which will remove any observations corresponding to time steps where at least one of the joined objects is missing a row. This will basically keep only the rows where the time index of the left side and the right side intersect.

Temporal.ijoin — Function.

ijoin(x::TS,y::TS)::TS

Inner join two TS objects by index.

Equivalent to x INNER JOIN y on x.index = y.index.

...

Arguments

x::TS: Left side of the join.

y::TS: Right side of the join.

...

source

Example

julia> x = TS(rand(252))
252x1 TS{Float64,Dates.Date}: 2018-06-07 to 2019-02-13

Index       A
2018-06-07  0.8143
2018-06-08  0.1012
2018-06-09  0.6722
2018-06-10  0.7695
2018-06-11  0.4448
2018-06-12  0.0679
2018-06-13  0.8697
⋮
2019-02-06  0.0494
2019-02-07  0.0465
2019-02-08  0.4432
2019-02-09  0.9642
2019-02-10  0.3251
2019-02-11  0.4945
2019-02-12  0.0665
2019-02-13  0.7541

julia> y = TS(rand(252), x.index .- Month(6))
252x1 TS{Float64,Dates.Date}: 2017-12-07 to 2018-08-13

Index       A
2017-12-07  0.2295
2017-12-08  0.6387
2017-12-09  0.4364
2017-12-10  0.3994
2017-12-11  0.3279
2017-12-12  0.1056
2017-12-13  0.2056
⋮
2018-08-06  0.1629
2018-08-07  0.5451
2018-08-08  0.3297
2018-08-09  0.3765
2018-08-10  0.312
2018-08-11  0.2476
2018-08-12  0.4637
2018-08-13  0.6835

julia> ijoin(x, y)
ERROR: ArgumentError: number of rows of each array must match (got (68, 69))

Left/Right Joins

Left and right joins are performed similarly to inner joins and the typical SQL join queries using the index field each object as the joining key.

Left Join: keep all observations of the left side of the join, fill the right side with NaN's where missing the corresponding time index

Right Join: keep all observations of the right side of the join, fill the left side with NaN's where missing the corresponding time index

Temporal.ljoin — Function.

ljoin(x::TS, y::TS)::TS

Left join two TS objects by index.

Equivalent to x LEFT JOIN y ON x.index = y.index.

...

Arguments

x::TS: Left side of the join.

y::TS: Right side of the join.

...

source

Temporal.rjoin — Function.

rjoin(x::TS, y::TS)::TS

Right join two TS objects by index.

Equivalent to x RIGHT JOIN y ON x.index = y.index.

...

Arguments

x::TS: Left side of the join.

y::TS: Right side of the join.

...

source

Example

julia> x = TS(rand(252))
252x1 TS{Float64,Dates.Date}: 2018-06-07 to 2019-02-13

Index       A
2018-06-07  0.8512
2018-06-08  0.1481
2018-06-09  0.4314
2018-06-10  0.1024
2018-06-11  0.1197
2018-06-12  0.661
2018-06-13  0.4512
⋮
2019-02-06  0.0622
2019-02-07  0.6154
2019-02-08  0.6997
2019-02-09  0.5679
2019-02-10  0.7445
2019-02-11  0.1244
2019-02-12  0.4485
2019-02-13  0.851

julia> y = TS(rand(252), x.index .- Month(6))
252x1 TS{Float64,Dates.Date}: 2017-12-07 to 2018-08-13

Index       A
2017-12-07  0.6546
2017-12-08  0.26
2017-12-09  0.679
2017-12-10  0.239
2017-12-11  0.5081
2017-12-12  0.2185
2017-12-13  0.1606
⋮
2018-08-06  0.9235
2018-08-07  0.0777
2018-08-08  0.2605
2018-08-09  0.011
2018-08-10  0.5285
2018-08-11  0.0926
2018-08-12  0.4714
2018-08-13  0.8796

julia> ljoin(x, y)
252x2 TS{Float64,Dates.Date}: 2018-06-07 to 2019-02-13

Index       A       A
2018-06-07  0.8512  0.3206
2018-06-08  0.1481  0.1457
2018-06-09  0.4314  0.4759
2018-06-10  0.1024  0.147
2018-06-11  0.1197  0.9474
2018-06-12  0.661   0.1721
2018-06-13  0.4512  0.1645
⋮
2019-02-06  0.0622  NaN
2019-02-07  0.6154  NaN
2019-02-08  0.6997  NaN
2019-02-09  0.5679  NaN
2019-02-10  0.7445  NaN
2019-02-11  0.1244  NaN
2019-02-12  0.4485  NaN
2019-02-13  0.851   NaN

julia> rjoin(x, y)
247x2 TS{Float64,Dates.Date}: 2017-12-07 to 2018-08-13

Index       A       A
2017-12-07  NaN     0.6546
2017-12-08  NaN     0.26
2017-12-09  NaN     0.679
2017-12-10  NaN     0.239
2017-12-11  NaN     0.5081
2017-12-12  NaN     0.2185
2017-12-13  NaN     0.1606
⋮
2018-08-06  0.9854  0.1981
2018-08-07  0.6368  0.7681
2018-08-08  0.8593  0.0867
2018-08-09  0.3122  0.9343
2018-08-10  0.5698  0.4698
2018-08-11  0.6103  0.9235
2018-08-12  0.2365  0.0777
2018-08-13  0.8692  0.2605