Joins

Joins

Outer Joins

One can perform a full outer join on the time indexes of two TS objects $x$ and $y$ in the following ways:

merge(x, y)

ojoin(x, y)

[x y]

hcat(x, y)

Where there are dates in the index of one that do not exist in the other, values will be filled with NaN objects. As the missing functionality matures in Julia's base syntax, it will eventually replace NaN in this context, since unfortunately NaN is only applicable for Float64 element types.

Base.merge — Function.

merge(x::TS,y::TS;join::Char='o')::TS

Merge two time series objects together by index with an optionally specified join type parameter.

...

Arguments

x::TS: Left side of the join.

y::TS: Right side of the join.

Optional args:

join::Char='o'::TS: Specifies the logic used to perform the merge, and may take on the values 'o' (outer join), 'i' (inner join), 'l' (left join), or 'r' (right join). Defaults to outer join, whose result is the same as hcat(x, y) or [x y].

...

source

Temporal.ojoin — Function.

ojoin(x::TS,y::TS)::TS

Outer join two TS objects by index.

Equivalent to x OUTER JOIN y ON x.index = y.index.

...

Arguments

x::TS: Left side of the join.

y::TS: Right side of the join.

...

source

Example

julia> x = TS(rand(252))
252x1 TS{Float64,Dates.Date}: 2018-06-06 to 2019-02-12

Index       A
2018-06-06  0.8517
2018-06-07  0.608
2018-06-08  0.7228
2018-06-09  0.1463
2018-06-10  0.371
2018-06-11  0.6801
2018-06-12  0.7865
⋮
2019-02-05  0.6732
2019-02-06  0.5204
2019-02-07  0.7349
2019-02-08  0.3217
2019-02-09  0.9711
2019-02-10  0.9715
2019-02-11  0.4886
2019-02-12  0.6526

julia> y = TS(rand(252), x.index .- Month(6))
252x1 TS{Float64,Dates.Date}: 2017-12-06 to 2018-08-12

Index       A
2017-12-06  0.9059
2017-12-07  0.3446
2017-12-08  0.5177
2017-12-09  0.3029
2017-12-10  0.4193
2017-12-11  0.5206
2017-12-12  0.1795
⋮
2018-08-05  0.2705
2018-08-06  0.196
2018-08-07  0.0182
2018-08-08  0.495
2018-08-09  0.4932
2018-08-10  0.4096
2018-08-11  0.7614
2018-08-12  0.101

julia> [x y]
431x2 TS{Float64,Dates.Date}: 2017-12-06 to 2019-02-12

Index       A       A
2017-12-06  NaN     0.9059
2017-12-07  NaN     0.3446
2017-12-08  NaN     0.5177
2017-12-09  NaN     0.3029
2017-12-10  NaN     0.4193
2017-12-11  NaN     0.5206
2017-12-12  NaN     0.1795
⋮
2019-02-05  0.6732  NaN
2019-02-06  0.5204  NaN
2019-02-07  0.7349  NaN
2019-02-08  0.3217  NaN
2019-02-09  0.9711  NaN
2019-02-10  0.9715  NaN
2019-02-11  0.4886  NaN
2019-02-12  0.6526  NaN

Inner Joins

You can do inner joins on TS objects using the ijoin function, which will remove any observations corresponding to time steps where at least one of the joined objects is missing a row. This will basically keep only the rows where the time index of the left side and the right side intersect.

Temporal.ijoin — Function.

ijoin(x::TS,y::TS)::TS

Inner join two TS objects by index.

Equivalent to x INNER JOIN y on x.index = y.index.

...

Arguments

x::TS: Left side of the join.

y::TS: Right side of the join.

...

source

Example

julia> x = TS(rand(252))
252x1 TS{Float64,Dates.Date}: 2018-06-06 to 2019-02-12

Index       A
2018-06-06  0.3928
2018-06-07  0.4375
2018-06-08  0.5661
2018-06-09  0.6737
2018-06-10  0.2428
2018-06-11  0.4955
2018-06-12  0.7658
⋮
2019-02-05  0.7169
2019-02-06  0.4476
2019-02-07  0.3595
2019-02-08  0.0671
2019-02-09  0.0963
2019-02-10  0.2366
2019-02-11  0.3645
2019-02-12  0.3626

julia> y = TS(rand(252), x.index .- Month(6))
252x1 TS{Float64,Dates.Date}: 2017-12-06 to 2018-08-12

Index       A
2017-12-06  0.6492
2017-12-07  0.278
2017-12-08  0.9662
2017-12-09  0.3451
2017-12-10  0.4075
2017-12-11  0.0742
2017-12-12  0.6188
⋮
2018-08-05  0.7185
2018-08-06  0.7211
2018-08-07  0.7737
2018-08-08  0.4397
2018-08-09  0.2962
2018-08-10  0.0258
2018-08-11  0.4593
2018-08-12  0.9291

julia> ijoin(x, y)
ERROR: ArgumentError: number of rows of each array must match (got (68, 69))

Left/Right Joins

Left and right joins are performed similarly to inner joins and the typical SQL join queries using the index field each object as the joining key.

Left Join: keep all observations of the left side of the join, fill the right side with NaN's where missing the corresponding time index

Right Join: keep all observations of the right side of the join, fill the left side with NaN's where missing the corresponding time index

Temporal.ljoin — Function.

ljoin(x::TS, y::TS)::TS

Left join two TS objects by index.

Equivalent to x LEFT JOIN y ON x.index = y.index.

...

Arguments

x::TS: Left side of the join.

y::TS: Right side of the join.

...

source

Temporal.rjoin — Function.

rjoin(x::TS, y::TS)::TS

Right join two TS objects by index.

Equivalent to x RIGHT JOIN y ON x.index = y.index.

...

Arguments

x::TS: Left side of the join.

y::TS: Right side of the join.

...

source

Example

julia> x = TS(rand(252))
252x1 TS{Float64,Dates.Date}: 2018-06-06 to 2019-02-12

Index       A
2018-06-06  0.0085
2018-06-07  0.3964
2018-06-08  0.4876
2018-06-09  0.128
2018-06-10  0.285
2018-06-11  0.9123
2018-06-12  0.1213
⋮
2019-02-05  0.6349
2019-02-06  0.7037
2019-02-07  0.7545
2019-02-08  0.934
2019-02-09  0.2784
2019-02-10  0.2181
2019-02-11  0.8229
2019-02-12  0.9147

julia> y = TS(rand(252), x.index .- Month(6))
252x1 TS{Float64,Dates.Date}: 2017-12-06 to 2018-08-12

Index       A
2017-12-06  0.9123
2017-12-07  0.6881
2017-12-08  0.7244
2017-12-09  0.0065
2017-12-10  0.1072
2017-12-11  0.7309
2017-12-12  0.6504
⋮
2018-08-05  0.8142
2018-08-06  0.68
2018-08-07  0.4828
2018-08-08  0.9939
2018-08-09  0.7769
2018-08-10  0.5026
2018-08-11  0.58
2018-08-12  0.8006

julia> ljoin(x, y)
252x2 TS{Float64,Dates.Date}: 2018-06-06 to 2019-02-12

Index       A       A
2018-06-06  0.0085  0.7529
2018-06-07  0.3964  0.797
2018-06-08  0.4876  0.7244
2018-06-09  0.128   0.6978
2018-06-10  0.285   0.8919
2018-06-11  0.9123  0.162
2018-06-12  0.1213  0.4546
⋮
2019-02-05  0.6349  NaN
2019-02-06  0.7037  NaN
2019-02-07  0.7545  NaN
2019-02-08  0.934   NaN
2019-02-09  0.2784  NaN
2019-02-10  0.2181  NaN
2019-02-11  0.8229  NaN
2019-02-12  0.9147  NaN

julia> rjoin(x, y)
247x2 TS{Float64,Dates.Date}: 2017-12-06 to 2018-08-12

Index       A       A
2017-12-06  NaN     0.9123
2017-12-07  NaN     0.6881
2017-12-08  NaN     0.7244
2017-12-09  NaN     0.0065
2017-12-10  NaN     0.1072
2017-12-11  NaN     0.7309
2017-12-12  NaN     0.6504
⋮
2018-08-05  0.9062  0.1812
2018-08-06  0.1841  0.4143
2018-08-07  0.1195  0.6304
2018-08-08  0.065   0.0054
2018-08-09  0.0405  0.6639
2018-08-10  0.5442  0.8142
2018-08-11  0.2272  0.68
2018-08-12  0.8205  0.4828