Joins

Outer Joins

One can perform a full outer join on the time indexes of two TS objects $x$ and $y$ in the following ways:

merge(x, y)

ojoin(x, y)

[x y]

hcat(x, y)

Where there are dates in the index of one that do not exist in the other, values will be filled with NaN objects. As the missing functionality matures in Julia's base syntax, it will eventually replace NaN in this context, since unfortunately NaN is only applicable for Float64 element types.

Base.merge — Function

merge(x::TS,y::TS;join::Char='o')::TS

Merge two time series objects together by index with an optionally specified join type parameter.

...

Arguments

x::TS: Left side of the join.

y::TS: Right side of the join.

Optional args:

join::Char='o'::TS: Specifies the logic used to perform the merge, and may take on the values 'o' (outer join), 'i' (inner join), 'l' (left join), or 'r' (right join). Defaults to outer join, whose result is the same as hcat(x, y) or [x y].

...

Temporal.ojoin — Function

ojoin(x::TS,y::TS)::TS

Outer join two TS objects by index.

Equivalent to x OUTER JOIN y ON x.index = y.index.

...

Arguments

x::TS: Left side of the join.

y::TS: Right side of the join.

...

Example

julia> x = TS(rand(252))
252x1 TS{Float64,Dates.Date}: 2019-04-11 to 2019-12-18

Index       A
2019-04-11  0.5762
2019-04-12  0.7168
2019-04-13  0.8916
2019-04-14  0.4089
2019-04-15  0.6511
2019-04-16  0.7499
2019-04-17  0.6387
⋮
2019-12-11  0.2629
2019-12-12  0.7203
2019-12-13  0.7312
2019-12-14  0.027
2019-12-15  0.6887
2019-12-16  0.0805
2019-12-17  0.4434
2019-12-18  0.7567

julia> y = TS(rand(252), x.index .- Month(6))
252x1 TS{Float64,Dates.Date}: 2018-10-11 to 2019-06-18

Index       A
2018-10-11  0.4895
2018-10-12  0.8831
2018-10-13  0.3967
2018-10-14  0.9068
2018-10-15  0.7323
2018-10-16  0.304
2018-10-17  0.2971
⋮
2019-06-11  0.9413
2019-06-12  0.6157
2019-06-13  0.0277
2019-06-14  0.5304
2019-06-15  0.2026
2019-06-16  0.8154
2019-06-17  0.7328
2019-06-18  0.9477

julia> [x y]
431x2 TS{Float64,Dates.Date}: 2018-10-11 to 2019-12-18

Index       A       A
2018-10-11  NaN     0.4895
2018-10-12  NaN     0.8831
2018-10-13  NaN     0.3967
2018-10-14  NaN     0.9068
2018-10-15  NaN     0.7323
2018-10-16  NaN     0.304
2018-10-17  NaN     0.2971
⋮
2019-12-11  0.2629  NaN
2019-12-12  0.7203  NaN
2019-12-13  0.7312  NaN
2019-12-14  0.027   NaN
2019-12-15  0.6887  NaN
2019-12-16  0.0805  NaN
2019-12-17  0.4434  NaN
2019-12-18  0.7567  NaN

Inner Joins

You can do inner joins on TS objects using the ijoin function, which will remove any observations corresponding to time steps where at least one of the joined objects is missing a row. This will basically keep only the rows where the time index of the left side and the right side intersect.

Temporal.ijoin — Function

ijoin(x::TS,y::TS)::TS

Inner join two TS objects by index.

Equivalent to x INNER JOIN y on x.index = y.index.

...

Arguments

x::TS: Left side of the join.

y::TS: Right side of the join.

...

Example

julia> x = TS(rand(252))
252x1 TS{Float64,Dates.Date}: 2019-04-11 to 2019-12-18

Index       A
2019-04-11  0.6998
2019-04-12  0.5847
2019-04-13  0.5917
2019-04-14  0.7245
2019-04-15  0.3389
2019-04-16  0.7364
2019-04-17  0.9641
⋮
2019-12-11  0.1715
2019-12-12  0.1075
2019-12-13  0.0319
2019-12-14  0.114
2019-12-15  0.7662
2019-12-16  0.2461
2019-12-17  0.0081
2019-12-18  0.2081

julia> y = TS(rand(252), x.index .- Month(6))
252x1 TS{Float64,Dates.Date}: 2018-10-11 to 2019-06-18

Index       A
2018-10-11  0.9594
2018-10-12  0.016
2018-10-13  0.1666
2018-10-14  0.3902
2018-10-15  0.9745
2018-10-16  0.9993
2018-10-17  0.4295
⋮
2019-06-11  0.4917
2019-06-12  0.2359
2019-06-13  0.3153
2019-06-14  0.9211
2019-06-15  0.1482
2019-06-16  0.2172
2019-06-17  0.5754
2019-06-18  0.5905

julia> ijoin(x, y)
ERROR: ArgumentError: number of rows of each array must match (got (68, 69))

Left/Right Joins

Left and right joins are performed similarly to inner joins and the typical SQL join queries using the index field each object as the joining key.

Left Join: keep all observations of the left side of the join, fill the right side with NaN's where missing the corresponding time index

Right Join: keep all observations of the right side of the join, fill the left side with NaN's where missing the corresponding time index

Temporal.ljoin — Function

ljoin(x::TS, y::TS)::TS

Left join two TS objects by index.

Equivalent to x LEFT JOIN y ON x.index = y.index.

...

Arguments

x::TS: Left side of the join.

y::TS: Right side of the join.

...

Temporal.rjoin — Function

rjoin(x::TS, y::TS)::TS

Right join two TS objects by index.

Equivalent to x RIGHT JOIN y ON x.index = y.index.

...

Arguments

x::TS: Left side of the join.

y::TS: Right side of the join.

...

Example

julia> x = TS(rand(252))
252x1 TS{Float64,Dates.Date}: 2019-04-11 to 2019-12-18

Index       A
2019-04-11  0.2813
2019-04-12  0.0525
2019-04-13  0.4935
2019-04-14  0.1587
2019-04-15  0.1686
2019-04-16  0.8587
2019-04-17  0.46
⋮
2019-12-11  0.1644
2019-12-12  0.3248
2019-12-13  0.9817
2019-12-14  0.6984
2019-12-15  0.8109
2019-12-16  0.642
2019-12-17  0.4082
2019-12-18  0.9852

julia> y = TS(rand(252), x.index .- Month(6))
252x1 TS{Float64,Dates.Date}: 2018-10-11 to 2019-06-18

Index       A
2018-10-11  0.5313
2018-10-12  0.4151
2018-10-13  0.28
2018-10-14  0.7547
2018-10-15  0.7529
2018-10-16  0.6527
2018-10-17  0.3376
⋮
2019-06-11  0.0609
2019-06-12  0.2947
2019-06-13  0.0068
2019-06-14  0.8482
2019-06-15  0.6426
2019-06-16  0.8464
2019-06-17  0.0287
2019-06-18  0.3619

julia> ljoin(x, y)
252x2 TS{Float64,Dates.Date}: 2019-04-11 to 2019-12-18

Index       A       A
2019-04-11  0.2813  0.21
2019-04-12  0.0525  0.8447
2019-04-13  0.4935  0.1181
2019-04-14  0.1587  0.4396
2019-04-15  0.1686  0.9276
2019-04-16  0.8587  0.8895
2019-04-17  0.46    0.9315
⋮
2019-12-11  0.1644  NaN
2019-12-12  0.3248  NaN
2019-12-13  0.9817  NaN
2019-12-14  0.6984  NaN
2019-12-15  0.8109  NaN
2019-12-16  0.642   NaN
2019-12-17  0.4082  NaN
2019-12-18  0.9852  NaN

julia> rjoin(x, y)
247x2 TS{Float64,Dates.Date}: 2018-10-11 to 2019-06-18

Index       A       A
2018-10-11  NaN     0.5313
2018-10-12  NaN     0.4151
2018-10-13  NaN     0.28
2018-10-14  NaN     0.7547
2018-10-15  NaN     0.7529
2018-10-16  NaN     0.6527
2018-10-17  NaN     0.3376
⋮
2019-06-11  0.0761  0.683
2019-06-12  0.5945  0.8226
2019-06-13  0.761   0.1868
2019-06-14  0.2029  0.9452
2019-06-15  0.935   0.6182
2019-06-16  0.4919  0.0609
2019-06-17  0.5071  0.2947
2019-06-18  0.2414  0.0068