Combining

Joins

Outer Joins

One can perform a full outer join on the time indexes of two TS objects $x$ and $y$ in the following ways:

merge(x, y)

ojoin(x, y)

[x y]

hcat(x, y)

Where there are dates in the index of one that do not exist in the other, values will be filled with NaN objects. As the missing functionality matures in Julia's base syntax, it will eventually replace NaN in this context, since unfortunately NaN is only applicable for Float64 element types.

Base.merge — Function.

merge(x::TS,y::TS;join::Char='o')::TS

Merge two time series objects together by index with an optionally specified join type parameter.

...

Arguments

x::TS: Left side of the join.

y::TS: Right side of the join.

Optional args:

join::Char='o'::TS: Specifies the logic used to perform the merge, and may take on the values 'o' (outer join), 'i' (inner join), 'l' (left join), or 'r' (right join). Defaults to outer join, whose result is the same as hcat(x, y) or [x y].

...

source

Temporal.ojoin — Function.

ojoin(x::TS,y::TS)::TS

Outer join two TS objects by index.

Equivalent to x OUTER JOIN y ON x.index = y.index.

...

Arguments

x::TS: Left side of the join.

y::TS: Right side of the join.

...

source

Example

julia> x = TS(rand(252))
252x1 TS{Float64,Dates.Date}: 2018-06-07 to 2019-02-13

Index       A
2018-06-07  0.0577
2018-06-08  0.4246
2018-06-09  0.7098
2018-06-10  0.2693
2018-06-11  0.1829
2018-06-12  0.3882
2018-06-13  0.6969
⋮
2019-02-06  0.467
2019-02-07  0.5367
2019-02-08  0.402
2019-02-09  0.0143
2019-02-10  0.4519
2019-02-11  0.1306
2019-02-12  0.9225
2019-02-13  0.7875

julia> y = TS(rand(252), x.index .- Month(6))
252x1 TS{Float64,Dates.Date}: 2017-12-07 to 2018-08-13

Index       A
2017-12-07  0.4265
2017-12-08  0.2563
2017-12-09  0.4492
2017-12-10  0.7788
2017-12-11  0.2047
2017-12-12  0.0999
2017-12-13  0.6087
⋮
2018-08-06  0.5312
2018-08-07  0.3467
2018-08-08  0.1935
2018-08-09  0.229
2018-08-10  0.5098
2018-08-11  0.3449
2018-08-12  0.0439
2018-08-13  0.407

julia> [x y]
431x2 TS{Float64,Dates.Date}: 2017-12-07 to 2019-02-13

Index       A       A
2017-12-07  NaN     0.4265
2017-12-08  NaN     0.2563
2017-12-09  NaN     0.4492
2017-12-10  NaN     0.7788
2017-12-11  NaN     0.2047
2017-12-12  NaN     0.0999
2017-12-13  NaN     0.6087
⋮
2019-02-06  0.467   NaN
2019-02-07  0.5367  NaN
2019-02-08  0.402   NaN
2019-02-09  0.0143  NaN
2019-02-10  0.4519  NaN
2019-02-11  0.1306  NaN
2019-02-12  0.9225  NaN
2019-02-13  0.7875  NaN

Inner Joins

You can do inner joins on TS objects using the ijoin function, which will remove any observations corresponding to time steps where at least one of the joined objects is missing a row. This will basically keep only the rows where the time index of the left side and the right side intersect.

Temporal.ijoin — Function.

ijoin(x::TS,y::TS)::TS

Inner join two TS objects by index.

Equivalent to x INNER JOIN y on x.index = y.index.

...

Arguments

x::TS: Left side of the join.

y::TS: Right side of the join.

...

source

Example

julia> x = TS(rand(252))
252x1 TS{Float64,Dates.Date}: 2018-06-07 to 2019-02-13

Index       A
2018-06-07  0.7896
2018-06-08  0.2376
2018-06-09  0.8841
2018-06-10  0.8991
2018-06-11  0.6147
2018-06-12  0.6498
2018-06-13  0.9691
⋮
2019-02-06  0.9606
2019-02-07  0.85
2019-02-08  0.8129
2019-02-09  0.6863
2019-02-10  0.4202
2019-02-11  0.8075
2019-02-12  0.9516
2019-02-13  0.7282

julia> y = TS(rand(252), x.index .- Month(6))
252x1 TS{Float64,Dates.Date}: 2017-12-07 to 2018-08-13

Index       A
2017-12-07  0.5739
2017-12-08  0.5375
2017-12-09  0.2898
2017-12-10  0.0303
2017-12-11  0.7072
2017-12-12  0.4317
2017-12-13  0.8083
⋮
2018-08-06  0.6517
2018-08-07  0.8852
2018-08-08  0.008
2018-08-09  0.9681
2018-08-10  0.187
2018-08-11  0.0807
2018-08-12  0.3807
2018-08-13  0.3025

julia> ijoin(x, y)
ERROR: ArgumentError: number of rows of each array must match (got (68, 69))

Left/Right Joins

Left and right joins are performed similarly to inner joins and the typical SQL join queries using the index field each object as the joining key.

Left Join: keep all observations of the left side of the join, fill the right side with NaN's where missing the corresponding time index

Right Join: keep all observations of the right side of the join, fill the left side with NaN's where missing the corresponding time index

Temporal.ljoin — Function.

ljoin(x::TS, y::TS)::TS

Left join two TS objects by index.

Equivalent to x LEFT JOIN y ON x.index = y.index.

...

Arguments

x::TS: Left side of the join.

y::TS: Right side of the join.

...

source

Temporal.rjoin — Function.

rjoin(x::TS, y::TS)::TS

Right join two TS objects by index.

Equivalent to x RIGHT JOIN y ON x.index = y.index.

...

Arguments

x::TS: Left side of the join.

y::TS: Right side of the join.

...

source

Example

julia> x = TS(rand(252))
252x1 TS{Float64,Dates.Date}: 2018-06-07 to 2019-02-13

Index       A
2018-06-07  0.5987
2018-06-08  0.406
2018-06-09  0.3804
2018-06-10  0.826
2018-06-11  0.1453
2018-06-12  0.8527
2018-06-13  0.6398
⋮
2019-02-06  0.6058
2019-02-07  0.6045
2019-02-08  0.9415
2019-02-09  0.3024
2019-02-10  0.9001
2019-02-11  0.9907
2019-02-12  0.2615
2019-02-13  0.2466

julia> y = TS(rand(252), x.index .- Month(6))
252x1 TS{Float64,Dates.Date}: 2017-12-07 to 2018-08-13

Index       A
2017-12-07  0.6958
2017-12-08  0.8849
2017-12-09  0.3466
2017-12-10  0.5727
2017-12-11  0.7534
2017-12-12  0.4686
2017-12-13  0.7947
⋮
2018-08-06  0.507
2018-08-07  0.1382
2018-08-08  0.8979
2018-08-09  0.7981
2018-08-10  0.5994
2018-08-11  0.2149
2018-08-12  0.613
2018-08-13  0.2009

julia> ljoin(x, y)
252x2 TS{Float64,Dates.Date}: 2018-06-07 to 2019-02-13

Index       A       A
2018-06-07  0.5987  0.1157
2018-06-08  0.406   0.719
2018-06-09  0.3804  0.2471
2018-06-10  0.826   0.9302
2018-06-11  0.1453  0.2395
2018-06-12  0.8527  0.8637
2018-06-13  0.6398  0.0513
⋮
2019-02-06  0.6058  NaN
2019-02-07  0.6045  NaN
2019-02-08  0.9415  NaN
2019-02-09  0.3024  NaN
2019-02-10  0.9001  NaN
2019-02-11  0.9907  NaN
2019-02-12  0.2615  NaN
2019-02-13  0.2466  NaN

julia> rjoin(x, y)
247x2 TS{Float64,Dates.Date}: 2017-12-07 to 2018-08-13

Index       A       A
2017-12-07  NaN     0.6958
2017-12-08  NaN     0.8849
2017-12-09  NaN     0.3466
2017-12-10  NaN     0.5727
2017-12-11  NaN     0.7534
2017-12-12  NaN     0.4686
2017-12-13  NaN     0.7947
⋮
2018-08-06  0.1317  0.4008
2018-08-07  0.3082  0.6635
2018-08-08  0.321   0.1438
2018-08-09  0.2278  0.8673
2018-08-10  0.0966  0.659
2018-08-11  0.369   0.507
2018-08-12  0.5943  0.1382
2018-08-13  0.5153  0.8979