Combining

Joins

Outer Joins

One can perform a full outer join on the time indexes of two TS objects $x$ and $y$ in the following ways:

merge(x, y)

ojoin(x, y)

[x y]

hcat(x, y)

Where there are dates in the index of one that do not exist in the other, values will be filled with NaN objects. As the missing functionality matures in Julia's base syntax, it will eventually replace NaN in this context, since unfortunately NaN is only applicable for Float64 element types.

Base.merge — Function.

merge(x::TS,y::TS;join::Char='o')::TS

Merge two time series objects together by index with an optionally specified join type parameter.

...

Arguments

x::TS: Left side of the join.

y::TS: Right side of the join.

Optional args:

join::Char='o'::TS: Specifies the logic used to perform the merge, and may take on the values 'o' (outer join), 'i' (inner join), 'l' (left join), or 'r' (right join). Defaults to outer join, whose result is the same as hcat(x, y) or [x y].

...

source

Temporal.ojoin — Function.

ojoin(x::TS,y::TS)::TS

Outer join two TS objects by index.

Equivalent to x OUTER JOIN y ON x.index = y.index.

...

Arguments

x::TS: Left side of the join.

y::TS: Right side of the join.

...

source

Example

julia> x = TS(rand(252))
252x1 TS{Float64,Dates.Date}: 2018-06-28 to 2019-03-06

Index       A
2018-06-28  0.3053
2018-06-29  0.8563
2018-06-30  0.5967
2018-07-01  0.6342
2018-07-02  0.1956
2018-07-03  0.4037
2018-07-04  0.1044
⋮
2019-02-27  0.9285
2019-02-28  0.4225
2019-03-01  0.1727
2019-03-02  0.1808
2019-03-03  0.4889
2019-03-04  0.5263
2019-03-05  0.0235
2019-03-06  0.6255

julia> y = TS(rand(252), x.index .- Month(6))
252x1 TS{Float64,Dates.Date}: 2017-12-28 to 2018-09-06

Index       A
2017-12-28  0.4625
2017-12-29  0.3433
2017-12-30  0.0099
2018-01-01  0.2273
2018-01-02  0.1083
2018-01-03  0.0423
2018-01-04  0.6887
⋮
2018-08-27  0.3228
2018-08-28  0.1085
2018-09-01  0.3243
2018-09-02  0.3321
2018-09-03  0.687
2018-09-04  0.7485
2018-09-05  0.9688
2018-09-06  0.7301

julia> [x y]
431x2 TS{Float64,Dates.Date}: 2017-12-28 to 2019-03-06

Index       A       A
2017-12-28  NaN     0.4625
2017-12-29  NaN     0.3433
2017-12-30  NaN     0.0099
2018-01-01  NaN     0.2273
2018-01-02  NaN     0.1083
2018-01-03  NaN     0.0423
2018-01-04  NaN     0.6887
⋮
2019-02-27  0.9285  NaN
2019-02-28  0.4225  NaN
2019-03-01  0.1727  NaN
2019-03-02  0.1808  NaN
2019-03-03  0.4889  NaN
2019-03-04  0.5263  NaN
2019-03-05  0.0235  NaN
2019-03-06  0.6255  NaN

Inner Joins

You can do inner joins on TS objects using the ijoin function, which will remove any observations corresponding to time steps where at least one of the joined objects is missing a row. This will basically keep only the rows where the time index of the left side and the right side intersect.

Temporal.ijoin — Function.

ijoin(x::TS,y::TS)::TS

Inner join two TS objects by index.

Equivalent to x INNER JOIN y on x.index = y.index.

...

Arguments

x::TS: Left side of the join.

y::TS: Right side of the join.

...

source

Example

julia> x = TS(rand(252))
252x1 TS{Float64,Dates.Date}: 2018-06-28 to 2019-03-06

Index       A
2018-06-28  0.8615
2018-06-29  0.0003
2018-06-30  0.664
2018-07-01  0.6872
2018-07-02  0.7282
2018-07-03  0.6146
2018-07-04  0.1999
⋮
2019-02-27  0.0392
2019-02-28  0.9802
2019-03-01  0.0321
2019-03-02  0.8267
2019-03-03  0.8533
2019-03-04  0.8809
2019-03-05  0.0058
2019-03-06  0.0735

julia> y = TS(rand(252), x.index .- Month(6))
252x1 TS{Float64,Dates.Date}: 2017-12-28 to 2018-09-06

Index       A
2017-12-28  0.1504
2017-12-29  0.047
2017-12-30  0.8987
2018-01-01  0.8937
2018-01-02  0.0089
2018-01-03  0.4677
2018-01-04  0.7026
⋮
2018-08-27  0.7301
2018-08-28  0.4309
2018-09-01  0.7739
2018-09-02  0.7154
2018-09-03  0.1268
2018-09-04  0.7074
2018-09-05  0.8635
2018-09-06  0.8999

julia> ijoin(x, y)
ERROR: ArgumentError: number of rows of each array must match (got (68, 69))

Left/Right Joins

Left and right joins are performed similarly to inner joins and the typical SQL join queries using the index field each object as the joining key.

Left Join: keep all observations of the left side of the join, fill the right side with NaN's where missing the corresponding time index

Right Join: keep all observations of the right side of the join, fill the left side with NaN's where missing the corresponding time index

Temporal.ljoin — Function.

ljoin(x::TS, y::TS)::TS

Left join two TS objects by index.

Equivalent to x LEFT JOIN y ON x.index = y.index.

...

Arguments

x::TS: Left side of the join.

y::TS: Right side of the join.

...

source

Temporal.rjoin — Function.

rjoin(x::TS, y::TS)::TS

Right join two TS objects by index.

Equivalent to x RIGHT JOIN y ON x.index = y.index.

...

Arguments

x::TS: Left side of the join.

y::TS: Right side of the join.

...

source

Example

julia> x = TS(rand(252))
252x1 TS{Float64,Dates.Date}: 2018-06-28 to 2019-03-06

Index       A
2018-06-28  0.4421
2018-06-29  0.4506
2018-06-30  0.7857
2018-07-01  0.3702
2018-07-02  0.3079
2018-07-03  0.9789
2018-07-04  0.0279
⋮
2019-02-27  0.7617
2019-02-28  0.2027
2019-03-01  0.4491
2019-03-02  0.0092
2019-03-03  0.1002
2019-03-04  0.3618
2019-03-05  0.6682
2019-03-06  0.3093

julia> y = TS(rand(252), x.index .- Month(6))
252x1 TS{Float64,Dates.Date}: 2017-12-28 to 2018-09-06

Index       A
2017-12-28  0.5093
2017-12-29  0.89
2017-12-30  0.285
2018-01-01  0.0972
2018-01-02  0.8424
2018-01-03  0.765
2018-01-04  0.9593
⋮
2018-08-27  0.0189
2018-08-28  0.5844
2018-09-01  0.7358
2018-09-02  0.0412
2018-09-03  0.8823
2018-09-04  0.7676
2018-09-05  0.7084
2018-09-06  0.903

julia> ljoin(x, y)
252x2 TS{Float64,Dates.Date}: 2018-06-28 to 2019-03-06

Index       A       A
2018-06-28  0.4421  0.2007
2018-06-29  0.4506  0.9164
2018-06-30  0.7857  0.7244
2018-07-01  0.3702  0.8643
2018-07-02  0.3079  0.7545
2018-07-03  0.9789  0.2834
2018-07-04  0.0279  0.736
⋮
2019-02-27  0.7617  NaN
2019-02-28  0.2027  NaN
2019-03-01  0.4491  NaN
2019-03-02  0.0092  NaN
2019-03-03  0.1002  NaN
2019-03-04  0.3618  NaN
2019-03-05  0.6682  NaN
2019-03-06  0.3093  NaN

julia> rjoin(x, y)
247x2 TS{Float64,Dates.Date}: 2017-12-28 to 2018-09-06

Index       A       A
2017-12-28  NaN     0.5093
2017-12-29  NaN     0.89
2017-12-30  NaN     0.285
2018-01-01  NaN     0.0972
2018-01-02  NaN     0.8424
2018-01-03  NaN     0.765
2018-01-04  NaN     0.9593
⋮
2018-08-27  0.6993  0.1614
2018-08-28  0.2925  0.9521
2018-09-01  0.4231  0.5371
2018-09-02  0.4794  0.9744
2018-09-03  0.8341  0.7818
2018-09-04  0.8986  0.0189
2018-09-05  0.9885  0.5844
2018-09-06  0.941   0.7358