Joins

Joins

Outer Joins

One can perform a full outer join on the time indexes of two TS objects $x$ and $y$ in the following ways:

merge(x, y)

ojoin(x, y)

[x y]

hcat(x, y)

Where there are dates in the index of one that do not exist in the other, values will be filled with NaN objects. As the missing functionality matures in Julia's base syntax, it will eventually replace NaN in this context, since unfortunately NaN is only applicable for Float64 element types.

Base.merge — Function.

merge(x::TS,y::TS;join::Char='o')::TS

Merge two time series objects together by index with an optionally specified join type parameter.

...

Arguments

x::TS: Left side of the join.

y::TS: Right side of the join.

Optional args:

join::Char='o'::TS: Specifies the logic used to perform the merge, and may take on the values 'o' (outer join), 'i' (inner join), 'l' (left join), or 'r' (right join). Defaults to outer join, whose result is the same as hcat(x, y) or [x y].

...

source

Temporal.ojoin — Function.

ojoin(x::TS,y::TS)::TS

Outer join two TS objects by index.

Equivalent to x OUTER JOIN y ON x.index = y.index.

...

Arguments

x::TS: Left side of the join.

y::TS: Right side of the join.

...

source

Example

julia> x = TS(rand(252))
252x1 TS{Float64,Dates.Date}: 2018-06-05 to 2019-02-11

Index       A
2018-06-05  0.0086
2018-06-06  0.0263
2018-06-07  0.3475
2018-06-08  0.1195
2018-06-09  0.056
2018-06-10  0.9845
2018-06-11  0.085
⋮
2019-02-04  0.8221
2019-02-05  0.3042
2019-02-06  0.158
2019-02-07  0.2956
2019-02-08  0.9347
2019-02-09  0.0693
2019-02-10  0.0028
2019-02-11  0.3489

julia> y = TS(rand(252), x.index .- Month(6))
252x1 TS{Float64,Dates.Date}: 2017-12-05 to 2018-08-11

Index       A
2017-12-05  0.3412
2017-12-06  0.6055
2017-12-07  0.5814
2017-12-08  0.6388
2017-12-09  0.377
2017-12-10  0.4009
2017-12-11  0.5189
⋮
2018-08-04  0.1913
2018-08-05  0.9406
2018-08-06  0.6995
2018-08-07  0.5263
2018-08-08  0.809
2018-08-09  0.1279
2018-08-10  0.3711
2018-08-11  0.2879

julia> [x y]
431x2 TS{Float64,Dates.Date}: 2017-12-05 to 2019-02-11

Index       A       A
2017-12-05  NaN     0.3412
2017-12-06  NaN     0.6055
2017-12-07  NaN     0.5814
2017-12-08  NaN     0.6388
2017-12-09  NaN     0.377
2017-12-10  NaN     0.4009
2017-12-11  NaN     0.5189
⋮
2019-02-04  0.8221  NaN
2019-02-05  0.3042  NaN
2019-02-06  0.158   NaN
2019-02-07  0.2956  NaN
2019-02-08  0.9347  NaN
2019-02-09  0.0693  NaN
2019-02-10  0.0028  NaN
2019-02-11  0.3489  NaN

Inner Joins

You can do inner joins on TS objects using the ijoin function, which will remove any observations corresponding to time steps where at least one of the joined objects is missing a row. This will basically keep only the rows where the time index of the left side and the right side intersect.

Temporal.ijoin — Function.

ijoin(x::TS,y::TS)::TS

Inner join two TS objects by index.

Equivalent to x INNER JOIN y on x.index = y.index.

...

Arguments

x::TS: Left side of the join.

y::TS: Right side of the join.

...

source

Example

julia> x = TS(rand(252))
252x1 TS{Float64,Dates.Date}: 2018-06-05 to 2019-02-11

Index       A
2018-06-05  0.9092
2018-06-06  0.2833
2018-06-07  0.9177
2018-06-08  0.4381
2018-06-09  0.8166
2018-06-10  0.342
2018-06-11  0.3111
⋮
2019-02-04  0.0219
2019-02-05  0.9364
2019-02-06  0.4664
2019-02-07  0.1267
2019-02-08  0.7836
2019-02-09  0.1595
2019-02-10  0.1416
2019-02-11  0.7003

julia> y = TS(rand(252), x.index .- Month(6))
252x1 TS{Float64,Dates.Date}: 2017-12-05 to 2018-08-11

Index       A
2017-12-05  0.9114
2017-12-06  0.841
2017-12-07  0.6322
2017-12-08  0.4301
2017-12-09  0.6844
2017-12-10  0.4174
2017-12-11  0.5385
⋮
2018-08-04  0.8549
2018-08-05  0.0997
2018-08-06  0.4203
2018-08-07  0.3163
2018-08-08  0.2397
2018-08-09  0.6774
2018-08-10  0.7312
2018-08-11  0.79

julia> ijoin(x, y)
ERROR: ArgumentError: number of rows of each array must match (got (68, 69))

Left/Right Joins

Left and right joins are performed similarly to inner joins and the typical SQL join queries using the index field each object as the joining key.

Left Join: keep all observations of the left side of the join, fill the right side with NaN's where missing the corresponding time index

Right Join: keep all observations of the right side of the join, fill the left side with NaN's where missing the corresponding time index

Temporal.ljoin — Function.

ljoin(x::TS, y::TS)::TS

Left join two TS objects by index.

Equivalent to x LEFT JOIN y ON x.index = y.index.

...

Arguments

x::TS: Left side of the join.

y::TS: Right side of the join.

...

source

Temporal.rjoin — Function.

rjoin(x::TS, y::TS)::TS

Right join two TS objects by index.

Equivalent to x RIGHT JOIN y ON x.index = y.index.

...

Arguments

x::TS: Left side of the join.

y::TS: Right side of the join.

...

source

Example

julia> x = TS(rand(252))
252x1 TS{Float64,Dates.Date}: 2018-06-05 to 2019-02-11

Index       A
2018-06-05  0.0738
2018-06-06  0.0532
2018-06-07  0.6825
2018-06-08  0.9555
2018-06-09  0.9731
2018-06-10  0.4566
2018-06-11  0.6315
⋮
2019-02-04  0.5434
2019-02-05  0.0229
2019-02-06  0.098
2019-02-07  0.6277
2019-02-08  0.1496
2019-02-09  0.3981
2019-02-10  0.2357
2019-02-11  0.2517

julia> y = TS(rand(252), x.index .- Month(6))
252x1 TS{Float64,Dates.Date}: 2017-12-05 to 2018-08-11

Index       A
2017-12-05  0.8563
2017-12-06  0.6315
2017-12-07  0.1772
2017-12-08  0.8627
2017-12-09  0.5181
2017-12-10  0.7642
2017-12-11  0.7018
⋮
2018-08-04  0.6956
2018-08-05  0.1721
2018-08-06  0.5439
2018-08-07  0.0176
2018-08-08  0.5227
2018-08-09  0.3054
2018-08-10  0.5324
2018-08-11  0.7751

julia> ljoin(x, y)
252x2 TS{Float64,Dates.Date}: 2018-06-05 to 2019-02-11

Index       A       A
2018-06-05  0.0738  0.9281
2018-06-06  0.0532  0.9396
2018-06-07  0.6825  0.9657
2018-06-08  0.9555  0.9478
2018-06-09  0.9731  0.1414
2018-06-10  0.4566  0.5298
2018-06-11  0.6315  0.2443
⋮
2019-02-04  0.5434  NaN
2019-02-05  0.0229  NaN
2019-02-06  0.098   NaN
2019-02-07  0.6277  NaN
2019-02-08  0.1496  NaN
2019-02-09  0.3981  NaN
2019-02-10  0.2357  NaN
2019-02-11  0.2517  NaN

julia> rjoin(x, y)
247x2 TS{Float64,Dates.Date}: 2017-12-05 to 2018-08-11

Index       A       A
2017-12-05  NaN     0.8563
2017-12-06  NaN     0.6315
2017-12-07  NaN     0.1772
2017-12-08  NaN     0.8627
2017-12-09  NaN     0.5181
2017-12-10  NaN     0.7642
2017-12-11  NaN     0.7018
⋮
2018-08-04  0.6024  0.8236
2018-08-05  0.0609  0.6151
2018-08-06  0.4485  0.3041
2018-08-07  0.9206  0.8931
2018-08-08  0.9666  0.1606
2018-08-09  0.1033  0.6956
2018-08-10  0.5112  0.1721
2018-08-11  0.9091  0.5439