Joins

Joins

Outer Joins

One can perform a full outer join on the time indexes of two TS objects $x$ and $y$ in the following ways:

merge(x, y)

ojoin(x, y)

[x y]

hcat(x, y)

Where there are dates in the index of one that do not exist in the other, values will be filled with NaN objects. As the missing functionality matures in Julia's base syntax, it will eventually replace NaN in this context, since unfortunately NaN is only applicable for Float64 element types.

Base.merge — Function.

merge(x::TS,y::TS;join::Char='o')::TS

Merge two time series objects together by index with an optionally specified join type parameter.

...

Arguments

x::TS: Left side of the join.

y::TS: Right side of the join.

Optional args:

join::Char='o'::TS: Specifies the logic used to perform the merge, and may take on the values 'o' (outer join), 'i' (inner join), 'l' (left join), or 'r' (right join). Defaults to outer join, whose result is the same as hcat(x, y) or [x y].

...

source

Temporal.ojoin — Function.

ojoin(x::TS,y::TS)::TS

Outer join two TS objects by index.

Equivalent to x OUTER JOIN y ON x.index = y.index.

...

Arguments

x::TS: Left side of the join.

y::TS: Right side of the join.

...

source

Example

julia> x = TS(rand(252))
252x1 TS{Float64,Dates.Date}: 2018-06-07 to 2019-02-13

Index       A
2018-06-07  0.1577
2018-06-08  0.3882
2018-06-09  0.3486
2018-06-10  0.6758
2018-06-11  0.6375
2018-06-12  0.9437
2018-06-13  0.3484
⋮
2019-02-06  0.012
2019-02-07  0.4825
2019-02-08  0.6579
2019-02-09  0.0882
2019-02-10  0.5729
2019-02-11  0.7125
2019-02-12  0.8357
2019-02-13  0.6706

julia> y = TS(rand(252), x.index .- Month(6))
252x1 TS{Float64,Dates.Date}: 2017-12-07 to 2018-08-13

Index       A
2017-12-07  0.9761
2017-12-08  0.238
2017-12-09  0.0979
2017-12-10  0.4398
2017-12-11  0.3691
2017-12-12  0.1148
2017-12-13  0.0401
⋮
2018-08-06  0.9298
2018-08-07  0.4043
2018-08-08  0.5309
2018-08-09  0.845
2018-08-10  0.518
2018-08-11  0.7991
2018-08-12  0.9118
2018-08-13  0.5247

julia> [x y]
431x2 TS{Float64,Dates.Date}: 2017-12-07 to 2019-02-13

Index       A       A
2017-12-07  NaN     0.9761
2017-12-08  NaN     0.238
2017-12-09  NaN     0.0979
2017-12-10  NaN     0.4398
2017-12-11  NaN     0.3691
2017-12-12  NaN     0.1148
2017-12-13  NaN     0.0401
⋮
2019-02-06  0.012   NaN
2019-02-07  0.4825  NaN
2019-02-08  0.6579  NaN
2019-02-09  0.0882  NaN
2019-02-10  0.5729  NaN
2019-02-11  0.7125  NaN
2019-02-12  0.8357  NaN
2019-02-13  0.6706  NaN

Inner Joins

You can do inner joins on TS objects using the ijoin function, which will remove any observations corresponding to time steps where at least one of the joined objects is missing a row. This will basically keep only the rows where the time index of the left side and the right side intersect.

Temporal.ijoin — Function.

ijoin(x::TS,y::TS)::TS

Inner join two TS objects by index.

Equivalent to x INNER JOIN y on x.index = y.index.

...

Arguments

x::TS: Left side of the join.

y::TS: Right side of the join.

...

source

Example

julia> x = TS(rand(252))
252x1 TS{Float64,Dates.Date}: 2018-06-07 to 2019-02-13

Index       A
2018-06-07  0.1735
2018-06-08  0.5469
2018-06-09  0.6654
2018-06-10  0.5587
2018-06-11  0.6661
2018-06-12  0.9937
2018-06-13  0.5636
⋮
2019-02-06  0.3138
2019-02-07  0.1658
2019-02-08  0.5956
2019-02-09  0.6943
2019-02-10  0.3965
2019-02-11  0.8393
2019-02-12  0.1033
2019-02-13  0.3984

julia> y = TS(rand(252), x.index .- Month(6))
252x1 TS{Float64,Dates.Date}: 2017-12-07 to 2018-08-13

Index       A
2017-12-07  0.3514
2017-12-08  0.7977
2017-12-09  0.2584
2017-12-10  0.8271
2017-12-11  0.8452
2017-12-12  0.2088
2017-12-13  0.6672
⋮
2018-08-06  0.2367
2018-08-07  0.7096
2018-08-08  0.0076
2018-08-09  0.4202
2018-08-10  0.1419
2018-08-11  0.882
2018-08-12  0.1923
2018-08-13  0.1881

julia> ijoin(x, y)
ERROR: ArgumentError: number of rows of each array must match (got (68, 69))

Left/Right Joins

Left and right joins are performed similarly to inner joins and the typical SQL join queries using the index field each object as the joining key.

Left Join: keep all observations of the left side of the join, fill the right side with NaN's where missing the corresponding time index

Right Join: keep all observations of the right side of the join, fill the left side with NaN's where missing the corresponding time index

Temporal.ljoin — Function.

ljoin(x::TS, y::TS)::TS

Left join two TS objects by index.

Equivalent to x LEFT JOIN y ON x.index = y.index.

...

Arguments

x::TS: Left side of the join.

y::TS: Right side of the join.

...

source

Temporal.rjoin — Function.

rjoin(x::TS, y::TS)::TS

Right join two TS objects by index.

Equivalent to x RIGHT JOIN y ON x.index = y.index.

...

Arguments

x::TS: Left side of the join.

y::TS: Right side of the join.

...

source

Example

julia> x = TS(rand(252))
252x1 TS{Float64,Dates.Date}: 2018-06-07 to 2019-02-13

Index       A
2018-06-07  0.0525
2018-06-08  0.2136
2018-06-09  0.5898
2018-06-10  0.8543
2018-06-11  0.2607
2018-06-12  0.4745
2018-06-13  0.8715
⋮
2019-02-06  0.6985
2019-02-07  0.6648
2019-02-08  0.6992
2019-02-09  0.057
2019-02-10  0.1991
2019-02-11  0.6597
2019-02-12  0.7783
2019-02-13  0.0396

julia> y = TS(rand(252), x.index .- Month(6))
252x1 TS{Float64,Dates.Date}: 2017-12-07 to 2018-08-13

Index       A
2017-12-07  0.1009
2017-12-08  0.685
2017-12-09  0.5894
2017-12-10  0.1115
2017-12-11  0.2609
2017-12-12  0.2655
2017-12-13  0.3255
⋮
2018-08-06  0.7235
2018-08-07  0.9126
2018-08-08  0.6784
2018-08-09  0.0316
2018-08-10  0.9599
2018-08-11  0.7466
2018-08-12  0.3329
2018-08-13  0.3136

julia> ljoin(x, y)
252x2 TS{Float64,Dates.Date}: 2018-06-07 to 2019-02-13

Index       A       A
2018-06-07  0.0525  0.3687
2018-06-08  0.2136  0.1913
2018-06-09  0.5898  0.4628
2018-06-10  0.8543  0.9094
2018-06-11  0.2607  0.8243
2018-06-12  0.4745  0.4356
2018-06-13  0.8715  0.6856
⋮
2019-02-06  0.6985  NaN
2019-02-07  0.6648  NaN
2019-02-08  0.6992  NaN
2019-02-09  0.057   NaN
2019-02-10  0.1991  NaN
2019-02-11  0.6597  NaN
2019-02-12  0.7783  NaN
2019-02-13  0.0396  NaN

julia> rjoin(x, y)
247x2 TS{Float64,Dates.Date}: 2017-12-07 to 2018-08-13

Index       A       A
2017-12-07  NaN     0.1009
2017-12-08  NaN     0.685
2017-12-09  NaN     0.5894
2017-12-10  NaN     0.1115
2017-12-11  NaN     0.2609
2017-12-12  NaN     0.2655
2017-12-13  NaN     0.3255
⋮
2018-08-06  0.7715  0.0682
2018-08-07  0.6049  0.2311
2018-08-08  0.9051  0.5816
2018-08-09  0.972   0.7798
2018-08-10  0.152   0.744
2018-08-11  0.6541  0.7235
2018-08-12  0.9754  0.9126
2018-08-13  0.7462  0.6784