Joins

Outer Joins

One can perform a full outer join on the time indexes of two TS objects $x$ and $y$ in the following ways:

merge(x, y)

ojoin(x, y)

[x y]

hcat(x, y)

Where there are dates in the index of one that do not exist in the other, values will be filled with NaN objects. As the missing functionality matures in Julia's base syntax, it will eventually replace NaN in this context, since unfortunately NaN is only applicable for Float64 element types.

Base.merge — Function

merge(x::TS,y::TS;join::Char='o')::TS

Merge two time series objects together by index with an optionally specified join type parameter.

...

Arguments

x::TS: Left side of the join.

y::TS: Right side of the join.

Optional args:

join::Char='o'::TS: Specifies the logic used to perform the merge, and may take on the values 'o' (outer join), 'i' (inner join), 'l' (left join), or 'r' (right join). Defaults to outer join, whose result is the same as hcat(x, y) or [x y].

...

Temporal.ojoin — Function

ojoin(x::TS,y::TS)::TS

Outer join two TS objects by index.

Equivalent to x OUTER JOIN y ON x.index = y.index.

...

Arguments

x::TS: Left side of the join.

y::TS: Right side of the join.

...

Example

julia> x = TS(rand(252))
252x1 TS{Float64,Dates.Date}: 2019-05-10 to 2020-01-16

Index       A
2019-05-10  0.1531
2019-05-11  0.9075
2019-05-12  0.0585
2019-05-13  0.3019
2019-05-14  0.5302
2019-05-15  0.4623
2019-05-16  0.7043
⋮
2020-01-09  0.4113
2020-01-10  0.7326
2020-01-11  0.3331
2020-01-12  0.8231
2020-01-13  0.7334
2020-01-14  0.3095
2020-01-15  0.9203
2020-01-16  0.376

julia> y = TS(rand(252), x.index .- Month(6))
252x1 TS{Float64,Dates.Date}: 2018-11-10 to 2019-07-16

Index       A
2018-11-10  0.8036
2018-11-11  0.6731
2018-11-12  0.255
2018-11-13  0.6114
2018-11-14  0.0873
2018-11-15  0.6051
2018-11-16  0.4576
⋮
2019-07-09  0.9832
2019-07-10  0.9171
2019-07-11  0.9729
2019-07-12  0.6037
2019-07-13  0.7787
2019-07-14  0.9225
2019-07-15  0.7139
2019-07-16  0.323

julia> [x y]
431x2 TS{Float64,Dates.Date}: 2018-11-10 to 2020-01-16

Index       A       A
2018-11-10  NaN     0.8036
2018-11-11  NaN     0.6731
2018-11-12  NaN     0.255
2018-11-13  NaN     0.6114
2018-11-14  NaN     0.0873
2018-11-15  NaN     0.6051
2018-11-16  NaN     0.4576
⋮
2020-01-09  0.4113  NaN
2020-01-10  0.7326  NaN
2020-01-11  0.3331  NaN
2020-01-12  0.8231  NaN
2020-01-13  0.7334  NaN
2020-01-14  0.3095  NaN
2020-01-15  0.9203  NaN
2020-01-16  0.376   NaN

Inner Joins

You can do inner joins on TS objects using the ijoin function, which will remove any observations corresponding to time steps where at least one of the joined objects is missing a row. This will basically keep only the rows where the time index of the left side and the right side intersect.

Temporal.ijoin — Function

ijoin(x::TS,y::TS)::TS

Inner join two TS objects by index.

Equivalent to x INNER JOIN y on x.index = y.index.

...

Arguments

x::TS: Left side of the join.

y::TS: Right side of the join.

...

Example

julia> x = TS(rand(252))
252x1 TS{Float64,Dates.Date}: 2019-05-10 to 2020-01-16

Index       A
2019-05-10  0.7191
2019-05-11  0.5908
2019-05-12  0.6634
2019-05-13  0.2989
2019-05-14  0.0354
2019-05-15  0.0256
2019-05-16  0.039
⋮
2020-01-09  0.4442
2020-01-10  0.0498
2020-01-11  0.7989
2020-01-12  0.9304
2020-01-13  0.9447
2020-01-14  0.5299
2020-01-15  0.0498
2020-01-16  0.9011

julia> y = TS(rand(252), x.index .- Month(6))
252x1 TS{Float64,Dates.Date}: 2018-11-10 to 2019-07-16

Index       A
2018-11-10  0.2043
2018-11-11  0.6677
2018-11-12  0.5163
2018-11-13  0.2529
2018-11-14  0.9911
2018-11-15  0.6841
2018-11-16  0.4279
⋮
2019-07-09  0.6272
2019-07-10  0.4359
2019-07-11  0.1078
2019-07-12  0.4096
2019-07-13  0.0949
2019-07-14  0.29
2019-07-15  0.6113
2019-07-16  0.0197

julia> ijoin(x, y)
ERROR: ArgumentError: number of rows of each array must match (got (67, 68))

Left/Right Joins

Left and right joins are performed similarly to inner joins and the typical SQL join queries using the index field each object as the joining key.

Left Join: keep all observations of the left side of the join, fill the right side with NaN's where missing the corresponding time index

Right Join: keep all observations of the right side of the join, fill the left side with NaN's where missing the corresponding time index

Temporal.ljoin — Function

ljoin(x::TS, y::TS)::TS

Left join two TS objects by index.

Equivalent to x LEFT JOIN y ON x.index = y.index.

...

Arguments

x::TS: Left side of the join.

y::TS: Right side of the join.

...

Temporal.rjoin — Function

rjoin(x::TS, y::TS)::TS

Right join two TS objects by index.

Equivalent to x RIGHT JOIN y ON x.index = y.index.

...

Arguments

x::TS: Left side of the join.

y::TS: Right side of the join.

...

Example

julia> x = TS(rand(252))
252x1 TS{Float64,Dates.Date}: 2019-05-10 to 2020-01-16

Index       A
2019-05-10  0.6562
2019-05-11  0.7332
2019-05-12  0.7155
2019-05-13  0.2124
2019-05-14  0.0284
2019-05-15  0.411
2019-05-16  0.795
⋮
2020-01-09  0.8108
2020-01-10  0.1765
2020-01-11  0.1381
2020-01-12  0.7377
2020-01-13  0.8751
2020-01-14  0.5273
2020-01-15  0.9344
2020-01-16  0.5351

julia> y = TS(rand(252), x.index .- Month(6))
252x1 TS{Float64,Dates.Date}: 2018-11-10 to 2019-07-16

Index       A
2018-11-10  0.7625
2018-11-11  0.5287
2018-11-12  0.4983
2018-11-13  0.3308
2018-11-14  0.6428
2018-11-15  0.5758
2018-11-16  0.6214
⋮
2019-07-09  0.4644
2019-07-10  0.1313
2019-07-11  0.0195
2019-07-12  0.7062
2019-07-13  0.0995
2019-07-14  0.6459
2019-07-15  0.5082
2019-07-16  0.2948

julia> ljoin(x, y)
252x2 TS{Float64,Dates.Date}: 2019-05-10 to 2020-01-16

Index       A       A
2019-05-10  0.6562  0.2828
2019-05-11  0.7332  0.8713
2019-05-12  0.7155  0.2972
2019-05-13  0.2124  0.217
2019-05-14  0.0284  0.6421
2019-05-15  0.411   0.5479
2019-05-16  0.795   0.7587
⋮
2020-01-09  0.8108  NaN
2020-01-10  0.1765  NaN
2020-01-11  0.1381  NaN
2020-01-12  0.7377  NaN
2020-01-13  0.8751  NaN
2020-01-14  0.5273  NaN
2020-01-15  0.9344  NaN
2020-01-16  0.5351  NaN

julia> rjoin(x, y)
246x2 TS{Float64,Dates.Date}: 2018-11-10 to 2019-07-16

Index       A       A
2018-11-10  NaN     0.7625
2018-11-11  NaN     0.5287
2018-11-12  NaN     0.4983
2018-11-13  NaN     0.3308
2018-11-14  NaN     0.6428
2018-11-15  NaN     0.5758
2018-11-16  NaN     0.6214
⋮
2019-07-09  0.1973  0.8315
2019-07-10  0.8725  0.3614
2019-07-11  0.4498  0.834
2019-07-12  0.0918  0.4866
2019-07-13  0.4225  0.8039
2019-07-14  0.0944  0.4065
2019-07-15  0.4361  0.4644
2019-07-16  0.3131  0.1313