Joins

Joins

Outer Joins

One can perform a full outer join on the time indexes of two TS objects $x$ and $y$ in the following ways:

merge(x, y)

ojoin(x, y)

[x y]

hcat(x, y)

Where there are dates in the index of one that do not exist in the other, values will be filled with NaN objects. As the missing functionality matures in Julia's base syntax, it will eventually replace NaN in this context, since unfortunately NaN is only applicable for Float64 element types.

Base.merge — Function.

merge(x::TS,y::TS;join::Char='o')::TS

Merge two time series objects together by index with an optionally specified join type parameter.

...

Arguments

x::TS: Left side of the join.

y::TS: Right side of the join.

Optional args:

join::Char='o'::TS: Specifies the logic used to perform the merge, and may take on the values 'o' (outer join), 'i' (inner join), 'l' (left join), or 'r' (right join). Defaults to outer join, whose result is the same as hcat(x, y) or [x y].

...

source

Temporal.ojoin — Function.

ojoin(x::TS,y::TS)::TS

Outer join two TS objects by index.

Equivalent to x OUTER JOIN y ON x.index = y.index.

...

Arguments

x::TS: Left side of the join.

y::TS: Right side of the join.

...

source

Example

julia> x = TS(rand(252))
252x1 TS{Float64,Dates.Date}: 2018-06-06 to 2019-02-12

Index       A
2018-06-06  0.8509
2018-06-07  0.1578
2018-06-08  0.0349
2018-06-09  0.4601
2018-06-10  0.2752
2018-06-11  0.8345
2018-06-12  0.9091
⋮
2019-02-05  0.0982
2019-02-06  0.0747
2019-02-07  0.5954
2019-02-08  0.2011
2019-02-09  0.2308
2019-02-10  0.5432
2019-02-11  0.9793
2019-02-12  0.823

julia> y = TS(rand(252), x.index .- Month(6))
252x1 TS{Float64,Dates.Date}: 2017-12-06 to 2018-08-12

Index       A
2017-12-06  0.4447
2017-12-07  0.4038
2017-12-08  0.7708
2017-12-09  0.8691
2017-12-10  0.0437
2017-12-11  0.3629
2017-12-12  0.7803
⋮
2018-08-05  0.4935
2018-08-06  0.8683
2018-08-07  0.5094
2018-08-08  0.6607
2018-08-09  0.1693
2018-08-10  0.4577
2018-08-11  0.0424
2018-08-12  0.323

julia> [x y]
431x2 TS{Float64,Dates.Date}: 2017-12-06 to 2019-02-12

Index       A       A
2017-12-06  NaN     0.4447
2017-12-07  NaN     0.4038
2017-12-08  NaN     0.7708
2017-12-09  NaN     0.8691
2017-12-10  NaN     0.0437
2017-12-11  NaN     0.3629
2017-12-12  NaN     0.7803
⋮
2019-02-05  0.0982  NaN
2019-02-06  0.0747  NaN
2019-02-07  0.5954  NaN
2019-02-08  0.2011  NaN
2019-02-09  0.2308  NaN
2019-02-10  0.5432  NaN
2019-02-11  0.9793  NaN
2019-02-12  0.823   NaN

Inner Joins

You can do inner joins on TS objects using the ijoin function, which will remove any observations corresponding to time steps where at least one of the joined objects is missing a row. This will basically keep only the rows where the time index of the left side and the right side intersect.

Temporal.ijoin — Function.

ijoin(x::TS,y::TS)::TS

Inner join two TS objects by index.

Equivalent to x INNER JOIN y on x.index = y.index.

...

Arguments

x::TS: Left side of the join.

y::TS: Right side of the join.

...

source

Example

julia> x = TS(rand(252))
252x1 TS{Float64,Dates.Date}: 2018-06-06 to 2019-02-12

Index       A
2018-06-06  0.5479
2018-06-07  0.0538
2018-06-08  0.5754
2018-06-09  0.4448
2018-06-10  0.8933
2018-06-11  0.6434
2018-06-12  0.24
⋮
2019-02-05  0.6961
2019-02-06  0.8992
2019-02-07  0.9098
2019-02-08  0.7548
2019-02-09  0.051
2019-02-10  0.6253
2019-02-11  0.2602
2019-02-12  0.82

julia> y = TS(rand(252), x.index .- Month(6))
252x1 TS{Float64,Dates.Date}: 2017-12-06 to 2018-08-12

Index       A
2017-12-06  0.05
2017-12-07  0.0843
2017-12-08  0.2516
2017-12-09  0.1426
2017-12-10  0.014
2017-12-11  0.8841
2017-12-12  0.6744
⋮
2018-08-05  0.8945
2018-08-06  0.0728
2018-08-07  0.4503
2018-08-08  0.7191
2018-08-09  0.2781
2018-08-10  0.9217
2018-08-11  0.8482
2018-08-12  0.1161

julia> ijoin(x, y)
ERROR: ArgumentError: number of rows of each array must match (got (68, 69))

Left/Right Joins

Left and right joins are performed similarly to inner joins and the typical SQL join queries using the index field each object as the joining key.

Left Join: keep all observations of the left side of the join, fill the right side with NaN's where missing the corresponding time index

Right Join: keep all observations of the right side of the join, fill the left side with NaN's where missing the corresponding time index

Temporal.ljoin — Function.

ljoin(x::TS, y::TS)::TS

Left join two TS objects by index.

Equivalent to x LEFT JOIN y ON x.index = y.index.

...

Arguments

x::TS: Left side of the join.

y::TS: Right side of the join.

...

source

Temporal.rjoin — Function.

rjoin(x::TS, y::TS)::TS

Right join two TS objects by index.

Equivalent to x RIGHT JOIN y ON x.index = y.index.

...

Arguments

x::TS: Left side of the join.

y::TS: Right side of the join.

...

source

Example

julia> x = TS(rand(252))
252x1 TS{Float64,Dates.Date}: 2018-06-06 to 2019-02-12

Index       A
2018-06-06  0.9075
2018-06-07  0.648
2018-06-08  0.7892
2018-06-09  0.8856
2018-06-10  0.614
2018-06-11  0.0468
2018-06-12  0.6237
⋮
2019-02-05  0.4318
2019-02-06  0.4797
2019-02-07  0.4958
2019-02-08  0.4935
2019-02-09  0.9972
2019-02-10  0.9733
2019-02-11  0.4916
2019-02-12  0.9071

julia> y = TS(rand(252), x.index .- Month(6))
252x1 TS{Float64,Dates.Date}: 2017-12-06 to 2018-08-12

Index       A
2017-12-06  0.3516
2017-12-07  0.2481
2017-12-08  0.8218
2017-12-09  0.2921
2017-12-10  0.5234
2017-12-11  0.9563
2017-12-12  0.9524
⋮
2018-08-05  0.3463
2018-08-06  0.3168
2018-08-07  0.0122
2018-08-08  0.5183
2018-08-09  0.0234
2018-08-10  0.6804
2018-08-11  0.2423
2018-08-12  0.0791

julia> ljoin(x, y)
252x2 TS{Float64,Dates.Date}: 2018-06-06 to 2019-02-12

Index       A       A
2018-06-06  0.9075  0.8614
2018-06-07  0.648   0.8553
2018-06-08  0.7892  0.4933
2018-06-09  0.8856  0.356
2018-06-10  0.614   0.2849
2018-06-11  0.0468  0.3849
2018-06-12  0.6237  0.5608
⋮
2019-02-05  0.4318  NaN
2019-02-06  0.4797  NaN
2019-02-07  0.4958  NaN
2019-02-08  0.4935  NaN
2019-02-09  0.9972  NaN
2019-02-10  0.9733  NaN
2019-02-11  0.4916  NaN
2019-02-12  0.9071  NaN

julia> rjoin(x, y)
247x2 TS{Float64,Dates.Date}: 2017-12-06 to 2018-08-12

Index       A       A
2017-12-06  NaN     0.3516
2017-12-07  NaN     0.2481
2017-12-08  NaN     0.8218
2017-12-09  NaN     0.2921
2017-12-10  NaN     0.5234
2017-12-11  NaN     0.9563
2017-12-12  NaN     0.9524
⋮
2018-08-05  0.5539  0.9399
2018-08-06  0.9657  0.323
2018-08-07  0.3918  0.2232
2018-08-08  0.4656  0.8476
2018-08-09  0.8625  0.7762
2018-08-10  0.1403  0.3463
2018-08-11  0.2184  0.3168
2018-08-12  0.8806  0.0122