Joins

Joins

Outer Joins

One can perform a full outer join on the time indexes of two TS objects $x$ and $y$ in the following ways:

merge(x, y)

ojoin(x, y)

[x y]

hcat(x, y)

Where there are dates in the index of one that do not exist in the other, values will be filled with NaN objects. As the missing functionality matures in Julia's base syntax, it will eventually replace NaN in this context, since unfortunately NaN is only applicable for Float64 element types.

Base.merge — Function.

merge(x::TS,y::TS;join::Char='o')::TS

Merge two time series objects together by index with an optionally specified join type parameter.

...

Arguments

x::TS: Left side of the join.

y::TS: Right side of the join.

Optional args:

join::Char='o'::TS: Specifies the logic used to perform the merge, and may take on the values 'o' (outer join), 'i' (inner join), 'l' (left join), or 'r' (right join). Defaults to outer join, whose result is the same as hcat(x, y) or [x y].

...

source

Temporal.ojoin — Function.

ojoin(x::TS,y::TS)::TS

Outer join two TS objects by index.

Equivalent to x OUTER JOIN y ON x.index = y.index.

...

Arguments

x::TS: Left side of the join.

y::TS: Right side of the join.

...

source

Example

julia> x = TS(rand(252))
252x1 TS{Float64,Dates.Date}: 2018-06-03 to 2019-02-09

Index       A
2018-06-03  0.6653
2018-06-04  0.1229
2018-06-05  0.3973
2018-06-06  0.9107
2018-06-07  0.7024
2018-06-08  0.5832
2018-06-09  0.8553
⋮
2019-02-02  0.9228
2019-02-03  0.0255
2019-02-04  0.7428
2019-02-05  0.2172
2019-02-06  0.3268
2019-02-07  0.0153
2019-02-08  0.1005
2019-02-09  0.0336

julia> y = TS(rand(252), x.index .- Month(6))
252x1 TS{Float64,Dates.Date}: 2017-12-03 to 2018-08-09

Index       A
2017-12-03  0.3101
2017-12-04  0.5896
2017-12-05  0.2251
2017-12-06  0.3119
2017-12-07  0.8419
2017-12-08  0.1083
2017-12-09  0.146
⋮
2018-08-02  0.7684
2018-08-03  0.6252
2018-08-04  0.8279
2018-08-05  0.4671
2018-08-06  0.3785
2018-08-07  0.5076
2018-08-08  0.7661
2018-08-09  0.8268

julia> [x y]
431x2 TS{Float64,Dates.Date}: 2017-12-03 to 2019-02-09

Index       A       A
2017-12-03  NaN     0.3101
2017-12-04  NaN     0.5896
2017-12-05  NaN     0.2251
2017-12-06  NaN     0.3119
2017-12-07  NaN     0.8419
2017-12-08  NaN     0.1083
2017-12-09  NaN     0.146
⋮
2019-02-02  0.9228  NaN
2019-02-03  0.0255  NaN
2019-02-04  0.7428  NaN
2019-02-05  0.2172  NaN
2019-02-06  0.3268  NaN
2019-02-07  0.0153  NaN
2019-02-08  0.1005  NaN
2019-02-09  0.0336  NaN

Inner Joins

You can do inner joins on TS objects using the ijoin function, which will remove any observations corresponding to time steps where at least one of the joined objects is missing a row. This will basically keep only the rows where the time index of the left side and the right side intersect.

Temporal.ijoin — Function.

ijoin(x::TS,y::TS)::TS

Inner join two TS objects by index.

Equivalent to x INNER JOIN y on x.index = y.index.

...

Arguments

x::TS: Left side of the join.

y::TS: Right side of the join.

...

source

Example

julia> x = TS(rand(252))
252x1 TS{Float64,Dates.Date}: 2018-06-03 to 2019-02-09

Index       A
2018-06-03  0.6847
2018-06-04  0.7831
2018-06-05  0.8982
2018-06-06  0.1572
2018-06-07  0.5166
2018-06-08  0.3876
2018-06-09  0.2105
⋮
2019-02-02  0.2306
2019-02-03  0.3061
2019-02-04  0.7288
2019-02-05  0.6155
2019-02-06  0.8262
2019-02-07  0.6823
2019-02-08  0.9247
2019-02-09  0.38

julia> y = TS(rand(252), x.index .- Month(6))
252x1 TS{Float64,Dates.Date}: 2017-12-03 to 2018-08-09

Index       A
2017-12-03  0.3317
2017-12-04  0.1371
2017-12-05  0.9699
2017-12-06  0.8786
2017-12-07  0.5413
2017-12-08  0.5554
2017-12-09  0.0033
⋮
2018-08-02  0.0832
2018-08-03  0.9989
2018-08-04  0.0366
2018-08-05  0.391
2018-08-06  0.4451
2018-08-07  0.2951
2018-08-08  0.2809
2018-08-09  0.2694

julia> ijoin(x, y)
ERROR: ArgumentError: number of rows of each array must match (got (68, 69))

Left/Right Joins

Left and right joins are performed similarly to inner joins and the typical SQL join queries using the index field each object as the joining key.

Left Join: keep all observations of the left side of the join, fill the right side with NaN's where missing the corresponding time index

Right Join: keep all observations of the right side of the join, fill the left side with NaN's where missing the corresponding time index

Temporal.ljoin — Function.

ljoin(x::TS, y::TS)::TS

Left join two TS objects by index.

Equivalent to x LEFT JOIN y ON x.index = y.index.

...

Arguments

x::TS: Left side of the join.

y::TS: Right side of the join.

...

source

Temporal.rjoin — Function.

rjoin(x::TS, y::TS)::TS

Right join two TS objects by index.

Equivalent to x RIGHT JOIN y ON x.index = y.index.

...

Arguments

x::TS: Left side of the join.

y::TS: Right side of the join.

...

source

Example

julia> x = TS(rand(252))
252x1 TS{Float64,Dates.Date}: 2018-06-03 to 2019-02-09

Index       A
2018-06-03  0.4441
2018-06-04  0.8945
2018-06-05  0.8395
2018-06-06  0.4227
2018-06-07  0.8016
2018-06-08  0.2901
2018-06-09  0.8809
⋮
2019-02-02  0.3353
2019-02-03  0.2668
2019-02-04  0.6567
2019-02-05  0.781
2019-02-06  0.0971
2019-02-07  0.2161
2019-02-08  0.1629
2019-02-09  0.8418

julia> y = TS(rand(252), x.index .- Month(6))
252x1 TS{Float64,Dates.Date}: 2017-12-03 to 2018-08-09

Index       A
2017-12-03  0.9594
2017-12-04  0.6777
2017-12-05  0.9214
2017-12-06  0.9365
2017-12-07  0.1843
2017-12-08  0.9314
2017-12-09  0.7393
⋮
2018-08-02  0.1746
2018-08-03  0.1147
2018-08-04  0.0794
2018-08-05  0.7522
2018-08-06  0.5318
2018-08-07  0.9859
2018-08-08  0.7796
2018-08-09  0.0582

julia> ljoin(x, y)
252x2 TS{Float64,Dates.Date}: 2018-06-03 to 2019-02-09

Index       A       A
2018-06-03  0.4441  0.1405
2018-06-04  0.8945  0.2369
2018-06-05  0.8395  0.5148
2018-06-06  0.4227  0.2331
2018-06-07  0.8016  0.5876
2018-06-08  0.2901  0.3549
2018-06-09  0.8809  0.8472
⋮
2019-02-02  0.3353  NaN
2019-02-03  0.2668  NaN
2019-02-04  0.6567  NaN
2019-02-05  0.781   NaN
2019-02-06  0.0971  NaN
2019-02-07  0.2161  NaN
2019-02-08  0.1629  NaN
2019-02-09  0.8418  NaN

julia> rjoin(x, y)
247x2 TS{Float64,Dates.Date}: 2017-12-03 to 2018-08-09

Index       A       A
2017-12-03  NaN     0.9594
2017-12-04  NaN     0.6777
2017-12-05  NaN     0.9214
2017-12-06  NaN     0.9365
2017-12-07  NaN     0.1843
2017-12-08  NaN     0.9314
2017-12-09  NaN     0.7393
⋮
2018-08-02  0.118   0.4943
2018-08-03  0.0242  0.6489
2018-08-04  0.579   0.6857
2018-08-05  0.1797  0.5615
2018-08-06  0.7142  0.009
2018-08-07  0.1989  0.1746
2018-08-08  0.6052  0.1147
2018-08-09  0.874   0.0794