Combining

Joins

Outer Joins

One can perform a full outer join on the time indexes of two TS objects $x$ and $y$ in the following ways:

merge(x, y)

ojoin(x, y)

[x y]

hcat(x, y)

Where there are dates in the index of one that do not exist in the other, values will be filled with NaN objects. As the missing functionality matures in Julia's base syntax, it will eventually replace NaN in this context, since unfortunately NaN is only applicable for Float64 element types.

Base.merge — Function.

merge(x::TS,y::TS;join::Char='o')::TS

Merge two time series objects together by index with an optionally specified join type parameter.

...

Arguments

x::TS: Left side of the join.

y::TS: Right side of the join.

Optional args:

join::Char='o'::TS: Specifies the logic used to perform the merge, and may take on the values 'o' (outer join), 'i' (inner join), 'l' (left join), or 'r' (right join). Defaults to outer join, whose result is the same as hcat(x, y) or [x y].

...

source

Temporal.ojoin — Function.

ojoin(x::TS,y::TS)::TS

Outer join two TS objects by index.

Equivalent to x OUTER JOIN y ON x.index = y.index.

...

Arguments

x::TS: Left side of the join.

y::TS: Right side of the join.

...

source

Example

julia> x = TS(rand(252))
252x1 TS{Float64,Dates.Date}: 2018-07-10 to 2019-03-18

Index       A
2018-07-10  0.5331
2018-07-11  0.6924
2018-07-12  0.6981
2018-07-13  0.8046
2018-07-14  0.5111
2018-07-15  0.1963
2018-07-16  0.7507
⋮
2019-03-11  0.5484
2019-03-12  0.583
2019-03-13  0.4503
2019-03-14  0.5066
2019-03-15  0.1869
2019-03-16  0.2102
2019-03-17  0.702
2019-03-18  0.9141

julia> y = TS(rand(252), x.index .- Month(6))
252x1 TS{Float64,Dates.Date}: 2018-01-10 to 2018-09-18

Index       A
2018-01-10  0.098
2018-01-11  0.7445
2018-01-12  0.7986
2018-01-13  0.9858
2018-01-14  0.5427
2018-01-15  0.0807
2018-01-16  0.869
⋮
2018-09-11  0.4611
2018-09-12  0.8499
2018-09-13  0.8602
2018-09-14  0.8216
2018-09-15  0.9701
2018-09-16  0.7213
2018-09-17  0.3744
2018-09-18  0.4722

julia> [x y]
431x2 TS{Float64,Dates.Date}: 2018-01-10 to 2019-03-18

Index       A       A
2018-01-10  NaN     0.098
2018-01-11  NaN     0.7445
2018-01-12  NaN     0.7986
2018-01-13  NaN     0.9858
2018-01-14  NaN     0.5427
2018-01-15  NaN     0.0807
2018-01-16  NaN     0.869
⋮
2019-03-11  0.5484  NaN
2019-03-12  0.583   NaN
2019-03-13  0.4503  NaN
2019-03-14  0.5066  NaN
2019-03-15  0.1869  NaN
2019-03-16  0.2102  NaN
2019-03-17  0.702   NaN
2019-03-18  0.9141  NaN

Inner Joins

You can do inner joins on TS objects using the ijoin function, which will remove any observations corresponding to time steps where at least one of the joined objects is missing a row. This will basically keep only the rows where the time index of the left side and the right side intersect.

Temporal.ijoin — Function.

ijoin(x::TS,y::TS)::TS

Inner join two TS objects by index.

Equivalent to x INNER JOIN y on x.index = y.index.

...

Arguments

x::TS: Left side of the join.

y::TS: Right side of the join.

...

source

Example

julia> x = TS(rand(252))
252x1 TS{Float64,Dates.Date}: 2018-07-10 to 2019-03-18

Index       A
2018-07-10  0.2302
2018-07-11  0.7022
2018-07-12  0.5898
2018-07-13  0.8122
2018-07-14  0.8863
2018-07-15  0.2176
2018-07-16  0.5669
⋮
2019-03-11  0.0807
2019-03-12  0.986
2019-03-13  0.0418
2019-03-14  0.2773
2019-03-15  0.0178
2019-03-16  0.6787
2019-03-17  0.9436
2019-03-18  0.7351

julia> y = TS(rand(252), x.index .- Month(6))
252x1 TS{Float64,Dates.Date}: 2018-01-10 to 2018-09-18

Index       A
2018-01-10  0.4059
2018-01-11  0.1806
2018-01-12  0.6081
2018-01-13  0.4713
2018-01-14  0.3874
2018-01-15  0.783
2018-01-16  0.9973
⋮
2018-09-11  0.1722
2018-09-12  0.3688
2018-09-13  0.1082
2018-09-14  0.2098
2018-09-15  0.4393
2018-09-16  0.1287
2018-09-17  0.4061
2018-09-18  0.8187

julia> ijoin(x, y)
68x2 TS{Float64,Dates.Date}: 2018-07-10 to 2018-09-18

Index       A       A
2018-07-10  0.2302  0.9589
2018-07-11  0.7022  0.5992
2018-07-12  0.5898  0.7854
2018-07-13  0.8122  0.6843
2018-07-14  0.8863  0.3745
2018-07-15  0.2176  0.1664
2018-07-16  0.5669  0.4121
⋮
2018-09-11  0.3345  0.1722
2018-09-12  0.5513  0.3688
2018-09-13  0.6067  0.1082
2018-09-14  0.0223  0.2098
2018-09-15  0.1701  0.4393
2018-09-16  0.2965  0.1287
2018-09-17  0.3502  0.4061
2018-09-18  0.2224  0.8187

Left/Right Joins

Left and right joins are performed similarly to inner joins and the typical SQL join queries using the index field each object as the joining key.

Left Join: keep all observations of the left side of the join, fill the right side with NaN's where missing the corresponding time index

Right Join: keep all observations of the right side of the join, fill the left side with NaN's where missing the corresponding time index

Temporal.ljoin — Function.

ljoin(x::TS, y::TS)::TS

Left join two TS objects by index.

Equivalent to x LEFT JOIN y ON x.index = y.index.

...

Arguments

x::TS: Left side of the join.

y::TS: Right side of the join.

...

source

Temporal.rjoin — Function.

rjoin(x::TS, y::TS)::TS

Right join two TS objects by index.

Equivalent to x RIGHT JOIN y ON x.index = y.index.

...

Arguments

x::TS: Left side of the join.

y::TS: Right side of the join.

...

source

Example

julia> x = TS(rand(252))
252x1 TS{Float64,Dates.Date}: 2018-07-10 to 2019-03-18

Index       A
2018-07-10  0.8407
2018-07-11  0.1583
2018-07-12  0.9802
2018-07-13  0.1953
2018-07-14  0.6642
2018-07-15  0.4358
2018-07-16  0.5388
⋮
2019-03-11  0.6714
2019-03-12  0.813
2019-03-13  0.6495
2019-03-14  0.6146
2019-03-15  0.8379
2019-03-16  0.5979
2019-03-17  0.9048
2019-03-18  0.6063

julia> y = TS(rand(252), x.index .- Month(6))
252x1 TS{Float64,Dates.Date}: 2018-01-10 to 2018-09-18

Index       A
2018-01-10  0.0358
2018-01-11  0.8958
2018-01-12  0.3047
2018-01-13  0.0745
2018-01-14  0.9976
2018-01-15  0.5448
2018-01-16  0.0891
⋮
2018-09-11  0.0751
2018-09-12  0.8446
2018-09-13  0.2774
2018-09-14  0.1401
2018-09-15  0.9971
2018-09-16  0.4777
2018-09-17  0.0099
2018-09-18  0.0142

julia> ljoin(x, y)
252x2 TS{Float64,Dates.Date}: 2018-07-10 to 2019-03-18

Index       A       A
2018-07-10  0.8407  0.1084
2018-07-11  0.1583  0.1674
2018-07-12  0.9802  0.3165
2018-07-13  0.1953  0.5837
2018-07-14  0.6642  0.0806
2018-07-15  0.4358  0.6993
2018-07-16  0.5388  0.0867
⋮
2019-03-11  0.6714  NaN
2019-03-12  0.813   NaN
2019-03-13  0.6495  NaN
2019-03-14  0.6146  NaN
2019-03-15  0.8379  NaN
2019-03-16  0.5979  NaN
2019-03-17  0.9048  NaN
2019-03-18  0.6063  NaN

julia> rjoin(x, y)
247x2 TS{Float64,Dates.Date}: 2018-01-10 to 2018-09-18

Index       A       A
2018-01-10  NaN     0.0358
2018-01-11  NaN     0.8958
2018-01-12  NaN     0.3047
2018-01-13  NaN     0.0745
2018-01-14  NaN     0.9976
2018-01-15  NaN     0.5448
2018-01-16  NaN     0.0891
⋮
2018-09-11  0.6897  0.0708
2018-09-12  0.6232  0.8821
2018-09-13  0.4037  0.5305
2018-09-14  0.6437  0.2335
2018-09-15  0.4471  0.6611
2018-09-16  0.3086  0.0751
2018-09-17  0.5969  0.8446
2018-09-18  0.5808  0.2774