Combining

Joins

Outer Joins

One can perform a full outer join on the time indexes of two TS objects $x$ and $y$ in the following ways:

merge(x, y)

ojoin(x, y)

[x y]

hcat(x, y)

Where there are dates in the index of one that do not exist in the other, values will be filled with NaN objects. As the missing functionality matures in Julia's base syntax, it will eventually replace NaN in this context, since unfortunately NaN is only applicable for Float64 element types.

Base.merge — Function.

merge(x::TS,y::TS;join::Char='o')::TS

Merge two time series objects together by index with an optionally specified join type parameter.

...

Arguments

x::TS: Left side of the join.

y::TS: Right side of the join.

Optional args:

join::Char='o'::TS: Specifies the logic used to perform the merge, and may take on the values 'o' (outer join), 'i' (inner join), 'l' (left join), or 'r' (right join). Defaults to outer join, whose result is the same as hcat(x, y) or [x y].

...

source

Temporal.ojoin — Function.

ojoin(x::TS,y::TS)::TS

Outer join two TS objects by index.

Equivalent to x OUTER JOIN y ON x.index = y.index.

...

Arguments

x::TS: Left side of the join.

y::TS: Right side of the join.

...

source

Example

julia> x = TS(rand(252))
252x1 TS{Float64,Dates.Date}: 2018-06-28 to 2019-03-06

Index       A
2018-06-28  0.6635
2018-06-29  0.9726
2018-06-30  0.7033
2018-07-01  0.5292
2018-07-02  0.3878
2018-07-03  0.4196
2018-07-04  0.5528
⋮
2019-02-27  0.3805
2019-02-28  0.815
2019-03-01  0.8773
2019-03-02  0.6833
2019-03-03  0.7262
2019-03-04  0.2452
2019-03-05  0.5668
2019-03-06  0.7602

julia> y = TS(rand(252), x.index .- Month(6))
252x1 TS{Float64,Dates.Date}: 2017-12-28 to 2018-09-06

Index       A
2017-12-28  0.7039
2017-12-29  0.5739
2017-12-30  0.1876
2018-01-01  0.6255
2018-01-02  0.2646
2018-01-03  0.1014
2018-01-04  0.3257
⋮
2018-08-27  0.167
2018-08-28  0.9303
2018-09-01  0.4088
2018-09-02  0.7705
2018-09-03  0.3636
2018-09-04  0.443
2018-09-05  0.9754
2018-09-06  0.2472

julia> [x y]
431x2 TS{Float64,Dates.Date}: 2017-12-28 to 2019-03-06

Index       A       A
2017-12-28  NaN     0.7039
2017-12-29  NaN     0.5739
2017-12-30  NaN     0.1876
2018-01-01  NaN     0.6255
2018-01-02  NaN     0.2646
2018-01-03  NaN     0.1014
2018-01-04  NaN     0.3257
⋮
2019-02-27  0.3805  NaN
2019-02-28  0.815   NaN
2019-03-01  0.8773  NaN
2019-03-02  0.6833  NaN
2019-03-03  0.7262  NaN
2019-03-04  0.2452  NaN
2019-03-05  0.5668  NaN
2019-03-06  0.7602  NaN

Inner Joins

You can do inner joins on TS objects using the ijoin function, which will remove any observations corresponding to time steps where at least one of the joined objects is missing a row. This will basically keep only the rows where the time index of the left side and the right side intersect.

Temporal.ijoin — Function.

ijoin(x::TS,y::TS)::TS

Inner join two TS objects by index.

Equivalent to x INNER JOIN y on x.index = y.index.

...

Arguments

x::TS: Left side of the join.

y::TS: Right side of the join.

...

source

Example

julia> x = TS(rand(252))
252x1 TS{Float64,Dates.Date}: 2018-06-28 to 2019-03-06

Index       A
2018-06-28  0.0443
2018-06-29  0.121
2018-06-30  0.7169
2018-07-01  0.4032
2018-07-02  0.1406
2018-07-03  0.9834
2018-07-04  0.5073
⋮
2019-02-27  0.7116
2019-02-28  0.7196
2019-03-01  0.6286
2019-03-02  0.22
2019-03-03  0.9257
2019-03-04  0.3368
2019-03-05  0.4556
2019-03-06  0.2018

julia> y = TS(rand(252), x.index .- Month(6))
252x1 TS{Float64,Dates.Date}: 2017-12-28 to 2018-09-06

Index       A
2017-12-28  0.6262
2017-12-29  0.2031
2017-12-30  0.3304
2018-01-01  0.0766
2018-01-02  0.2426
2018-01-03  0.4102
2018-01-04  0.439
⋮
2018-08-27  0.1135
2018-08-28  0.9828
2018-09-01  0.0086
2018-09-02  0.6441
2018-09-03  0.9197
2018-09-04  0.6018
2018-09-05  0.4166
2018-09-06  0.4101

julia> ijoin(x, y)
ERROR: ArgumentError: number of rows of each array must match (got (68, 69))

Left/Right Joins

Left and right joins are performed similarly to inner joins and the typical SQL join queries using the index field each object as the joining key.

Left Join: keep all observations of the left side of the join, fill the right side with NaN's where missing the corresponding time index

Right Join: keep all observations of the right side of the join, fill the left side with NaN's where missing the corresponding time index

Temporal.ljoin — Function.

ljoin(x::TS, y::TS)::TS

Left join two TS objects by index.

Equivalent to x LEFT JOIN y ON x.index = y.index.

...

Arguments

x::TS: Left side of the join.

y::TS: Right side of the join.

...

source

Temporal.rjoin — Function.

rjoin(x::TS, y::TS)::TS

Right join two TS objects by index.

Equivalent to x RIGHT JOIN y ON x.index = y.index.

...

Arguments

x::TS: Left side of the join.

y::TS: Right side of the join.

...

source

Example

julia> x = TS(rand(252))
252x1 TS{Float64,Dates.Date}: 2018-06-28 to 2019-03-06

Index       A
2018-06-28  0.2095
2018-06-29  0.3018
2018-06-30  0.3735
2018-07-01  0.2151
2018-07-02  0.3348
2018-07-03  0.6373
2018-07-04  0.4155
⋮
2019-02-27  0.0837
2019-02-28  0.5726
2019-03-01  0.2886
2019-03-02  0.7072
2019-03-03  0.6088
2019-03-04  0.9643
2019-03-05  0.8663
2019-03-06  0.5803

julia> y = TS(rand(252), x.index .- Month(6))
252x1 TS{Float64,Dates.Date}: 2017-12-28 to 2018-09-06

Index       A
2017-12-28  0.1762
2017-12-29  0.7818
2017-12-30  0.0435
2018-01-01  0.6497
2018-01-02  0.7968
2018-01-03  0.3353
2018-01-04  0.8654
⋮
2018-08-27  0.9735
2018-08-28  0.8662
2018-09-01  0.1211
2018-09-02  0.4368
2018-09-03  0.58
2018-09-04  0.3686
2018-09-05  0.8057
2018-09-06  0.1414

julia> ljoin(x, y)
252x2 TS{Float64,Dates.Date}: 2018-06-28 to 2019-03-06

Index       A       A
2018-06-28  0.2095  0.2486
2018-06-29  0.3018  0.233
2018-06-30  0.3735  0.2345
2018-07-01  0.2151  0.9941
2018-07-02  0.3348  0.2787
2018-07-03  0.6373  0.6358
2018-07-04  0.4155  0.2873
⋮
2019-02-27  0.0837  NaN
2019-02-28  0.5726  NaN
2019-03-01  0.2886  NaN
2019-03-02  0.7072  NaN
2019-03-03  0.6088  NaN
2019-03-04  0.9643  NaN
2019-03-05  0.8663  NaN
2019-03-06  0.5803  NaN

julia> rjoin(x, y)
247x2 TS{Float64,Dates.Date}: 2017-12-28 to 2018-09-06

Index       A       A
2017-12-28  NaN     0.1762
2017-12-29  NaN     0.7818
2017-12-30  NaN     0.0435
2018-01-01  NaN     0.6497
2018-01-02  NaN     0.7968
2018-01-03  NaN     0.3353
2018-01-04  NaN     0.8654
⋮
2018-08-27  0.5197  0.7332
2018-08-28  0.7537  0.7633
2018-09-01  0.8925  0.5937
2018-09-02  0.4181  0.5619
2018-09-03  0.3003  0.9406
2018-09-04  0.1091  0.9735
2018-09-05  0.3101  0.8662
2018-09-06  0.8578  0.1211