Combining

Joins

Outer Joins

One can perform a full outer join on the time indexes of two TS objects $x$ and $y$ in the following ways:

merge(x, y)

ojoin(x, y)

[x y]

hcat(x, y)

Where there are dates in the index of one that do not exist in the other, values will be filled with NaN objects. As the missing functionality matures in Julia's base syntax, it will eventually replace NaN in this context, since unfortunately NaN is only applicable for Float64 element types.

Base.merge — Function.

merge(x::TS,y::TS;join::Char='o')::TS

Merge two time series objects together by index with an optionally specified join type parameter.

...

Arguments

x::TS: Left side of the join.

y::TS: Right side of the join.

Optional args:

join::Char='o'::TS: Specifies the logic used to perform the merge, and may take on the values 'o' (outer join), 'i' (inner join), 'l' (left join), or 'r' (right join). Defaults to outer join, whose result is the same as hcat(x, y) or [x y].

...

source

Temporal.ojoin — Function.

ojoin(x::TS,y::TS)::TS

Outer join two TS objects by index.

Equivalent to x OUTER JOIN y ON x.index = y.index.

...

Arguments

x::TS: Left side of the join.

y::TS: Right side of the join.

...

source

Example

julia> x = TS(rand(252))
252x1 TS{Float64,Dates.Date}: 2019-02-21 to 2019-10-30

Index       A
2019-02-21  0.6019
2019-02-22  0.738
2019-02-23  0.3249
2019-02-24  0.0275
2019-02-25  0.0814
2019-02-26  0.9212
2019-02-27  0.3487
⋮
2019-10-23  0.0879
2019-10-24  0.4418
2019-10-25  0.4769
2019-10-26  0.9349
2019-10-27  0.9914
2019-10-28  0.9548
2019-10-29  0.0551
2019-10-30  0.8577

julia> y = TS(rand(252), x.index .- Month(6))
252x1 TS{Float64,Dates.Date}: 2018-08-21 to 2019-04-30

Index       A
2018-08-21  0.4667
2018-08-22  0.4071
2018-08-23  0.5146
2018-08-24  0.0375
2018-08-25  0.6033
2018-08-26  0.5832
2018-08-27  0.4554
⋮
2019-04-23  0.9488
2019-04-24  0.6247
2019-04-25  0.369
2019-04-26  0.0034
2019-04-27  0.8983
2019-04-28  0.3078
2019-04-29  0.5828
2019-04-30  0.6081

julia> [x y]
431x2 TS{Float64,Dates.Date}: 2018-08-21 to 2019-10-30

Index       A       A
2018-08-21  NaN     0.4667
2018-08-22  NaN     0.4071
2018-08-23  NaN     0.5146
2018-08-24  NaN     0.0375
2018-08-25  NaN     0.6033
2018-08-26  NaN     0.5832
2018-08-27  NaN     0.4554
⋮
2019-10-23  0.0879  NaN
2019-10-24  0.4418  NaN
2019-10-25  0.4769  NaN
2019-10-26  0.9349  NaN
2019-10-27  0.9914  NaN
2019-10-28  0.9548  NaN
2019-10-29  0.0551  NaN
2019-10-30  0.8577  NaN

Inner Joins

You can do inner joins on TS objects using the ijoin function, which will remove any observations corresponding to time steps where at least one of the joined objects is missing a row. This will basically keep only the rows where the time index of the left side and the right side intersect.

Temporal.ijoin — Function.

ijoin(x::TS,y::TS)::TS

Inner join two TS objects by index.

Equivalent to x INNER JOIN y on x.index = y.index.

...

Arguments

x::TS: Left side of the join.

y::TS: Right side of the join.

...

source

Example

julia> x = TS(rand(252))
252x1 TS{Float64,Dates.Date}: 2019-02-21 to 2019-10-30

Index       A
2019-02-21  0.9945
2019-02-22  0.3099
2019-02-23  0.6868
2019-02-24  0.8592
2019-02-25  0.4654
2019-02-26  0.35
2019-02-27  0.4946
⋮
2019-10-23  0.7059
2019-10-24  0.2508
2019-10-25  0.3637
2019-10-26  0.4007
2019-10-27  0.9315
2019-10-28  0.4714
2019-10-29  0.8622
2019-10-30  0.7216

julia> y = TS(rand(252), x.index .- Month(6))
252x1 TS{Float64,Dates.Date}: 2018-08-21 to 2019-04-30

Index       A
2018-08-21  0.4692
2018-08-22  0.8751
2018-08-23  0.6117
2018-08-24  0.5905
2018-08-25  0.4138
2018-08-26  0.4333
2018-08-27  0.1858
⋮
2019-04-23  0.3717
2019-04-24  0.1158
2019-04-25  0.7372
2019-04-26  0.1076
2019-04-27  0.8438
2019-04-28  0.9739
2019-04-29  0.6257
2019-04-30  0.3233

julia> ijoin(x, y)
ERROR: ArgumentError: number of rows of each array must match (got (68, 71))

Left/Right Joins

Left and right joins are performed similarly to inner joins and the typical SQL join queries using the index field each object as the joining key.

Left Join: keep all observations of the left side of the join, fill the right side with NaN's where missing the corresponding time index

Right Join: keep all observations of the right side of the join, fill the left side with NaN's where missing the corresponding time index

Temporal.ljoin — Function.

ljoin(x::TS, y::TS)::TS

Left join two TS objects by index.

Equivalent to x LEFT JOIN y ON x.index = y.index.

...

Arguments

x::TS: Left side of the join.

y::TS: Right side of the join.

...

source

Temporal.rjoin — Function.

rjoin(x::TS, y::TS)::TS

Right join two TS objects by index.

Equivalent to x RIGHT JOIN y ON x.index = y.index.

...

Arguments

x::TS: Left side of the join.

y::TS: Right side of the join.

...

source

Example

julia> x = TS(rand(252))
252x1 TS{Float64,Dates.Date}: 2019-02-21 to 2019-10-30

Index       A
2019-02-21  0.9084
2019-02-22  0.0324
2019-02-23  0.5048
2019-02-24  0.8746
2019-02-25  0.8284
2019-02-26  0.3145
2019-02-27  0.261
⋮
2019-10-23  0.3542
2019-10-24  0.3195
2019-10-25  0.859
2019-10-26  0.1503
2019-10-27  0.853
2019-10-28  0.7523
2019-10-29  0.2245
2019-10-30  0.5534

julia> y = TS(rand(252), x.index .- Month(6))
252x1 TS{Float64,Dates.Date}: 2018-08-21 to 2019-04-30

Index       A
2018-08-21  0.1224
2018-08-22  0.0194
2018-08-23  0.2063
2018-08-24  0.3048
2018-08-25  0.7281
2018-08-26  0.5465
2018-08-27  0.4309
⋮
2019-04-23  0.3113
2019-04-24  0.5336
2019-04-25  0.894
2019-04-26  0.7678
2019-04-27  0.6407
2019-04-28  0.7333
2019-04-29  0.8162
2019-04-30  0.0372

julia> ljoin(x, y)
252x2 TS{Float64,Dates.Date}: 2019-02-21 to 2019-10-30

Index       A       A
2019-02-21  0.9084  0.8005
2019-02-22  0.0324  0.6707
2019-02-23  0.5048  0.5958
2019-02-24  0.8746  0.2865
2019-02-25  0.8284  0.7494
2019-02-26  0.3145  0.6461
2019-02-27  0.261   0.6743
⋮
2019-10-23  0.3542  NaN
2019-10-24  0.3195  NaN
2019-10-25  0.859   NaN
2019-10-26  0.1503  NaN
2019-10-27  0.853   NaN
2019-10-28  0.7523  NaN
2019-10-29  0.2245  NaN
2019-10-30  0.5534  NaN

julia> rjoin(x, y)
247x2 TS{Float64,Dates.Date}: 2018-08-21 to 2019-04-30

Index       A       A
2018-08-21  NaN     0.1224
2018-08-22  NaN     0.0194
2018-08-23  NaN     0.2063
2018-08-24  NaN     0.3048
2018-08-25  NaN     0.7281
2018-08-26  NaN     0.5465
2018-08-27  NaN     0.4309
⋮
2019-04-23  0.3097  0.6746
2019-04-24  0.6689  0.7247
2019-04-25  0.0658  0.1335
2019-04-26  0.9819  0.0139
2019-04-27  0.9279  0.0216
2019-04-28  0.9753  0.3113
2019-04-29  0.377   0.5336
2019-04-30  0.1691  0.894