Joins

Outer Joins

One can perform a full outer join on the time indexes of two TS objects $x$ and $y$ in the following ways:

merge(x, y)

ojoin(x, y)

[x y]

hcat(x, y)

Where there are dates in the index of one that do not exist in the other, values will be filled with NaN objects. As the missing functionality matures in Julia's base syntax, it will eventually replace NaN in this context, since unfortunately NaN is only applicable for Float64 element types.

Base.merge — Function

merge(x::TS,y::TS;join::Char='o')::TS

Merge two time series objects together by index with an optionally specified join type parameter.

...

Arguments

x::TS: Left side of the join.

y::TS: Right side of the join.

Optional args:

join::Char='o'::TS: Specifies the logic used to perform the merge, and may take on the values 'o' (outer join), 'i' (inner join), 'l' (left join), or 'r' (right join). Defaults to outer join, whose result is the same as hcat(x, y) or [x y].

...

Temporal.ojoin — Function

ojoin(x::TS,y::TS)::TS

Outer join two TS objects by index.

Equivalent to x OUTER JOIN y ON x.index = y.index.

...

Arguments

x::TS: Left side of the join.

y::TS: Right side of the join.

...

Example

julia> x = TS(rand(252))
252x1 TS{Float64,Dates.Date}: 2019-04-11 to 2019-12-18

Index       A
2019-04-11  0.0509
2019-04-12  0.1125
2019-04-13  0.6439
2019-04-14  0.7214
2019-04-15  0.5509
2019-04-16  0.39
2019-04-17  0.0028
⋮
2019-12-11  0.2974
2019-12-12  0.4077
2019-12-13  0.0745
2019-12-14  0.4187
2019-12-15  0.5841
2019-12-16  0.5552
2019-12-17  0.2959
2019-12-18  0.3171

julia> y = TS(rand(252), x.index .- Month(6))
252x1 TS{Float64,Dates.Date}: 2018-10-11 to 2019-06-18

Index       A
2018-10-11  0.473
2018-10-12  0.8509
2018-10-13  0.7007
2018-10-14  0.3527
2018-10-15  0.1154
2018-10-16  0.5913
2018-10-17  0.5844
⋮
2019-06-11  0.4125
2019-06-12  0.2417
2019-06-13  0.4204
2019-06-14  0.7101
2019-06-15  0.3761
2019-06-16  0.9988
2019-06-17  0.299
2019-06-18  0.088

julia> [x y]
431x2 TS{Float64,Dates.Date}: 2018-10-11 to 2019-12-18

Index       A       A
2018-10-11  NaN     0.473
2018-10-12  NaN     0.8509
2018-10-13  NaN     0.7007
2018-10-14  NaN     0.3527
2018-10-15  NaN     0.1154
2018-10-16  NaN     0.5913
2018-10-17  NaN     0.5844
⋮
2019-12-11  0.2974  NaN
2019-12-12  0.4077  NaN
2019-12-13  0.0745  NaN
2019-12-14  0.4187  NaN
2019-12-15  0.5841  NaN
2019-12-16  0.5552  NaN
2019-12-17  0.2959  NaN
2019-12-18  0.3171  NaN

Inner Joins

You can do inner joins on TS objects using the ijoin function, which will remove any observations corresponding to time steps where at least one of the joined objects is missing a row. This will basically keep only the rows where the time index of the left side and the right side intersect.

Temporal.ijoin — Function

ijoin(x::TS,y::TS)::TS

Inner join two TS objects by index.

Equivalent to x INNER JOIN y on x.index = y.index.

...

Arguments

x::TS: Left side of the join.

y::TS: Right side of the join.

...

Example

julia> x = TS(rand(252))
252x1 TS{Float64,Dates.Date}: 2019-04-11 to 2019-12-18

Index       A
2019-04-11  0.5133
2019-04-12  0.1692
2019-04-13  0.9335
2019-04-14  0.3243
2019-04-15  0.2011
2019-04-16  0.1531
2019-04-17  0.8056
⋮
2019-12-11  0.5708
2019-12-12  0.204
2019-12-13  0.7116
2019-12-14  0.7056
2019-12-15  0.3345
2019-12-16  0.9506
2019-12-17  0.754
2019-12-18  0.0461

julia> y = TS(rand(252), x.index .- Month(6))
252x1 TS{Float64,Dates.Date}: 2018-10-11 to 2019-06-18

Index       A
2018-10-11  0.5388
2018-10-12  0.8961
2018-10-13  0.5813
2018-10-14  0.6349
2018-10-15  0.2205
2018-10-16  0.4201
2018-10-17  0.7371
⋮
2019-06-11  0.8975
2019-06-12  0.9626
2019-06-13  0.2838
2019-06-14  0.1194
2019-06-15  0.6148
2019-06-16  0.4575
2019-06-17  0.1581
2019-06-18  0.7157

julia> ijoin(x, y)
ERROR: ArgumentError: number of rows of each array must match (got (68, 69))

Left/Right Joins

Left and right joins are performed similarly to inner joins and the typical SQL join queries using the index field each object as the joining key.

Left Join: keep all observations of the left side of the join, fill the right side with NaN's where missing the corresponding time index

Right Join: keep all observations of the right side of the join, fill the left side with NaN's where missing the corresponding time index

Temporal.ljoin — Function

ljoin(x::TS, y::TS)::TS

Left join two TS objects by index.

Equivalent to x LEFT JOIN y ON x.index = y.index.

...

Arguments

x::TS: Left side of the join.

y::TS: Right side of the join.

...

Temporal.rjoin — Function

rjoin(x::TS, y::TS)::TS

Right join two TS objects by index.

Equivalent to x RIGHT JOIN y ON x.index = y.index.

...

Arguments

x::TS: Left side of the join.

y::TS: Right side of the join.

...

Example

julia> x = TS(rand(252))
252x1 TS{Float64,Dates.Date}: 2019-04-11 to 2019-12-18

Index       A
2019-04-11  0.0831
2019-04-12  0.1706
2019-04-13  0.953
2019-04-14  0.0713
2019-04-15  0.3297
2019-04-16  0.485
2019-04-17  0.2349
⋮
2019-12-11  0.7077
2019-12-12  0.5814
2019-12-13  0.3632
2019-12-14  0.3485
2019-12-15  0.972
2019-12-16  0.6417
2019-12-17  0.7384
2019-12-18  0.5882

julia> y = TS(rand(252), x.index .- Month(6))
252x1 TS{Float64,Dates.Date}: 2018-10-11 to 2019-06-18

Index       A
2018-10-11  0.3052
2018-10-12  0.7378
2018-10-13  0.9483
2018-10-14  0.5678
2018-10-15  0.2285
2018-10-16  0.7352
2018-10-17  0.1002
⋮
2019-06-11  0.52
2019-06-12  0.7595
2019-06-13  0.4799
2019-06-14  0.8339
2019-06-15  0.8041
2019-06-16  0.719
2019-06-17  0.2676
2019-06-18  0.6497

julia> ljoin(x, y)
252x2 TS{Float64,Dates.Date}: 2019-04-11 to 2019-12-18

Index       A       A
2019-04-11  0.0831  0.1329
2019-04-12  0.1706  0.6376
2019-04-13  0.953   0.9687
2019-04-14  0.0713  0.8105
2019-04-15  0.3297  0.0059
2019-04-16  0.485   0.5718
2019-04-17  0.2349  0.1753
⋮
2019-12-11  0.7077  NaN
2019-12-12  0.5814  NaN
2019-12-13  0.3632  NaN
2019-12-14  0.3485  NaN
2019-12-15  0.972   NaN
2019-12-16  0.6417  NaN
2019-12-17  0.7384  NaN
2019-12-18  0.5882  NaN

julia> rjoin(x, y)
247x2 TS{Float64,Dates.Date}: 2018-10-11 to 2019-06-18

Index       A       A
2018-10-11  NaN     0.3052
2018-10-12  NaN     0.7378
2018-10-13  NaN     0.9483
2018-10-14  NaN     0.5678
2018-10-15  NaN     0.2285
2018-10-16  NaN     0.7352
2018-10-17  NaN     0.1002
⋮
2019-06-11  0.6573  0.3952
2019-06-12  0.6333  0.2849
2019-06-13  0.5994  0.8085
2019-06-14  0.7907  0.2548
2019-06-15  0.4612  0.7566
2019-06-16  0.9237  0.52
2019-06-17  0.0478  0.7595
2019-06-18  0.2657  0.4799