Combining

Joins

Outer Joins

One can perform a full outer join on the time indexes of two TS objects $x$ and $y$ in the following ways:

merge(x, y)

ojoin(x, y)

[x y]

hcat(x, y)

Where there are dates in the index of one that do not exist in the other, values will be filled with NaN objects. As the missing functionality matures in Julia's base syntax, it will eventually replace NaN in this context, since unfortunately NaN is only applicable for Float64 element types.

Base.merge — Function.

merge(x::TS,y::TS;join::Char='o')::TS

Merge two time series objects together by index with an optionally specified join type parameter.

...

Arguments

x::TS: Left side of the join.

y::TS: Right side of the join.

Optional args:

join::Char='o'::TS: Specifies the logic used to perform the merge, and may take on the values 'o' (outer join), 'i' (inner join), 'l' (left join), or 'r' (right join). Defaults to outer join, whose result is the same as hcat(x, y) or [x y].

...

source

Temporal.ojoin — Function.

ojoin(x::TS,y::TS)::TS

Outer join two TS objects by index.

Equivalent to x OUTER JOIN y ON x.index = y.index.

...

Arguments

x::TS: Left side of the join.

y::TS: Right side of the join.

...

source

Example

julia> x = TS(rand(252))
252x1 TS{Float64,Dates.Date}: 2019-02-22 to 2019-10-31

Index       A
2019-02-22  0.201
2019-02-23  0.5739
2019-02-24  0.2416
2019-02-25  0.7551
2019-02-26  0.5793
2019-02-27  0.3507
2019-02-28  0.3616
⋮
2019-10-24  0.0239
2019-10-25  0.523
2019-10-26  0.9617
2019-10-27  0.9891
2019-10-28  0.9501
2019-10-29  0.3224
2019-10-30  0.8774
2019-10-31  0.1845

julia> y = TS(rand(252), x.index .- Month(6))
252x1 TS{Float64,Dates.Date}: 2018-08-22 to 2019-04-30

Index       A
2018-08-22  0.6206
2018-08-23  0.3206
2018-08-24  0.3999
2018-08-25  0.5876
2018-08-26  0.6722
2018-08-27  0.2933
2018-08-28  0.8454
⋮
2019-04-24  0.5913
2019-04-25  0.4901
2019-04-26  0.1855
2019-04-27  0.369
2019-04-28  0.0509
2019-04-29  0.9975
2019-04-30  0.5423
2019-04-30  0.6129

julia> [x y]
431x2 TS{Float64,Dates.Date}: 2018-08-22 to 2019-10-31

Index       A       A
2018-08-22  NaN     0.6206
2018-08-23  NaN     0.3206
2018-08-24  NaN     0.3999
2018-08-25  NaN     0.5876
2018-08-26  NaN     0.6722
2018-08-27  NaN     0.2933
2018-08-28  NaN     0.8454
⋮
2019-10-24  0.0239  NaN
2019-10-25  0.523   NaN
2019-10-26  0.9617  NaN
2019-10-27  0.9891  NaN
2019-10-28  0.9501  NaN
2019-10-29  0.3224  NaN
2019-10-30  0.8774  NaN
2019-10-31  0.1845  NaN

Inner Joins

You can do inner joins on TS objects using the ijoin function, which will remove any observations corresponding to time steps where at least one of the joined objects is missing a row. This will basically keep only the rows where the time index of the left side and the right side intersect.

Temporal.ijoin — Function.

ijoin(x::TS,y::TS)::TS

Inner join two TS objects by index.

Equivalent to x INNER JOIN y on x.index = y.index.

...

Arguments

x::TS: Left side of the join.

y::TS: Right side of the join.

...

source

Example

julia> x = TS(rand(252))
252x1 TS{Float64,Dates.Date}: 2019-02-22 to 2019-10-31

Index       A
2019-02-22  0.8242
2019-02-23  0.912
2019-02-24  0.4887
2019-02-25  0.8836
2019-02-26  0.5874
2019-02-27  0.3669
2019-02-28  0.9205
⋮
2019-10-24  0.8899
2019-10-25  0.5614
2019-10-26  0.0245
2019-10-27  0.3372
2019-10-28  0.2149
2019-10-29  0.0317
2019-10-30  0.8673
2019-10-31  0.2732

julia> y = TS(rand(252), x.index .- Month(6))
252x1 TS{Float64,Dates.Date}: 2018-08-22 to 2019-04-30

Index       A
2018-08-22  0.3171
2018-08-23  0.8452
2018-08-24  0.2634
2018-08-25  0.5524
2018-08-26  0.264
2018-08-27  0.2104
2018-08-28  0.8465
⋮
2019-04-24  0.0308
2019-04-25  0.7666
2019-04-26  0.1378
2019-04-27  0.2632
2019-04-28  0.01
2019-04-29  0.2135
2019-04-30  0.6518
2019-04-30  0.6493

julia> ijoin(x, y)
ERROR: ArgumentError: number of rows of each array must match (got (67, 71))

Left/Right Joins

Left and right joins are performed similarly to inner joins and the typical SQL join queries using the index field each object as the joining key.

Left Join: keep all observations of the left side of the join, fill the right side with NaN's where missing the corresponding time index

Right Join: keep all observations of the right side of the join, fill the left side with NaN's where missing the corresponding time index

Temporal.ljoin — Function.

ljoin(x::TS, y::TS)::TS

Left join two TS objects by index.

Equivalent to x LEFT JOIN y ON x.index = y.index.

...

Arguments

x::TS: Left side of the join.

y::TS: Right side of the join.

...

source

Temporal.rjoin — Function.

rjoin(x::TS, y::TS)::TS

Right join two TS objects by index.

Equivalent to x RIGHT JOIN y ON x.index = y.index.

...

Arguments

x::TS: Left side of the join.

y::TS: Right side of the join.

...

source

Example

julia> x = TS(rand(252))
252x1 TS{Float64,Dates.Date}: 2019-02-22 to 2019-10-31

Index       A
2019-02-22  0.9863
2019-02-23  0.6912
2019-02-24  0.5126
2019-02-25  0.0167
2019-02-26  0.7757
2019-02-27  0.9955
2019-02-28  0.8706
⋮
2019-10-24  0.1868
2019-10-25  0.0664
2019-10-26  0.7405
2019-10-27  0.3865
2019-10-28  0.1123
2019-10-29  0.981
2019-10-30  0.855
2019-10-31  0.0679

julia> y = TS(rand(252), x.index .- Month(6))
252x1 TS{Float64,Dates.Date}: 2018-08-22 to 2019-04-30

Index       A
2018-08-22  0.2459
2018-08-23  0.2923
2018-08-24  0.4342
2018-08-25  0.9704
2018-08-26  0.4159
2018-08-27  0.9255
2018-08-28  0.7903
⋮
2019-04-24  0.351
2019-04-25  0.1436
2019-04-26  0.0142
2019-04-27  0.9602
2019-04-28  0.4226
2019-04-29  0.0873
2019-04-30  0.1011
2019-04-30  0.3672

julia> ljoin(x, y)
252x2 TS{Float64,Dates.Date}: 2019-02-22 to 2019-10-31

Index       A       A
2019-02-22  0.9863  0.1283
2019-02-23  0.6912  0.9533
2019-02-24  0.5126  0.4848
2019-02-25  0.0167  0.0695
2019-02-26  0.7757  0.9295
2019-02-27  0.9955  0.8641
2019-02-28  0.8706  0.9167
⋮
2019-10-24  0.1868  NaN
2019-10-25  0.0664  NaN
2019-10-26  0.7405  NaN
2019-10-27  0.3865  NaN
2019-10-28  0.1123  NaN
2019-10-29  0.981   NaN
2019-10-30  0.855   NaN
2019-10-31  0.0679  NaN

julia> rjoin(x, y)
246x2 TS{Float64,Dates.Date}: 2018-08-22 to 2019-04-30

Index       A       A
2018-08-22  NaN     0.2459
2018-08-23  NaN     0.2923
2018-08-24  NaN     0.4342
2018-08-25  NaN     0.9704
2018-08-26  NaN     0.4159
2018-08-27  NaN     0.9255
2018-08-28  NaN     0.7903
⋮
2019-04-23  0.5286  0.9408
2019-04-24  0.2968  0.3632
2019-04-25  0.5459  0.5681
2019-04-26  0.792   0.5421
2019-04-27  0.7835  0.5769
2019-04-28  0.1499  0.1549
2019-04-29  0.8203  0.351
2019-04-30  0.4488  0.1436